### 2025-09-04 5719ad5
Add resampling frequency  (#207)
* fix bug resampling

* fix bug select_days function

* json requires double quotes

* undo changes json
