### 2025-09-04 bcfaa8e
Add doi for GGIR, remove abstracts from bib
