+++ paper/paper.bib
@@ -11,11 +11,9 @@
   issn = {2373-8057},
   doi = {10.1038/s41531-023-00497-x},
   urldate = {2025-08-20},
-  abstract = {Digital health technologies can provide continuous monitoring and objective, real-world measures of Parkinson's disease (PD), but have primarily been evaluated in small, single-site studies. In this 12-month, multicenter observational study, we evaluated whether a smartwatch and smartphone application could measure features of early PD. 82 individuals with early, untreated PD and 50 age-matched controls wore research-grade sensors, a smartwatch, and a smartphone while performing standardized assessments in the clinic. At home, participants wore the smartwatch for seven days after each clinic visit and completed motor, speech and cognitive tasks on the smartphone every other week. Features derived from the devices, particularly arm swing, the proportion of time with tremor, and finger tapping, differed significantly between individuals with early PD and age-matched controls and had variable correlation with traditional assessments. Longitudinal assessments will inform the value of these digital measures for use in future clinical trials.},
   copyright = {2023 The Author(s)},
   langid = {english},
   keywords = {Parkinson's disease},
-  file = {C:\Users\z665206\Zotero\storage\EM8P33RA\Adams e.a. - 2023 - Using a smartwatch and smartphone to assess early Parkinson’s disease in the WATCH-PD study.pdf}
 }
 
 @article{bloemPersonalizedParkinsonProject2019a,
@@ -32,7 +30,6 @@
   doi = {10.1186/s12883-019-1394-3},
   urldate = {2025-08-11},
   langid = {english},
-  file = {C:\Users\z665206\Zotero\storage\VE7DP97W\Bloem e.a. - 2019 - The Personalized Parkinson Project examining disease progression through broad biomarkers in early.pdf}
 }
 
 @article{boyleActivityBehavioralRecognition2025,
@@ -48,10 +45,8 @@
   issn = {1424-8220},
   doi = {10.3390/s25030668},
   urldate = {2025-08-11},
-  abstract = {Background: With a progressively aging global population, the prevalence of Parkinson's Disease and dementia will increase, thus multiplying the healthcare burden worldwide. Sensing technology can complement the current measures used for symptom management and monitoring. The aim of this umbrella review is to provide future researchers with a synthesis of the current methodologies and metrics of sensing technologies for the management and monitoring of activities and behavioral symptoms in older adults with neurodegenerative disease. This is of key importance when considering the rapid obsolescence of and potential for future implementation of these technologies into real-world healthcare settings. Methods: Seven medical and technical databases were searched for systematic reviews (2018--2024) that met our inclusion/exclusion criteria. Articles were screened independently using Rayyan. PRISMA guidelines, the Cochrane Handbook for Systematic Reviews, and the Johanna Briggs Institute Critical Appraisal Checklist for Systematic Reviews were utilized for the assessment of bias, quality, and research synthesis. A narrative synthesis combines the study findings. Results: After screening 1458 articles, 9 systematic reviews were eligible for inclusion, synthesizing 402 primary studies. This umbrella review reveals that the use of sensing technologies for the observation and management of activities and behavioral symptoms is promising, however diversely applied, heterogenous in the methods used, and currently challenging to apply within clinical settings. Conclusions: Human activity and behavioral recognition requires true interdisciplinary collaborations between engineering, data science, and healthcare domains. The standardization of metrics, ethical AI development, and a culture of research-friendly technology and support are the next crucial developments needed for this rising field.},
   copyright = {https://creativecommons.org/licenses/by/4.0/},
   langid = {english},
[…]
