### 2025-09-04 daa1be3
Remove doi prefix
