+++ paper/paper.bib
@@ -176,7 +176,7 @@
   year = {2019},
   month = sep,
   issn = {2018-0063},
-  doi = {https://doi.org/10.1123/jmpb.2018-0063},
+  doi = {10.1123/jmpb.2018-0063},
   urldate = {2025-08-20},
   chapter = {Journal for the Measurement of Physical Behaviour},
   langid = {english},
