### 2025-09-05 7599487
incorporate modal tremor power in aggregate function
