### 2025-09-08 b8528a9
clarify variables
