### 2025-09-08 7247702
Add modal tremor power to feature statistics function (#212)
* fix bug resampling

* fix bug select_days function

* json requires double quotes

* undo changes json

* incorporate modal tremor power in aggregate function

* Incorporate suggestions

* fix type

* clarify variables
