+++ src/paradigma/config.py
@@ -223,7 +223,8 @@ class TremorConfig(IMUConfig):
         # -----------
         # Aggregation
         # -----------
-        self.aggregates_tremor_power: List[str] = ['mode', 'median', '90p']
+        self.aggregates_tremor_power: List[str] = ['mode_binned', 'median', '90p']
+        self.evaluation_points_tremor_power: np.ndarray = np.linspace(0, 6, 301) 
 
         # -----------------
         # TSDF data storage
