### 2025-09-08 7fc1739
Preparing for release - v103 (#213)
* V1.0.2 - Pulse rate consistency in naming (#199)

* update config

* add quantification step to tremor pipeline

* more efficient pipeline

* small typo

* fix documentation

* small changes in documentation

* add datetime column in step 4

* change order of columns in quantification step

* Remove capitalization and rename gait to arm swing during gait (#186)

* Remove capitalization and rename gait to arm swing during gait

* Change probability to predictions and update arrows in tremor pipeline

* Change probability to predictions and update arrows in 
[…]
