+++ src/paradigma/pipelines/gait_pipeline.py
@@ -1,3 +1,4 @@
+import logging
 import numpy as np
 import pandas as pd
 from scipy.signal import periodogram
@@ -10,9 +11,15 @@ from paradigma.feature_extraction import pca_transform_gyroscope, compute_angle,
     extract_angle_extremes, compute_range_of_motion, compute_peak_angular_velocity, compute_statistics, \
     compute_std_euclidean_norm, compute_power_in_bandwidth, compute_dominant_frequency, compute_mfccs, \
     compute_total_power
-from paradigma.segmenting import tabulate_windows, create_segments, discard_segments, categorize_segments, WindowedDataExtractor
+from paradigma.segmenting import tabulate_windows, create_segments, discard_segments, WindowedDataExtractor
 from paradigma.util import aggregate_parameter
 
+
+logger = logging.getLogger(__name__)
+
+# Only configure basic logging if no handlers exist
+if not logger.hasHandlers():
+    logging.basicConfig(level=logging.INFO, format='%(levelname)s: %(message)s')
 
 def extract_gait_features(
         df: pd.DataFrame,
@@ -351,19 +358,15 @@ def quantify_arm_swing(
     """
     # Group consecutive timestamps into segments, with new segments starting after a pre-specified gap.
     # Segments are made based on predicted gait
-    df[DataColumns.SEGMENT_NR] = create_segments(
+    df['unfiltered_segment_nr'] = create_segments(
         time_array=df[DataColumns.TIME], 
         max_segment_gap_s=max_segment_gap_s
     )
 
-    # Segment category is determined based on predicted gait, hence it is set
-    # before filtering the DataFrame to only include predicted no other arm activity
-    df[DataColumns.SEGMENT_CAT] = categorize_segments(df=df, fs=fs)
-
     # Remove segments that do not meet predetermined criteria
     df = discard_segments(
         df=df,
-        segment_nr_colname=DataColumns.SEGMENT_NR,
+        segment_nr_colname='unfiltered_segment_nr',
         min_segment_length_s=min_segment_length_s,
         fs=fs,
         format='timestamps'
@@ -371,7 +374,10 @@ def quantify_arm_swing(
 
     if df.empty:
         raise ValueError("No segments found in the input data after discarding segments of invalid shape.")
-
+    
+    # Create dictionary of gait segment number and duration
+    gait_segment_duration_dict = {segment_nr: len(group[DataColumns.TIME]) / fs for segment_nr, group in df.groupby('unfiltered_segment_nr', sort=False)}
+    
     # If no arm swing data is remaining, return an empty dictionary
     if filtered and df.loc[df[DataColumns.PRED_NO_OTHER_ARM_ACTIVITY]==1].empty:
         raise ValueError("No gait without other arm activities to quantify.")
@@ -380,7 +386,7 @@ def quantify_arm_swing(
         df = df.loc[df[DataColumns.PRED_NO_OTHER_ARM_ACTIVITY]==1].reset_index(drop=True)
 
         # Group consecutive timestamps into segments of filtered gait
-        df[DataColumns.SEGMENT_NR] = create_segments(
+        df['filtered_segment_nr'] = create_segments(
             time_array=df[DataColumns.TIME], 
             max_segment_gap_s=max_segment_gap_s
         )
@@ -388,13 +394,15 @@ def quantify_arm_swing(
         # Remove segments that do not meet predetermined criteria
         df = discard_segments(
             df=df,
-            segment_nr_colname=DataColumns.SEGMENT_NR,
+            segment_nr_colname='filtered_segment_nr',
             min_segment_length_s=min_segment_length_s,
             fs=fs,
         )
 
         if df.empty:
[…]
