+++ _sources/tutorials/gait_analysis.ipynb.txt
@@ -842,7 +842,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": null,
+   "execution_count": 5,
    "metadata": {},
    "outputs": [],
    "source": [
@@ -1013,7 +1013,7 @@
      "name": "stdout",
      "output_type": "stream",
      "text": [
-      "A total of 61 features have been extracted from 2749 3 - second windows with 2.25 seconds overlap.\n"
+      "A total of 61 features have been extracted from 2749 3-second windows with 2.25 seconds overlap.\n"
      ]
     },
     {
@@ -1254,7 +1254,7 @@
     "    config=config,\n",
     ")\n",
     "\n",
-    "print(f\"A total of {df_arm_activity.shape[1] - 1} features have been extracted from {df_arm_activity.shape[0]} {config.window_length_s} - second windows with {config.window_length_s - config.window_step_length_s} seconds overlap.\")\n",
+    "print(f\"A total of {df_arm_activity.shape[1] - 1} features have been extracted from {df_arm_activity.shape[0]} {config.window_length_s}-second windows with {config.window_length_s - config.window_step_length_s} seconds overlap.\")\n",
     "df_arm_activity.head()"
    ]
   },
@@ -1441,10 +1441,12 @@
       "A total of 84 filtered gait segments have been quantified.\n",
       "\n",
       "Metadata of the first gait segment:\n",
-      "{'duration_s': 9.0,\n",
+      "{'duration_filtered_segment_s': 9.0,\n",
+      " 'duration_unfiltered_segment_s': 9.0,\n",
       " 'end_time_s': 2230.74,\n",
-      " 'segment_category': 'moderately_long',\n",
       " 'start_time_s': 2221.75}\n",
+      "\n",
+      "Of this example, the filtered gait segment of 9.0 seconds is part of an unfiltered segment of 9.0 seconds, which is at least as large as the filtered gait segment.\n",
       "\n",
       "Individual arm swings of the first gait segment of the filtered dataset:\n"
      ]
@@ -1471,7 +1473,6 @@
        "    <tr style=\"text-align: right;\">\n",
        "      <th></th>\n",
        "      <th>segment_nr</th>\n",
-       "      <th>segment_category</th>\n",
        "      <th>range_of_motion</th>\n",
        "      <th>peak_velocity</th>\n",
        "    </tr>\n",
@@ -1480,91 +1481,78 @@
        "    <tr>\n",
        "      <th>0</th>\n",
        "      <td>1</td>\n",
-       "      <td>moderately_long</td>\n",
        "      <td>19.218491</td>\n",
        "      <td>90.807689</td>\n",
        "    </tr>\n",
        "    <tr>\n",
        "      <th>1</th>\n",
        "      <td>1</td>\n",
-       "      <td>moderately_long</td>\n",
        "      <td>21.267287</td>\n",
        "      <td>105.781357</td>\n",
        "    </tr>\n",
        "    <tr>\n",
        "      <th>2</th>\n",
        "      <td>1</td>\n",
-       "      <td>moderately_long</td>\n",
        "      <td>23.582098</td>\n",
        "      <td>103.932332</td>\n",
        "    </tr>\n",
        "    <tr>\n",
        "      <th>3</th>\n",
        "      <td>1</td>\n",
-       "      <td>moderately_long</td>\n",
        "      <td>23.757712</td>\n",
        "      <td>114.846304</td>\n",
        "    </tr>\n",
        "    <tr>\n",
        "      <th>4</th>\n",
        "      <td>1</td>\n",
-       "      <td>moderately_long</td>\n",
        "      <td>17.430734</td>\n",
        "      <td>63.297391</td>\n",
[…]
