+++ _sources/tutorials/data_preparation.ipynb.txt
@@ -972,9 +972,9 @@
  ],
  "metadata": {
   "kernelspec": {
-   "display_name": "paradigma-Fn6RLG4_-py3.11",
+   "display_name": "paradigma (Poetry)",
    "language": "python",
-   "name": "python3"
+   "name": "paradigma"
   },
   "language_info": {
    "codemirror_mode": {
@@ -986,7 +986,7 @@
    "name": "python",
    "nbconvert_exporter": "python",
    "pygments_lexer": "ipython3",
-   "version": "3.11.5"
+   "version": "3.12.0"
   }
  },
  "nbformat": 4,
