### 2025-09-17 b06e9f0
add tolerance to config
