### 2025-09-19 adb7343
Remove segment categorization
