+++ src/paradigma/segmenting.py
@@ -245,74 +245,6 @@ def discard_segments(
 
     return df
 
-
-def categorize_segments(df, fs, format='timestamps', window_step_length_s=None):
-    """
-    Categorize segments based on their duration.
-
-    This function categorizes segments into four categories based on their duration 
-    in seconds. The categories are defined as:
-    - Category 1: Segments shorter than 5 seconds
-    - Category 2: Segments between 5 and 10 seconds
-    - Category 3: Segments between 10 and 20 seconds
-    - Category 4: Segments longer than 20 seconds
-
-    The duration of each segment is calculated based on the sampling frequency and 
-    the number of rows (data points) in the segment.
-
-    Parameters
-    ----------
-    df : pd.DataFrame
-        The input DataFrame containing the segment column with segment numbers.
-    config : object
-        A configuration object containing `sampling_frequency`.
-    format : str, optional
-        The format of the input data, either 'timestamps' or 'windows'.
-
-    Returns
-    -------
-    pd.Series
-        A Series containing the category for each segment:
-        - 'short' for segments < 5 seconds
-        - 'moderately_long' for segments between 5 and 10 seconds
-        - 'long' for segments between 10 and 20 seconds
-        - 'very_long' for segments > 20 seconds
-    """
-    if format == 'windows' and window_step_length_s is None:
-        raise ValueError("Window step length must be provided for 'windows' format.")
-    
-    # Define duration thresholds in seconds
-    d_max_duration = {
-        'short': 5,
-        'moderately_long': 10,
-        'long': 20
-    }
-    
-    # Convert thresholds to rows if format is 'timestamps'
-    if format == 'timestamps':
-        d_max_duration = {k: v * fs for k, v in d_max_duration.items()}
-
-    # Count rows per segment
-    segment_sizes = df[DataColumns.SEGMENT_NR].value_counts()
-
-    # Convert segment sizes to duration in seconds
-    if format == 'windows':
-        segment_sizes *= window_step_length_s
-
-    # Group by the segment column and apply the categorization
-    def categorize(segment_size):
-        if segment_size < d_max_duration['short']:
-            return 'short'
-        elif segment_size < d_max_duration['moderately_long']:
-            return 'moderately_long'
-        elif segment_size < d_max_duration['long']:
-            return 'long'
-        else:
-            return 'very_long'
-
-    # Apply categorization to the DataFrame
-    return df[DataColumns.SEGMENT_NR].map(segment_sizes).map(categorize).astype('category')
-
 class WindowedDataExtractor:
     """
     A utility class for extracting specific column indices and slices 
