### 2025-09-22 2d10e54
Change logic of default configs and add ability to update
