+++ src/paradigma/config.py
@@ -1,6 +1,9 @@
+from dataclasses import asdict
 from typing import Dict, List
 from paradigma.constants import DataColumns, DataUnits
 import numpy as np
+import warnings
+
 
 class BaseConfig:
     def __init__(self) -> None:
@@ -26,44 +29,57 @@ class BaseConfig:
         self.values_filename = f"{prefix}_values.bin"
 
 class IMUConfig(BaseConfig):
+    """
+    IMU configuration that uses DataColumns() to dynamically map available channels.
+    Works even if only accelerometer or only gyroscope data is present.
+    """
 
-    def __init__(self) -> None:
+    def __init__(self, column_mapping: dict[str, str] | None = None) -> None:
         super().__init__()
-
         self.set_filenames('IMU')
 
         self.acceleration_units = DataUnits.ACCELERATION
         self.rotation_units = DataUnits.ROTATION
-
         self.axes = ["x", "y", "z"]
 
-        self.accelerometer_cols: List[str] = [
-            DataColumns.ACCELEROMETER_X,
-            DataColumns.ACCELEROMETER_Y,
-            DataColumns.ACCELEROMETER_Z,
-        ]
-        self.gyroscope_cols: List[str] = [
-            DataColumns.GYROSCOPE_X,
-            DataColumns.GYROSCOPE_Y,
-            DataColumns.GYROSCOPE_Z,
-        ]
-        self.gravity_cols: List[str] = [
-            DataColumns.GRAV_ACCELEROMETER_X,
-            DataColumns.GRAV_ACCELEROMETER_Y,
-            DataColumns.GRAV_ACCELEROMETER_Z,
-        ]
+        # Generate a default mapping or override with user-provided mapping
+        default_mapping = asdict(DataColumns())
+        self.column_mapping = {**default_mapping, **(column_mapping or {})}
 
-        self.d_channels_accelerometer = {
-            DataColumns.ACCELEROMETER_X: self.acceleration_units,
-            DataColumns.ACCELEROMETER_Y: self.acceleration_units,
-            DataColumns.ACCELEROMETER_Z: self.acceleration_units,
+        self.time_colname = self.column_mapping["TIME"]
+
+        self.accelerometer_colnames: List[str] = []
+        self.gyroscope_colnames: List[str] = []
+        self.gravity_colnames: List[str] = []
+
+        self.d_channels_accelerometer: Dict[str, str] = {}
+        self.d_channels_gyroscope: Dict[str, str] = {}
+
+        accel_keys = ["ACCELEROMETER_X", "ACCELEROMETER_Y", "ACCELEROMETER_Z"]
+        grav_keys = ["GRAV_ACCELEROMETER_X", "GRAV_ACCELEROMETER_Y", "GRAV_ACCELEROMETER_Z"]
+        gyro_keys = ["GYROSCOPE_X", "GYROSCOPE_Y", "GYROSCOPE_Z"]
+
+        if all(k in self.column_mapping for k in accel_keys):
+            self.accelerometer_colnames = [self.column_mapping[k] for k in accel_keys]
+
+            if all(k in self.column_mapping for k in grav_keys):
+                self.gravity_colnames = [self.column_mapping[k] for k in grav_keys]
+
+            self.d_channels_accelerometer = {
+                c: self.acceleration_units for c in self.accelerometer_colnames
+            }
+
+        if all(k in self.column_mapping for k in gyro_keys):
+            self.gyroscope_colnames = [self.column_mapping[k] for k in gyro_keys]
+
+            self.d_channels_gyroscope = {
+                c: self.rotation_units for c in self.gyroscope_colnames
+            }
+
+        self.d_channels_imu: Dict[str, str] = {
[…]
