+++ src/paradigma/preprocessing.py
@@ -1,16 +1,12 @@
-import json
 import numpy as np
 import pandas as pd
-import tsdf
-from pathlib import Path
 from scipy import signal
 from scipy.interpolate import interp1d
 from typing import List, Tuple, Union
 from datetime import datetime
 
-from paradigma.constants import TimeUnit, DataColumns
 from paradigma.config import PPGConfig, IMUConfig
-from paradigma.util import write_df_data, read_metadata, invert_watch_side
+from paradigma.util import invert_watch_side
 
 
 def resample_data(
@@ -194,18 +190,18 @@ def preprocess_imu_data(df: pd.DataFrame, config: IMUConfig, sensor: str, watch_
 
     # Extract sensor column
     if sensor == 'accelerometer':
-        values_colnames = config.accelerometer_cols
+        values_colnames = config.accelerometer_colnames
     elif sensor == 'gyroscope':
-        values_colnames = config.gyroscope_cols
+        values_colnames = config.gyroscope_colnames
     elif sensor == 'both':
-        values_colnames = config.accelerometer_cols + config.gyroscope_cols
+        values_colnames = config.accelerometer_colnames + config.gyroscope_colnames
     else:
         raise('Sensor should be either accelerometer, gyroscope, or both')
         
     # Resample the data to the specified frequency
     df = resample_data(
         df=df,
-        time_column=DataColumns.TIME,
+        time_column=config.time_col,
         values_column_names=values_colnames,
         sampling_frequency=config.sampling_frequency,
         resampling_frequency=config.resampling_frequency
@@ -217,12 +213,12 @@ def preprocess_imu_data(df: pd.DataFrame, config: IMUConfig, sensor: str, watch_
     if sensor in ['accelerometer', 'both']:
       
         # Extract accelerometer data for filtering
-        accel_data = df[config.accelerometer_cols].values
+        accel_data = df[config.accelerometer_colnames].values
 
         # Define filter configurations for high-pass and low-pass
         filter_renaming_configs = {
-        "hp": {"result_columns": config.accelerometer_cols, "replace_original": True},
-        "lp": {"result_columns": [f'{col}_grav' for col in config.accelerometer_cols], "replace_original": False},
+        "hp": {"result_columns": config.accelerometer_colnames, "replace_original": True},
+        "lp": {"result_columns": [f'{col}_grav' for col in config.accelerometer_colnames], "replace_original": False},
         }
 
         # Apply filters in a loop
@@ -238,9 +234,9 @@ def preprocess_imu_data(df: pd.DataFrame, config: IMUConfig, sensor: str, watch_
             # Replace or add new columns based on configuration
             df[filter_config["result_columns"]] = filtered_data
 
-        values_colnames += config.gravity_cols
+        values_colnames += config.gravity_colnames
 
-    df = df[[DataColumns.TIME, *values_colnames]]
+    df = df[[config.time_colname, *values_colnames]]
 
     return df
 
@@ -273,12 +269,19 @@ def preprocess_ppg_data(df_ppg: pd.DataFrame, df_acc: pd.DataFrame, ppg_config: 
     """
 
     # Extract overlapping segments
-    df_ppg_overlapping, df_acc_overlapping = extract_overlapping_segments(df_ppg, df_acc, start_time_ppg, start_time_imu)
+    df_ppg_overlapping, df_acc_overlapping = extract_overlapping_segments(
+        df_ppg=df_ppg, 
+        df_acc=df_acc, 
+        time_colname_ppg=ppg_config.time_colname, 
[…]
