### 2025-09-22 b853dcb
Add time column to config
