### 2025-09-22 f7d4e11
Adjust docstrings
