### 2025-09-23 00649e7
Update worflow - remove kernelspec
