### 2025-09-23 3cbb4a6
Add info on scale sensitivity and make accelerometer optional
