### 2025-09-23 f671861
update preprocessing also for optional accelerometer
