+++ src/paradigma/testing.py
@@ -1,54 +1,100 @@
 import json
-import numpy as np
 import os
-import pandas as pd
 from pathlib import Path
-import tsdf
 from typing import List
 
+import numpy as np
+import pandas as pd
+import tsdf
+
 from paradigma.classification import ClassifierPackage
-from paradigma.config import IMUConfig, PPGConfig, GaitConfig, TremorConfig, PulseRateConfig
+from paradigma.config import (
+    GaitConfig,
+    IMUConfig,
+    PPGConfig,
+    PulseRateConfig,
+    TremorConfig,
+)
 from paradigma.constants import DataColumns, TimeUnit
-from paradigma.pipelines.gait_pipeline import extract_gait_features, detect_gait, \
-    extract_arm_activity_features, filter_gait
-from paradigma.pipelines.tremor_pipeline import extract_tremor_features, detect_tremor, \
-    aggregate_tremor
-from paradigma.pipelines.pulse_rate_pipeline import extract_signal_quality_features, signal_quality_classification, \
-    aggregate_pulse_rate
+from paradigma.pipelines.gait_pipeline import (
+    detect_gait,
+    extract_arm_activity_features,
+    extract_gait_features,
+    filter_gait,
+)
+from paradigma.pipelines.pulse_rate_pipeline import (
+    aggregate_pulse_rate,
+    extract_signal_quality_features,
+)
+from paradigma.pipelines.tremor_pipeline import (
+    aggregate_tremor,
+    detect_tremor,
+    extract_tremor_features,
+)
 from paradigma.preprocessing import preprocess_imu_data, preprocess_ppg_data
-from paradigma.util import read_metadata, write_df_data, get_end_iso8601, merge_predictions_with_timestamps
+from paradigma.util import (
+    get_end_iso8601,
+    merge_predictions_with_timestamps,
+    read_metadata,
+    write_df_data,
+)
 
 
-def preprocess_imu_data_io(path_to_input: str | Path, path_to_output: str | Path, 
-                           config: IMUConfig, sensor: str, watch_side: str) -> None:
+def preprocess_imu_data_io(
+    path_to_input: str | Path,
+    path_to_output: str | Path,
+    config: IMUConfig,
+    sensor: str,
+    watch_side: str,
+) -> None:
     # Load data
-    metadata_time, metadata_values = read_metadata(str(path_to_input), str(config.meta_filename),
-                                                    str(config.time_filename), str(config.values_filename))
-    df = tsdf.load_dataframe_from_binaries([metadata_time, metadata_values], tsdf.constants.ConcatenationType.columns)
+    metadata_time, metadata_values = read_metadata(
+        str(path_to_input),
+        str(config.meta_filename),
+        str(config.time_filename),
+        str(config.values_filename),
+    )
+    df = tsdf.load_dataframe_from_binaries(
+        [metadata_time, metadata_values], tsdf.constants.ConcatenationType.columns
+    )
 
     # Preprocess data
     df = preprocess_imu_data(df=df, config=config, sensor=sensor, watch_side=watch_side)
 
     # Store data
-    for sensor, units in zip(['accelerometer', 'gyroscope'], ['g', config.rotation_units]):
+    for sensor, units in zip(
+        ["accelerometer", "gyroscope"], ["g", config.rotation_units]
+    ):
         if any(sensor in col for col in df.columns):
             df_sensor = df[[DataColumns.TIME] + [x for x in df.columns if sensor in x]]
 
             metadata_values.channels = [x for x in df.columns if sensor in x]
[…]
