+++ src/paradigma/pipelines/pulse_rate_pipeline.py
@@ -1,24 +1,38 @@
-import json
-import numpy as np
-import os
-import pandas as pd
 from pathlib import Path
-from scipy.signal import welch
-from scipy.signal.windows import hamming, hann
-import tsdf
 from typing import List
 
+import numpy as np
+import pandas as pd
+from scipy.signal import welch
+from scipy.signal.windows import hamming, hann
+
 from paradigma.classification import ClassifierPackage
-from paradigma.constants import DataColumns
 from paradigma.config import PulseRateConfig
-from paradigma.feature_extraction import compute_statistics, compute_signal_to_noise_ratio, compute_auto_correlation, \
-    compute_dominant_frequency, compute_relative_power, compute_spectral_entropy
-from paradigma.pipelines.pulse_rate_utils import assign_sqa_label, extract_pr_segments, extract_pr_from_segment
-from paradigma.segmenting import tabulate_windows, WindowedDataExtractor
+from paradigma.constants import DataColumns
+from paradigma.feature_extraction import (
+    compute_auto_correlation,
+    compute_dominant_frequency,
+    compute_relative_power,
+    compute_signal_to_noise_ratio,
+    compute_spectral_entropy,
+    compute_statistics,
+)
+from paradigma.pipelines.pulse_rate_utils import (
+    assign_sqa_label,
+    extract_pr_from_segment,
+    extract_pr_segments,
+)
+from paradigma.segmenting import WindowedDataExtractor, tabulate_windows
 from paradigma.util import aggregate_parameter
 
-def extract_signal_quality_features(df_ppg: pd.DataFrame, df_acc: pd.DataFrame, ppg_config: PulseRateConfig, acc_config: PulseRateConfig) -> pd.DataFrame:
-    """	
+
+def extract_signal_quality_features(
+    df_ppg: pd.DataFrame,
+    df_acc: pd.DataFrame,
+    ppg_config: PulseRateConfig,
+    acc_config: PulseRateConfig,
+) -> pd.DataFrame:
+    """
     Extract signal quality features from the PPG signal.
     The features are extracted from the temporal and spectral domain of the PPG signal.
     The temporal domain features include variance, mean, median, kurtosis, skewness, signal-to-noise ratio, and autocorrelation.
@@ -39,23 +53,25 @@ def extract_signal_quality_features(df_ppg: pd.DataFrame, df_acc: pd.DataFrame, 
     -------
     df_features : pd.DataFrame
         The DataFrame containing the extracted signal quality features.
-    
+
     """
     # Group sequences of timestamps into windows
     ppg_windowed_cols = [DataColumns.TIME, ppg_config.ppg_colname]
     ppg_windowed = tabulate_windows(
-        df=df_ppg, 
+        df=df_ppg,
         columns=ppg_windowed_cols,
         window_length_s=ppg_config.window_length_s,
         window_step_length_s=ppg_config.window_step_length_s,
-        fs=ppg_config.sampling_frequency
+        fs=ppg_config.sampling_frequency,
     )
 
     # Extract data from the windowed PPG signal
     extractor = WindowedDataExtractor(ppg_windowed_cols)
     idx_time = extractor.get_index(DataColumns.TIME)
     idx_ppg = extractor.get_index(ppg_config.ppg_colname)
-    start_time_ppg = np.min(ppg_windowed[:, :, idx_time], axis=1) # Start time of the window is relative to the first datapoint in the PPG data
+    start_time_ppg = np.min(
+        ppg_windowed[:, :, idx_time], axis=1
+    )  # Start time of the window is relative to the first datapoint in the PPG data
     ppg_values_windowed = ppg_windowed[:, :, idx_ppg]
 
[…]

[…]
