### 2025-09-25 4933d41
Modify hook to python specific
