+++ .pre-commit-config.yaml
@@ -9,14 +9,14 @@ repos:
     hooks:
       - id: black
         pass_filenames: true
-        always_run: true
+        types: [python]
 
   - repo: https://github.com/PyCQA/isort
     rev: 5.13.2
     hooks:
       - id: isort
-        always_run: true
         args: ["--profile", "black"]
+        types: [python]
 
   - repo: https://github.com/pre-commit/pre-commit-hooks
     rev: v4.5.0
