### 2025-09-25 e435e66
Remove json check as it blocks metadata
