+++ .pre-commit-config.yaml
@@ -34,7 +34,6 @@ repos:
     rev: v4.5.0
     hooks:
       - id: check-yaml
-      - id: check-json
       - id: end-of-file-fixer
       - id: trailing-whitespace
       - id: check-added-large-files
