### 2025-09-26 46045a5
Add tolerance for contiguous data (#223)
* fix bug resampling

* fix bug select_days function

* json requires double quotes

* undo changes json

* incorporate modal tremor power in aggregate function

* Incorporate suggestions

* fix type

* clarify variables

* add tolerance to config

* change preprocess_ppg_data too

* add suggestions

* run one segment for pulse rate only

* fix typos

* add suggestion
