### 2025-09-29 a47ff03
Adding pre-commit hooks (#222)
* Add pre-commit hooks and add Poetry scripts for simplifying dev process
