+++ src/paradigma/pipelines/gait_pipeline.py
@@ -1,30 +1,43 @@
 import logging
+from typing import List, Tuple
+
 import numpy as np
 import pandas as pd
 from scipy.signal import periodogram
-from typing import List, Tuple
 
 from paradigma.classification import ClassifierPackage
-from paradigma.constants import DataColumns
 from paradigma.config import GaitConfig
-from paradigma.feature_extraction import pca_transform_gyroscope, compute_angle, remove_moving_average_angle, \
-    extract_angle_extremes, compute_range_of_motion, compute_peak_angular_velocity, compute_statistics, \
-    compute_std_euclidean_norm, compute_power_in_bandwidth, compute_dominant_frequency, compute_mfccs, \
-    compute_total_power
-from paradigma.segmenting import tabulate_windows, create_segments, discard_segments, WindowedDataExtractor
+from paradigma.constants import DataColumns
+from paradigma.feature_extraction import (
+    compute_angle,
+    compute_dominant_frequency,
+    compute_mfccs,
+    compute_peak_angular_velocity,
+    compute_power_in_bandwidth,
+    compute_range_of_motion,
+    compute_statistics,
+    compute_std_euclidean_norm,
+    compute_total_power,
+    extract_angle_extremes,
+    pca_transform_gyroscope,
+    remove_moving_average_angle,
+)
+from paradigma.segmenting import (
+    WindowedDataExtractor,
+    create_segments,
+    discard_segments,
+    tabulate_windows,
+)
 from paradigma.util import aggregate_parameter
-
 
 logger = logging.getLogger(__name__)
 
 # Only configure basic logging if no handlers exist
 if not logger.hasHandlers():
-    logging.basicConfig(level=logging.INFO, format='%(levelname)s: %(message)s')
+    logging.basicConfig(level=logging.INFO, format="%(levelname)s: %(message)s")
 
-def extract_gait_features(
-        df: pd.DataFrame,
-        config: GaitConfig
-    ) -> pd.DataFrame:
+
+def extract_gait_features(df: pd.DataFrame, config: GaitConfig) -> pd.DataFrame:
     """
     Extracts gait features from accelerometer and gravity sensor data in the input DataFrame by computing temporal and spectral features.
 
@@ -51,7 +64,7 @@ def extract_gait_features(
         A DataFrame containing extracted gait features, including temporal and spectral domain features. The DataFrame will have
         columns corresponding to time, statistical features of the accelerometer and gravity data, and spectral features of the
         accelerometer data.
-    
+
     Notes
     -----
     - This function groups the data into windows based on timestamps and applies Fast Fourier Transform to compute spectral features.
@@ -66,11 +79,11 @@ def extract_gait_features(
     # Group sequences of timestamps into windows
     windowed_cols = [DataColumns.TIME] + config.accelerometer_cols + config.gravity_cols
     windowed_data = tabulate_windows(
-        df=df, 
+        df=df,
         columns=windowed_cols,
         window_length_s=config.window_length_s,
         window_step_length_s=config.window_step_length_s,
-        fs=config.sampling_frequency
+        fs=config.sampling_frequency,
     )
 
     extractor = WindowedDataExtractor(windowed_cols)
@@ -85,13 +98,13 @@ def extract_gait_features(
     windowed_grav = windowed_data[:, :, idx_grav]
 
     df_features = pd.DataFrame(start_time, columns=[DataColumns.TIME])
-    
+
     # Compute statistics of the temporal domain signals (mean, std) for accelerometer and gravity
     df_temporal_features = extract_temporal_domain_features(
-        config=config, 
+        config=config,
         windowed_acc=windowed_acc,
         windowed_grav=windowed_grav,
-        grav_stats=['mean', 'std']
+        grav_stats=["mean", "std"],
     )
 
[…]
