+++ src/paradigma/config.py
@@ -1,12 +1,15 @@
 from typing import Dict, List
-from paradigma.constants import DataColumns, DataUnits
+
 import numpy as np
+
+from paradigma.constants import DataColumns, DataUnits
+
 
 class BaseConfig:
     def __init__(self) -> None:
-        self.meta_filename = ''
-        self.values_filename = ''
-        self.time_filename = ''
+        self.meta_filename = ""
+        self.values_filename = ""
+        self.time_filename = ""
 
     def set_sensor(self, sensor: str) -> None:
         """Sets the sensor and derived filenames"""
@@ -15,7 +18,7 @@ class BaseConfig:
 
     def set_filenames(self, prefix: str) -> None:
         """Sets the filenames based on the prefix. This method is duplicated from `gaits_analysis_config.py`.
-        
+
         Parameters
         ----------
         prefix : str
@@ -25,12 +28,13 @@ class BaseConfig:
         self.time_filename = f"{prefix}_time.bin"
         self.values_filename = f"{prefix}_values.bin"
 
+
 class IMUConfig(BaseConfig):
 
     def __init__(self) -> None:
         super().__init__()
 
-        self.set_filenames('IMU')
+        self.set_filenames("IMU")
 
         self.acceleration_units = DataUnits.ACCELERATION
         self.rotation_units = DataUnits.ROTATION
@@ -63,7 +67,10 @@ class IMUConfig(BaseConfig):
             DataColumns.GYROSCOPE_Y: self.rotation_units,
             DataColumns.GYROSCOPE_Z: self.rotation_units,
         }
-        self.d_channels_imu = {**self.d_channels_accelerometer, **self.d_channels_gyroscope}
+        self.d_channels_imu = {
+            **self.d_channels_accelerometer,
+            **self.d_channels_gyroscope,
+        }
 
         self.sampling_frequency = 100
         self.resampling_frequency = 100
@@ -78,7 +85,7 @@ class PPGConfig(BaseConfig):
     def __init__(self) -> None:
         super().__init__()
 
-        self.set_filenames('PPG')
+        self.set_filenames("PPG")
 
         self.ppg_colname = DataColumns.PPG
 
@@ -89,9 +96,7 @@ class PPGConfig(BaseConfig):
         self.upper_cutoff_frequency = 3.5
         self.filter_order = 4
 
-        self.d_channels_ppg = {
-            DataColumns.PPG: DataUnits.NONE
-        }
+        self.d_channels_ppg = {DataColumns.PPG: DataUnits.NONE}
 
 
 # Domain base configs
@@ -100,7 +105,7 @@ class GaitConfig(IMUConfig):
     def __init__(self, step) -> None:
         super().__init__()
 
-        self.set_sensor('accelerometer')
+        self.set_sensor("accelerometer")
 
         # ----------
         # Segmenting
@@ -108,7 +113,7 @@ class GaitConfig(IMUConfig):
         self.max_segment_gap_s = 1.5
         self.min_segment_length_s = 1.5
 
-        if step == 'gait':
+        if step == "gait":
             self.window_length_s: float = 6
             self.window_step_length_s: float = 1
         else:
@@ -169,11 +174,15 @@ class GaitConfig(IMUConfig):
         }
 
         for mfcc_coef in range(1, self.mfcc_n_coefficients + 1):
-            self.d_channels_values[f"accelerometer_mfcc_{mfcc_coef}"] = DataUnits.GRAVITY
+            self.d_channels_values[f"accelerometer_mfcc_{mfcc_coef}"] = (
+                DataUnits.GRAVITY
+            )
 
-        if step == 'arm_activity':
+        if step == "arm_activity":
[…]
