+++ pytype.cfg
@@ -2,7 +2,7 @@
 # Path to source files
 inputs = .
 # Path to exclude from analysis
-exclude = 
+exclude =
     **/node_modules
     **/venv
     **/.venv

[…]
