### 2025-09-30 685a919
Resolve pre-commit suggestions
