### 2025-09-30 0ecd77a
Change workflow to be more flexible for kernelspecs
