### 2025-09-30 a80320e
Remove redundant poetry activation
