### 2025-09-30 7c1adf6
Change kernelspec
