+++ docs/conf.py
@@ -51,6 +51,11 @@ exclude_patterns = ["build", "Thumbs.db", ".DS_Store"]
 
 nbsphinx_execute_arguments = ["--ExecutePreprocessor.kernel_name=python3"]
 
+# myst-nb execution settings
+jupyter_execute_notebooks = "force"  # always (re)execute notebooks
+nb_kernel_name = "python3"  # force python3, ignore metadata
+nb_execution_timeout = 600  # bump timeout if needed
+
 # -- Options for HTML output -------------------------------------------------
 
 # The theme to use for HTML and HTML Help pages.  See the documentation for
