### 2025-09-30 6f42326
Add flexibility colnames (#219)
* Remove segment categorization

* Add recommendation to use pre-set colnames

* Adjust pipelines to flexible column names and config updates

* Change logic of default configs and add ability to update

* Update dependencies

* Change workflow to be more flexible for kernelspecs
