+++ src/paradigma/pipelines/pulse_rate_utils.py
@@ -7,10 +7,8 @@ from paradigma.config import PulseRateConfig
 
 
 def assign_sqa_label(
-        ppg_prob: np.ndarray, 
-        config: PulseRateConfig, 
-        acc_label: np.ndarray | None = None
-    ) -> np.ndarray:
+    ppg_prob: np.ndarray, config: PulseRateConfig, acc_label: np.ndarray | None = None
+) -> np.ndarray:
     """
     Assigns a signal quality label to every individual data point.
 
