+++ src/paradigma/classification.py
@@ -1,10 +1,10 @@
 import pickle
 from pathlib import Path
 from typing import Any, Optional
-from sklearn.preprocessing import StandardScaler
 
 import numpy as np
 from sklearn.base import BaseEstimator
+from sklearn.preprocessing import StandardScaler
 
 
 class ClassifierPackage:
@@ -47,7 +47,7 @@ class ClassifierPackage:
         if not self.scaler:
             return X
         return self.scaler.transform(X)
-    
+
     def update_scaler(self, x_train) -> None:
         """
         Update the standard scaler used for feature transformation.
@@ -58,7 +58,7 @@ class ClassifierPackage:
             The training data to fit the scaler.
         """
         scaler = StandardScaler()
-        self.scaler = scaler.fit(x_train)  
+        self.scaler = scaler.fit(x_train)
 
     def predict_proba(self, X) -> float:
         """
