### 2025-10-07 51c64bc
Ppg scale sensitivity + optional accelerometer (#220)
* Add info on scale sensitivity and make accelerometer optional

* update preprocessing also for optional accelerometer

* update feedback PR

* resolve merge conflicts

* minor bug fixes

* pre-commit_check

* pre-commit updates

* PR_req_changes

* update_copilot_review

* error_fix_tutorial

* update_markdown

* tutorial_clarification_sqa

* update_markdown

* Update src/paradigma/preprocessing.py

Co-authored-by: Erik Post <57133568+Erikpostt@users.noreply.github.com>

* remove optional

---------

Co-authored-by: Erik Post <57133568+Er
[…]
