+++ src/paradigma/pipelines/pulse_rate_pipeline.py
@@ -1,4 +1,3 @@
-from pathlib import Path
 from typing import List
 
 import numpy as np
@@ -28,9 +27,9 @@ from paradigma.util import aggregate_parameter
 
 def extract_signal_quality_features(
     df_ppg: pd.DataFrame,
-    df_acc: pd.DataFrame,
     ppg_config: PulseRateConfig,
-    acc_config: PulseRateConfig,
+    df_acc: pd.DataFrame | None = None,
+    acc_config: PulseRateConfig | None = None,
 ) -> pd.DataFrame:
     """
     Extract signal quality features from the PPG signal.
@@ -56,7 +55,7 @@ def extract_signal_quality_features(
 
     """
     # Group sequences of timestamps into windows
-    ppg_windowed_colnames = [DataColumns.TIME, ppg_config.ppg_colname]
+    ppg_windowed_colnames = [ppg_config.time_colname, ppg_config.ppg_colname]
     ppg_windowed = tabulate_windows(
         df=df_ppg,
         columns=ppg_windowed_colnames,
@@ -67,28 +66,40 @@ def extract_signal_quality_features(
 
     # Extract data from the windowed PPG signal
     extractor = WindowedDataExtractor(ppg_windowed_colnames)
-    idx_time = extractor.get_index(DataColumns.TIME)
+    idx_time = extractor.get_index(ppg_config.time_colname)
     idx_ppg = extractor.get_index(ppg_config.ppg_colname)
     start_time_ppg = np.min(
         ppg_windowed[:, :, idx_time], axis=1
     )  # Start time of the window is relative to the first datapoint in the PPG data
     ppg_values_windowed = ppg_windowed[:, :, idx_ppg]
 
-    acc_windowed_colnames = [DataColumns.TIME] + acc_config.accelerometer_colnames
-    acc_windowed = tabulate_windows(
-        df=df_acc,
-        columns=acc_windowed_colnames,
-        window_length_s=acc_config.window_length_s,
-        window_step_length_s=acc_config.window_step_length_s,
-        fs=acc_config.sampling_frequency,
-    )
+    df_features = pd.DataFrame(start_time_ppg, columns=[ppg_config.time_colname])
 
-    # Extract data from the windowed accelerometer signal
-    extractor = WindowedDataExtractor(acc_windowed_colnames)
-    idx_acc = extractor.get_slice(acc_config.accelerometer_colnames)
-    acc_values_windowed = acc_windowed[:, :, idx_acc]
+    if df_acc is not None and acc_config is not None:
 
-    df_features = pd.DataFrame(start_time_ppg, columns=[DataColumns.TIME])
+        acc_windowed_colnames = [
+            acc_config.time_colname
+        ] + acc_config.accelerometer_colnames
+        acc_windowed = tabulate_windows(
+            df=df_acc,
+            columns=acc_windowed_colnames,
+            window_length_s=acc_config.window_length_s,
+            window_step_length_s=acc_config.window_step_length_s,
+            fs=acc_config.sampling_frequency,
+        )
+
+        # Extract data from the windowed accelerometer signal
+        extractor = WindowedDataExtractor(acc_windowed_colnames)
+        idx_acc = extractor.get_slice(acc_config.accelerometer_colnames)
+        acc_values_windowed = acc_windowed[:, :, idx_acc]
+
+        # Compute periodicity feature of the accelerometer signal
+        df_accelerometer_feature = extract_accelerometer_feature(
+            acc_values_windowed, ppg_values_windowed, acc_config
+        )
+        # Combine the accelerometer feature with the previously computed features
+        df_features = pd.concat([df_features, df_accelerometer_feature], axis=1)
+
     # Compute features of the temporal domain of the PPG signal
     df_temporal_features = extract_temporal_domain_features(
[…]
