### 2025-10-07 e7ebbdd
Process pre-commit hooks
