### 2025-10-07 b87809f
Clarify installing pre-commit
