+++ .pre-commit-config.yaml
@@ -4,6 +4,12 @@ repos:
     hooks:
       - id: nbstripout
 
+  - repo: https://github.com/srstevenson/nb-clean
+    rev: 3.0.0
+    hooks:
+      - id: nb-clean
+        args: ["--remove-kernel-metadata"]
+
   - repo: https://github.com/psf/black
     rev: 24.8.0
     hooks:
