### 2025-10-08 a00b1c1
Add nb-clean, update pre-commit hooks
