+++ .pre-commit-config.yaml
@@ -8,6 +8,8 @@ repos:
     rev: 4.0.1
     hooks:
       - id: nb-clean
+        args:
+          - --remove-empty-cells
 
 
   - repo: https://github.com/psf/black
