### 2025-10-20 d2e7dee
Solve hooks
