+++ pyproject.toml
@@ -46,6 +46,7 @@ pre-commit = "^4.3.0"
 jupyter-client = "^8.6.3"
 nbstripout = "^0.8.1"
 nb-clean = "^4.0.1"
+jupyter = "^1.1.1"
 
 [tool.poetry.scripts]
 build-docs = "scripts.build_docs:main"
