### 2025-10-21 e580804
Serve pre-commit hooks
