+++ .pre-commit-config.yaml
@@ -3,13 +3,13 @@ repos:
     rev: 0.8.1
     hooks:
       - id: nbstripout
+        args: ["--extra-keys", "metadata.kernelspec", "--extra-keys", "metadata.language_info"]
 
   - repo: https://github.com/srstevenson/nb-clean
     rev: 4.0.1
     hooks:
       - id: nb-clean
-        args:
-          - --remove-empty-cells
+        args: ["--remove-empty-cells"]
 
 
   - repo: https://github.com/psf/black
