### 2025-10-21 6de3e26
Remove all kernelspec info
