### 2025-10-21 345c94c
Remove all metadata
