### 2025-10-22 09c0324
Release v1.0.4 (#228)
* Clarify installing pre-commit

* Add nb-clean, update pre-commit hooks

* Update test notebooks

* Remove all kernelspec info
