+++ pyproject.toml
@@ -45,6 +45,8 @@ ipykernel = "^6.30.1"
 pre-commit = "^4.3.0"
 jupyter-client = "^8.6.3"
 nbstripout = "^0.8.1"
+nb-clean = "^4.0.1"
+jupyter = "^1.1.1"
 
 [tool.poetry.scripts]
 build-docs = "scripts.build_docs:main"
