### 2025-10-27 548d9bb
Update readme ppg scaling
