### 2025-10-29 34d542d
Enable build-and-test for all branches
