### 2025-10-29 b07e79d
Enable build-and-test for all branches
