+++ .github/workflows/build-and-test.yml
@@ -2,7 +2,7 @@ name: Build and test
 
 on:
   push:
-    branches: [ '**' ]
+    branches: [ main ]
   pull_request:
     branches: [ '**' ]
   workflow_call:
