### 2025-11-11 54db93f
Sync main to release (#233)
* Incorporate modal tremor power in aggregate function

* Add tolerance for contiguous data to config

* Add pre-commit hooks and add Poetry scripts for simplifying dev process

* Remove segment categorization (set to legacy)

* Adjust pipelines to flexible column names and config updates

* Add info on scale sensitivity and make accelerometer optional

* Clarify installing pre-commit
