+++ tests/test_gait_analysis.py
@@ -1,39 +1,45 @@
-import pandas as pd
 from pathlib import Path
 
-from paradigma.testing import filter_gait_io, detect_gait_io, extract_arm_activity_features_io, extract_gait_features_io
-from paradigma.config import GaitConfig
 from test_notebooks import compare_data
 
+from paradigma.config import GaitConfig
+from paradigma.testing import (
+    detect_gait_io,
+    extract_arm_activity_features_io,
+    extract_gait_features_io,
+    filter_gait_io,
+)
 
 # Mappings between the metadata and binary files
 
 gait_binaries_pairs: list[tuple[str, str]] = [
-        ("gait_meta.json", "gait_time.bin"),
-        ("gait_meta.json", "gait_values.bin"),
-    ]
+    ("gait_meta.json", "gait_time.bin"),
+    ("gait_meta.json", "gait_values.bin"),
+]
 
 arm_activity_binaries_pairs: list[tuple[str, str]] = [
-        ("arm_activity_meta.json", "arm_activity_values.bin"),
-        ("arm_activity_meta.json", "arm_activity_time.bin"),
-    ]
+    ("arm_activity_meta.json", "arm_activity_values.bin"),
+    ("arm_activity_meta.json", "arm_activity_time.bin"),
+]
 
 arm_swing_binaries_pairs: list[tuple[str, str]] = [
-        ("arm_swing_meta.json", "arm_swing_values.bin"),
-        ("arm_swing_meta.json", "arm_swing_time.bin"),
-    ]
+    ("arm_swing_meta.json", "arm_swing_values.bin"),
+    ("arm_swing_meta.json", "arm_swing_time.bin"),
+]
 
 accelerometer_binaries_pairs: list[tuple[str, str]] = [
-        ("accelerometer_meta.json", "accelerometer_values.bin"),
-        ("accelerometer_meta.json", "accelerometer_time.bin"),
-    ]
+    ("accelerometer_meta.json", "accelerometer_values.bin"),
+    ("accelerometer_meta.json", "accelerometer_time.bin"),
+]
 gyroscope_binaries_pairs: list[tuple[str, str]] = [
-        ("gyroscope_meta.json", "gyroscope_values.bin"),
-        ("gyroscope_meta.json", "gyroscope_time.bin"),
-    ]
-imu_binaries_pairs: list[tuple[str, str]] = accelerometer_binaries_pairs + gyroscope_binaries_pairs
+    ("gyroscope_meta.json", "gyroscope_values.bin"),
+    ("gyroscope_meta.json", "gyroscope_time.bin"),
+]
+imu_binaries_pairs: list[tuple[str, str]] = (
+    accelerometer_binaries_pairs + gyroscope_binaries_pairs
+)
 
-path_to_assets = Path('src/paradigma/assets')
+path_to_assets = Path("src/paradigma/assets")
 
 
 def test_2_extract_features_gait_output(shared_datadir: Path):
@@ -52,14 +58,14 @@ def test_2_extract_features_gait_output(shared_datadir: Path):
     config = GaitConfig(step=data_type)
 
     extract_gait_features_io(
-        config=config, 
-        path_to_input=path_to_preprocessed_input, 
-        path_to_output=path_to_tested_output
+        config=config,
+        path_to_input=path_to_preprocessed_input,
+        path_to_output=path_to_tested_output,
     )
     compare_data(
-        reference_dir=path_to_reference_output, 
-        tested_dir=path_to_tested_output, 
-        binaries_pairs=gait_binaries_pairs
+        reference_dir=path_to_reference_output,
+        tested_dir=path_to_tested_output,
+        binaries_pairs=gait_binaries_pairs,
     )
 
 
[…]
