+++ src/paradigma/feature_extraction.py
@@ -1,7 +1,7 @@
-import numpy as np
-import pandas as pd
 from typing import List, Tuple
 
+import numpy as np
+import pandas as pd
 from scipy.integrate import cumulative_trapezoid
 from scipy.signal import find_peaks, windows
 from scipy.stats import kurtosis, skew
@@ -10,7 +10,9 @@ from sklearn.decomposition import PCA
 from paradigma.config import PulseRateConfig
 
 
-def compute_statistics(data: np.ndarray, statistic: str, abs_stats: bool=False) -> np.ndarray:
+def compute_statistics(
+    data: np.ndarray, statistic: str, abs_stats: bool = False
+) -> np.ndarray:
     """
     Compute a specific statistical measure along the timestamps of a 2D or 3D array.
 
@@ -29,7 +31,7 @@ def compute_statistics(data: np.ndarray, statistic: str, abs_stats: bool=False) 
         - 'kurtosis': Kurtosis.
         - 'skewness': Skewness.
     abs_stats : bool, optional
-        Whether to compute the statistics on the absolute values of the data for 
+        Whether to compute the statistics on the absolute values of the data for
         the mean and median (default: False).
 
     Returns
@@ -43,33 +45,42 @@ def compute_statistics(data: np.ndarray, statistic: str, abs_stats: bool=False) 
     ValueError
         If the specified `statistic` is not supported or if the input data has an invalid shape.
     """
-    if statistic not in ['mean', 'median', 'var', 'std', 'max', 'min', 'kurtosis', 'skewness']:
+    if statistic not in [
+        "mean",
+        "median",
+        "var",
+        "std",
+        "max",
+        "min",
+        "kurtosis",
+        "skewness",
+    ]:
         raise ValueError(f"Statistic '{statistic}' is not supported.")
-    
+
     if data.ndim > 3 or data.ndim < 2:
         raise ValueError("Input data must be a 1D, 2D or 3D array.")
 
-    if statistic == 'mean':
+    if statistic == "mean":
         if abs_stats:
             return np.mean(np.abs(data), axis=1)
         else:
             return np.mean(data, axis=1)
-    elif statistic == 'median':
+    elif statistic == "median":
         if abs_stats:
             return np.median(np.abs(data), axis=1)
         else:
             return np.median(data, axis=1)
-    elif statistic == 'var':
+    elif statistic == "var":
         return np.var(data, ddof=1, axis=1)
-    elif statistic == 'std':
+    elif statistic == "std":
         return np.std(data, axis=1)
-    elif statistic == 'max':
+    elif statistic == "max":
         return np.max(data, axis=1)
-    elif statistic == 'min':
+    elif statistic == "min":
         return np.min(data, axis=1)
-    elif statistic == 'kurtosis':
+    elif statistic == "kurtosis":
         return kurtosis(data, fisher=False, axis=1)
-    elif statistic == 'skewness':
+    elif statistic == "skewness":
         return skew(data, axis=1)
     else:
         raise ValueError(f"Statistic '{statistic}' is not supported.")
@@ -79,8 +90,8 @@ def compute_std_euclidean_norm(data: np.ndarray) -> np.ndarray:
     """
     Compute the standard deviation of the Euclidean norm for each window of sensor data.
 
-    The function calculates the Euclidean norm (L2 norm) across sensor axes for each 
-    timestamp within a window, and then computes the standard deviation of these norms 
+    The function calculates the Euclidean norm (L2 norm) across sensor axes for each
+    timestamp within a window, and then computes the standard deviation of these norms
     for each window.
 
     Parameters
[…]
