+++ tests/test_paradigma.py
@@ -1,10 +1,13 @@
 import tsdf
 
+
 def test_transform_time_array(shared_datadir):
     """
     The initial test to check if the preprocessing function works as expected. It checks the output dimensions and the type of the output.
     """
-    metadata_dict = tsdf.load_metadata_from_path(shared_datadir / '1.prepared_data/ppg/PPG_meta.json')
+    metadata_dict = tsdf.load_metadata_from_path(
+        shared_datadir / "1.prepared_data/ppg/PPG_meta.json"
+    )
 
     # Retrieve the metadata object we want, using the name of the binary as key
     metadata_values = metadata_dict["PPG_time.bin"]

[…]
