### 2025-11-24 #9502 @editorialbot
[REVIEW]: ParaDigMa: a Python toolbox for extracting Parkinson disease digital biomarkers from daily life wrist sensor data
**Submitting author:** @erikpostt (<a href="http://orcid.org/0000-0002-1937-9976">Erik Post</a>)
**Repository:** https://github.com/biomarkersParkinson/paradigma
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.1.1
