+++ pyproject.toml
@@ -16,11 +16,14 @@ readme = "README.md"
 [tool.poetry.dependencies]
 python = "^3.11"
 pandas = "^2.1.4"
-scikit-learn = ">=1.3.2,<1.6.1"
+scikit-learn = "^1.3.2"
 tsdf = "^0.5.2"
 pytype = "^2024.4.11"
 python-dateutil = "^2.9.0.post0"
 nbconvert = "^7.16.6"
+openmovement = { git = "https://github.com/digitalinteraction/openmovement-python.git", branch = "master" }
+pyarrow = "^22.0.0"
+avro = "^1.12.1"
 
 [tool.poetry.group.testing.dependencies]
 ipykernel = "^6.27.1"
