+++ pyproject.toml
@@ -18,7 +18,6 @@ python = "^3.11"
 pandas = "^2.1.4"
 scikit-learn = "^1.3.2"
 tsdf = { git = "https://github.com/biomarkersParkinson/tsdf.git", branch = "main" }
-pytype = "^2024.4.11"
 python-dateutil = "^2.9.0.post0"
 nbconvert = "^7.16.6"
 openmovement = { git = "https://github.com/digitalinteraction/openmovement-python.git", branch = "master" }
