+++ pyproject.toml
@@ -41,7 +41,6 @@ myst-parser = "^4.0.0"
 nbsphinx = "^0.9.6"
 
 [tool.poetry.group.dev.dependencies]
-pytype = "^2024.10.11"
 notebook = "^7.4.5"
 ipykernel = "^6.30.1"
 pre-commit = "^4.3.0"
