### 2025-11-28 adaa568
Increase efficiency of discarding segments
