### 2025-11-28 a8f32d5
Fix index issue in discarding segments
