### 2025-11-28 da95c6a
Fix index issue in discarding segments
