### 2025-12-01 e67e843
Simplify segmenting, fit tutorials to max line spacing, add device-specific tutorials to index
