+++ pyproject.toml
@@ -16,7 +16,7 @@ readme = "README.md"
 [tool.poetry.dependencies]
 python = "^3.11"
 pandas = "^2.1.4"
-scikit-learn = "^1.3.2"
+scikit-learn = ">=1.3.2,<1.6.1"
 tsdf = { git = "https://github.com/biomarkersParkinson/tsdf.git", branch = "main" }
 python-dateutil = "^2.9.0.post0"
 nbconvert = "^7.16.6"
