### 2025-12-03 443656b
Add axivity tutorial (#236)
* Release v1.0.3

* Enable build-and-test for all branches

* Enable build-and-test for all branches

* Sync main to release (#233)

* Incorporate modal tremor power in aggregate function

* Add tolerance for contiguous data to config

* Add pre-commit hooks and add Poetry scripts for simplifying dev process

* Remove segment categorization (set to legacy)

* Adjust pipelines to flexible column names and config updates

* Add info on scale sensitivity and make accelerometer optional

* Clarify installing pre-commit

* Update dependencies

* Update pathing in tutoria
[…]
