+++ pyproject.toml
@@ -17,10 +17,12 @@ readme = "README.md"
 python = "^3.11"
 pandas = "^2.1.4"
 scikit-learn = ">=1.3.2,<1.6.1"
-tsdf = "^0.5.2"
-pytype = "^2024.4.11"
+tsdf = { git = "https://github.com/biomarkersParkinson/tsdf.git", branch = "main" }
 python-dateutil = "^2.9.0.post0"
 nbconvert = "^7.16.6"
+openmovement = { git = "https://github.com/digitalinteraction/openmovement-python.git", branch = "master" }
+pyarrow = "^22.0.0"
+avro = "^1.12.1"
 
 [tool.poetry.group.testing.dependencies]
 ipykernel = "^6.27.1"
@@ -39,7 +41,6 @@ myst-parser = "^4.0.0"
 nbsphinx = "^0.9.6"
 
 [tool.poetry.group.dev.dependencies]
-pytype = "^2024.10.11"
 notebook = "^7.4.5"
 ipykernel = "^6.30.1"
 pre-commit = "^4.3.0"
