+++ paper/paper.md
@@ -40,8 +40,8 @@ authors:
       email: yordan.raykov@nottingham.ac.uk
       orcid: 0000-0002-8241-5087
       affiliation: "7, 8"
-    - given-names: Twan 
-      non-dropping-particle: van 
+    - given-names: Twan
+      non-dropping-particle: van
       family: Laarhoven
       orcid: 0000-0001-7597-0579
       affiliation: 2
@@ -50,7 +50,7 @@ authors:
       orcid: 0000-0002-8241-5087
       affiliation: "1, 2"
 affiliations:
-    - name: Center of Expertise for Parkinson and Movement Disorders, department of Neurology, Donders Institute for Brain, Cognition and Behaviour, Radboud University Medical Center, Nijmegen, the Netherlands
+    - name: Center of Expertise for Parkinson and Movement Disorders, Department of Neurology, Donders Institute for Brain, Cognition and Behaviour, Radboud University Medical Center, Nijmegen, the Netherlands
       index: 1
     - name: Institute for Computing and Information Sciences, Radboud University, Nijmegen, the Netherlands
       index: 2
@@ -72,25 +72,25 @@ bibliography: paper.bib
 
 # Summary
 
-ParaDigMa is an open-source Python package designed to extract digital biomarkers of Parkinson disease (PD) from wrist-worn sensor data collected during passive monitoring in daily life [@postParaDigMaToolboxDeriving2025a]. It processes raw accelerometer, gyroscope and photoplethysmography (PPG) signals using modular pipelines which address common challenges in real-world wearable sensor data analysis. The package currently targets three clinically relevant symptom domains: (1) tremor, (2) arm swing during gait, and (3) pulse rate (to derive digital biomarkers related to autonomic dysfunction). ParaDigMa enables researchers to derive interpretable and clinically meaningful digital biomarkers to monitor PD in daily life.
+PARkinson disease DIGital MArkers (ParaDigMa) is an open-source Python package designed to extract digital biomarkers of Parkinson disease (PD) from wrist-worn sensor data collected during passive monitoring in daily life [@postParaDigMaToolboxDeriving2025a]. It processes raw accelerometer, gyroscope, and photoplethysmography (PPG) signals using modular pipelines which address common challenges in real-world wearable sensor data analysis. The package currently targets three clinically relevant symptom domains: (1) tremor, (2) arm swing during gait, and (3) pulse rate (to derive digital biomarkers related to autonomic dysfunction). ParaDigMa enables researchers to derive interpretable and clinically meaningful digital biomarkers to monitor PD in daily life.
 
 # Statement of need
 
-The use of wearable sensors to monitor PD in daily life has grown exponentially over the past two decades [@debSystematicSurveyResearch2022], with mounting evidence supporting their utility for diagnosis, monitoring disease progression, and evaluating treatment efficacy [@sunDigitalBiomarkersPrecision2024]. However, the extraction of digital biomarkers from raw sensor signals often relies on proprietary or non-replicable pipelines applied to individual cohorts. To support their adoption in clinical and research practice, and to enable meaningful comparisons across studies, reproducible and standardized data processing workflows are essential [@sunDigitalBiomarkersPrecision2024; @boyleActivityBehavioralRecognition2025]. 
[…]
