### 2025-12-17 c3bc712
standardize columns in tremor analysis
