+++ src/paradigma/config.py
@@ -178,18 +178,18 @@ class GaitConfig(IMUConfig):
             "accelerometer_x_grav_std": DataUnits.GRAVITY,
             "accelerometer_y_grav_std": DataUnits.GRAVITY,
             "accelerometer_z_grav_std": DataUnits.GRAVITY,
-            "accelerometer_x_power_below_gait": DataUnits.POWER_SPECTRAL_DENSITY,
-            "accelerometer_y_power_below_gait": DataUnits.POWER_SPECTRAL_DENSITY,
-            "accelerometer_z_power_below_gait": DataUnits.POWER_SPECTRAL_DENSITY,
-            "accelerometer_x_power_gait": DataUnits.POWER_SPECTRAL_DENSITY,
-            "accelerometer_y_power_gait": DataUnits.POWER_SPECTRAL_DENSITY,
-            "accelerometer_z_power_gait": DataUnits.POWER_SPECTRAL_DENSITY,
-            "accelerometer_x_power_tremor": DataUnits.POWER_SPECTRAL_DENSITY,
-            "accelerometer_y_power_tremor": DataUnits.POWER_SPECTRAL_DENSITY,
-            "accelerometer_z_power_tremor": DataUnits.POWER_SPECTRAL_DENSITY,
-            "accelerometer_x_power_above_tremor": DataUnits.POWER_SPECTRAL_DENSITY,
-            "accelerometer_y_power_above_tremor": DataUnits.POWER_SPECTRAL_DENSITY,
-            "accelerometer_z_power_above_tremor": DataUnits.POWER_SPECTRAL_DENSITY,
+            "accelerometer_x_power_below_gait": DataUnits.POWER_SPECTRAL_DENSITY_ACC,
+            "accelerometer_y_power_below_gait": DataUnits.POWER_SPECTRAL_DENSITY_ACC,
+            "accelerometer_z_power_below_gait": DataUnits.POWER_SPECTRAL_DENSITY_ACC,
+            "accelerometer_x_power_gait": DataUnits.POWER_SPECTRAL_DENSITY_ACC,
+            "accelerometer_y_power_gait": DataUnits.POWER_SPECTRAL_DENSITY_ACC,
+            "accelerometer_z_power_gait": DataUnits.POWER_SPECTRAL_DENSITY_ACC,
+            "accelerometer_x_power_tremor": DataUnits.POWER_SPECTRAL_DENSITY_ACC,
+            "accelerometer_y_power_tremor": DataUnits.POWER_SPECTRAL_DENSITY_ACC,
+            "accelerometer_z_power_tremor": DataUnits.POWER_SPECTRAL_DENSITY_ACC,
+            "accelerometer_x_power_above_tremor": DataUnits.POWER_SPECTRAL_DENSITY_ACC,
+            "accelerometer_y_power_above_tremor": DataUnits.POWER_SPECTRAL_DENSITY_ACC,
+            "accelerometer_z_power_above_tremor": DataUnits.POWER_SPECTRAL_DENSITY_ACC,
             "accelerometer_x_dominant_frequency": DataUnits.FREQUENCY,
             "accelerometer_y_dominant_frequency": DataUnits.FREQUENCY,
             "accelerometer_z_dominant_frequency": DataUnits.FREQUENCY,
@@ -266,11 +266,12 @@ class TremorConfig(IMUConfig):
         if step == "features":
             self.d_channels_values: Dict[str, str] = {}
             for mfcc_coef in range(1, self.n_coefficients_mfcc + 1):
-                self.d_channels_values[f"mfcc_{mfcc_coef}"] = "unitless"
+                self.d_channels_values[f"mfcc_{mfcc_coef}"] = DataUnits.NONE
 
[…]
