### 2025-12-19 cdede16
change unit tremor power
