### 2025-12-22 10988a0
Add flexibility for tremor columns (#241)
* fix bug resampling

* fix bug select_days function

* json requires double quotes

* undo changes json

* incorporate modal tremor power in aggregate function

* Incorporate suggestions

* fix type

* clarify variables

* add tolerance to config

* change preprocess_ppg_data too

* add suggestions

* run one segment for pulse rate only

* fix typos

* add suggestion

* standardize columns in tremor analysis

* pre-commit hooks

* update poetry dependencies

* update md files

* change unit tremor power

* update lock file

* update markdown files

* 
[…]
