### 2025-12-23 ed01fff
Add cross-OS testing, split by test/quality
