### 2025-12-23 1d150ad
Ensure consistency in docstrings
