+++ src/paradigma/pipelines/gait_pipeline.py
@@ -39,42 +39,31 @@ if not logger.hasHandlers():
 
 def extract_gait_features(df: pd.DataFrame, config: GaitConfig) -> pd.DataFrame:
     """
-    Extracts gait features from accelerometer and gravity sensor data in the input DataFrame by computing temporal and spectral features.
+    Extracts gait features from accelerometer and gravity sensor data in the input DataFrame by
+    computing temporal and spectral features.
 
     This function performs the following steps:
     1. Groups sequences of timestamps into windows, using accelerometer and gravity data.
-    2. Computes temporal domain features such as mean and standard deviation for accelerometer and gravity data.
-    3. Transforms the signals from the temporal domain to the spectral domain using the Fast Fourier Transform (FFT).
+    2. Computes temporal domain features such as mean and standard deviation for accelerometer and
+    gravity data.
+    3. Transforms the signals from the temporal domain to the spectral domain using the Fast
+    Fourier Transform (FFT).
     4. Computes spectral domain features for the accelerometer data.
     5. Combines both temporal and spectral features into a final DataFrame.
 
     Parameters
     ----------
     df : pd.DataFrame
-        The input DataFrame containing gait data, which includes time, accelerometer, and gravity sensor data. The data should be
-        structured with the necessary columns as specified in the `config`.
-
-    onfig : GaitConfig
-        Configuration object containing parameters for feature extraction, including column names for time, accelerometer data, and
-        gravity data, as well as settings for windowing, and feature computation.
+        Input DataFrame containing gait data, including time, accelerometer, and gravity
+        sensor columns.
+    config : GaitConfig
+        Configuration object with parameters for feature extraction (columns names, windows
+        settings, etc.)
 
     Returns
     -------
     pd.DataFrame
-        A DataFrame containing extracted gait features, including temporal and spectral domain features. The DataFrame will have
-        columns corresponding to time, statistical features of the accelerometer and gravity data, and spectral features of the
-        accelerometer data.
-
-    Notes
-    -----
-    - This function groups the data into windows based on timestamps and applies Fast Fourier Transform to compute spectral features.
-    - The temporal features are extracted from the accelerometer and gravity data, and include statistics like mean and standard deviation.
-    - The input DataFrame must include columns as specified in the `config` object for proper feature extraction.
-
-    Raises
-    ------
-    ValueError
-        If the input DataFrame does not contain the required columns as specified in the configuration or if any step in the feature extraction fails.
+        DataFrame containing temporal and spectral gait features.
     """
     # Group sequences of timestamps into windows
     windowed_colnames = (
@@ -127,14 +116,10 @@ def detect_gait(
     df: pd.DataFrame, clf_package: ClassifierPackage, parallel: bool = False
 ) -> pd.Series:
     """
-    Detects gait activity in the input DataFrame using a pre-trained classifier and applies a threshold to classify results.
+    Detects gait activity in the input DataFrame using a pre-trained classifier.
 
-    This function performs the following steps:
-    1. Loads the pre-trained classifier and scaling parameters from the specified directory.
-    2. Scales the relevant features in the input DataFrame (`df`) using the loaded scaling parameters.
-    3. Predicts the probability of gait activity for each sample in the DataFrame using the classifier.
-    4. Applies a threshold to the predicted probabilities to determine whether gait activity is present.
-    5. Returns predicted probabilities
+    Features are scaled, probabilities predicted, and a threshold applied to classify
+    samples as gait activity or not. Supports optional parallel processing.
 
[…]
