+++ src/paradigma/preprocessing.py
@@ -24,36 +24,30 @@ def resample_data(
     Parameters
     ----------
     df : pd.DataFrame
-        The input DataFrame containing the sensor data.
+        Input DataFrame containing the sensor data.
     time_column : str
-        The name of the column containing the time data.
+        Name of column containing the time data.
     values_column_names : List[str]
-        A list of column names that should be resampled.
+        Names of columns to be resampled.
     sampling_frequency : int
-        The original sampling frequency of the data (in Hz).
+        Original sampling frequency of the data (in Hz).
     resampling_frequency : int
-        The frequency to which the data should be resampled (in Hz).
+        Frequency to which the data should be resampled (in Hz).
     tolerance : float, optional
-        The tolerance added to the expected difference when checking
-        for contiguous timestamps. If not provided, it defaults to the tolerance specified in IMUConfig.
+        Tolerance added to the expected difference when checking for contiguous
+        timestamps. If not provided, it defaults to the tolerance specified in IMUConfig.
 
     Returns
     -------
     pd.DataFrame
-        A DataFrame with the resampled data, where each column contains resampled values.
+        Resampled data, where each column contains resampled values.
         The time column will reflect the new resampling frequency.
-
-    Raises
-    ------
-    ValueError
-        If the time array is not strictly increasing.
 
     Notes
     -----
     - Uses cubic interpolation for smooth resampling if there are enough points.
     - If only two timestamps are available, it falls back to linear interpolation.
     """
-
     # Set default tolerance if not provided to tolerance specified in IMUConfig
     if tolerance is None:
         tolerance = IMUConfig().tolerance
@@ -108,38 +102,33 @@ def butterworth_filter(
     """
     Applies a Butterworth filter to 1D or 2D sensor data.
 
-    This function applies a low-pass, high-pass, or band-pass Butterworth filter to the
+    Applies a low-pass, high-pass, or band-pass Butterworth filter to the
     input data. The filter is designed using the specified order, cutoff frequency,
     and passband type. The function can handle both 1D and 2D data arrays.
 
     Parameters
     ----------
     data : np.ndarray
-        The sensor data to be filtered. Can be 1D (e.g., a single signal) or 2D
+        Sensor data to be filtered. Can be 1D (e.g., a single signal) or 2D
         (e.g., multi-axis sensor data).
     order : int
-        The order of the Butterworth filter. Higher values result in a steeper roll-off.
+        Order of the Butterworth filter. Higher values result in a steeper roll-off.
     cutoff_frequency : float or List[float]
-        The cutoff frequency (or frequencies) for the filter. For a low-pass or high-pass filter,
+        Cutoff frequency (or frequencies) for the filter. For a low-pass or high-pass filter,
         this is a single float. For a band-pass filter, this should be a list of two floats,
         specifying the lower and upper cutoff frequencies.
     passband : str
-        The type of passband to apply. Options are:
+        Type of passband to apply. Options are:
         - 'hp' : high-pass filter
         - 'lp' : low-pass filter
         - 'band' : band-pass filter
     sampling_frequency : int
-        The sampling frequency of the data in Hz. This is used to normalize the cutoff frequency.
+        Sampling frequency of the data in Hz. This is used to normalize the cutoff frequency.
 
     Returns
     -------
     np.ndarray
         The filtered sensor data. The shape of the output is the same as the input data.
-
-    Raises
-    ------
-    ValueError
-        If the input data has more than two dimensions, or if an invalid passband is specified.
 
     Notes
     -----
@@ -174,32 +163,31 @@ def preprocess_imu_data(
     Parameters
     ----------
     df : pd.DataFrame
[…]

[…]
