### 2025-12-23 ec88a83
Modify ruff for linting
