### 2025-12-23 #9502 @Erikpostt
@Asanto32, thank you again for taking the time to review the repository and manuscript.

**1. Support is claimed for Python 3.11+ but in the unit tests the only tested python version is 3.12.**
Thank you for catching this inconsistency. We have updated the CI configuration to explicitly test the package on both Python 3.11 and 3.12, aligning the test coverage with the stated supported Python versions.

**2. Additionally, there should be testing across multiple OS in the CI.**
Thank you for the suggestion. We have updated the CI to run the test suite across multiple operating systems (Linux, macOS, Windows) and multiple Python versions (3.11 and 3.12). Quality checks and builds are run separately, as they are OS-independent. Formatting, linting, and documentation builds are OS-agnostic and are therefore run once on Linux, while functional tests are executed across all targeted OS and Python combinations.

**3. numpy version in the lock file is 1.26 but Python 3.13 requires > 2.0, this may cause issues.**
Good point. We have updated the dependencies accordingly and will ensure these remain current in future releases.

**4. Inconsistent docstrings, in some cases “Raises” is included when Errors are raised, other times it is missing.**
Thank you for noticing. We have standardized the docstrings by removing the “Raises” sections for consistency. We may later revisit this decision to include them uniformly where appropriate.

**5. Linting could be quite useful for improving readability of the codebase.**
We appreciate the suggestion. Our pre-commit hooks already include Black, isort, and Ruff, which enforce consistent formatting and linting across the codebase. We believe major readability guidelines are covered automatically. Are there specific formatting issues you noticed that are not currently handled?

**6. There are quite a few places where there are “TODO” inline comments, which feels like something that should only be included in a development branch.**
You are correct. We have removed inline TODOs and moved them to our backlog.

**7. I feel like the claims about using this toolbox for raw sensor data/device agnostic processing are not accurate. I tried feeding in data I have from GENEActiv and Actigraph watches, however this would require me to extract data, separate the data from the timestamps, and create specific metadata json files. This requires raw data to be heavily modified into, you can’t simply provide a path to a “.bin” or “.gt3x” file, for example. I understand that in theory I just need a pandas dataframe to run the analysis, but I feel like some of the claims about only needing raw sensor data should be revisited.**
Thank you for testing Paradigma with your own data. The requirement is indeed a pandas DataFrame; the tutorial examples using `tsdf` files are just one format. Device-agnostic claims refer to the fact that results are consistent across different devices, not that every raw file type is automatically supported. Do you think that clarifying these claims is sufficient? With "... but I feel like some of the claims about only needing raw sensor data should be revisited", do you mean that in fact _Paradigma_ requires a pandas DataFrame in memory?

**8. My biggest issue is that there is no Python or CLI entrypoint. If this was truly a modular pipeline, we should be able to easily point to raw sensor data, and call any specific or combination of analysis pipelines. As it stands, unless I am missing something, a user would need to follow what seems to be quite a long series of notebooks in the tutorial. Perhaps I am not the target audience, but I feel this would make the toolbox much more usable.**
Great point! We agree this would improve usability. Currently, the target audience is researchers with minor Python experience, following tutorials to run analyses. Developing Python/CLI entrypoints is on our roadmap, and of high priority. Would you consider this a required feature for publication, or acceptable to implement shortly afterward?

**9. There are broken links in the webpage tutorial section.**
Thank you for pointing this out. We have corrected these links and are awaiting the upcoming release to reflect the changes.
