### 2025-12-23 2a05275
Update np.trapz to np.trapezoid for numpy v2.X
