### 2025-12-23 biomarkersParkinson/paradigma#243 @Erikpostt
Add Python/CLI entrypoint
As a user, I want to run pipelines using, e.g., `python run_gait_pipeline()` in CLI. Also, I want to point at a directory/file and run the pipeline without having to load in data.

Essentially, Notebooks should be examples, not the only way to run the toolbox.

**Describe the solution you'd like**
A high-level entrypoint function bridging pipelines (e.g., `run_pipeline(domain='gait')`.
* a Python function that can be imported and called, and/or
* a CLI command (`paradigma run …`)

Then, for modularity, the user should be able to:
* point to raw sensor data
* choose which pipeline(s) to run (gait, arm activity, etc.)
* run them without reading notebooks

For example:

```python
from paradigma import run_pipeline

run_pipeline(
    data_path="path/to/raw/data",
    pipelines=["gait", "arm_activity"],
    config="default",  # or config object / file
    output_dir="results/"
)
```

The function should live in in an obvious place, `paradigma/pipeline.py` or `paradigma/__init__.py`. For it to be an endpoint, we also need to add a console script in `pyproject.toml`:

```toml
[project.scripts]
paradigma = "paradigma.cli:main"
```

**Additional context**
This is a feature request raised by a reviewer for the JOSS paper: https://github.com/openjournals/joss-reviews/issues/9502.
