### 2025-12-24 biomarkersParkinson/paradigma#243 @Erikpostt
This also requires us to think about possible usages by users:
* Run a single file of a single subject
* Run multiple files in a directory of a single subject
* Run multiple files and multiple subjects

This does only allow serial computation, so we could also consider setting up a skeleton for parallel processing.
