### 2026-01-06 #9502 @Erikpostt
> ⚠️ [An error happened when generating the pdf](https://github.com/openjournals/joss-papers/actions/runs/20745045845?check_suite_focus=true).

@Kevin-Mattheus-Moerman the PDF build fails in `xuanxu/upload-files-action` during branch creation via Octokit. This appears to be a GitHub Actions permission issue rather than a problem with the manuscript. Are you familiar with this issue?
