### 2026-01-07 8af0137
Reframe index ToC, add dev option for building docs, add new preprints to readme
