+++ scripts/build_docs.py
@@ -1,3 +1,4 @@
+import argparse
 import subprocess
 import sys
 from pathlib import Path
@@ -6,7 +7,26 @@ from nbclient import NotebookClient
 from nbformat import read, write
 
 
+def run(cmd, **kwargs):
+    subprocess.run(cmd, check=True, **kwargs)
+
+
 def main():
+    parser = argparse.ArgumentParser()
+    parser.add_argument(
+        "--dev",
+        action="store_true",
+        help="Fast dev build: do not execute notebooks",
+    )
+
+    parser.add_argument(
+        "--no-nbconvert",
+        action="store_true",
+        help="Skip nbconvert step (assumes markdown already exists)",
+    )
+
+    args = parser.parse_args()
+
     project_root = Path(__file__).resolve().parent.parent
     tutorials_dir = project_root / "docs/tutorials"
     static_dir = tutorials_dir / "_static"
@@ -18,39 +38,50 @@ def main():
         return
 
     for nb_path in notebooks:
-        print(f"Executing {nb_path}...")
-        with nb_path.open("r", encoding="utf-8") as f:
-            nb = read(f, as_version=4)
+        md_path = static_dir / f"{nb_path.stem}.md"
 
-        client = NotebookClient(
-            nb, timeout=600, kernel_name=f"python{sys.version_info.major}"
-        )
-        client.execute(cwd=nb_path.parent)
+        if not args.dev:
+            print(f"Executing {nb_path}...")
+            with nb_path.open("r", encoding="utf-8") as f:
+                nb = read(f, as_version=4)
 
-        with nb_path.open("w", encoding="utf-8") as f:
-            write(nb, f)
+            client = NotebookClient(
+                nb, timeout=600, kernel_name=f"python{sys.version_info.major}"
+            )
+            client.execute(cwd=nb_path.parent)
 
-        print(f"Exporting {nb_path} to markdown...")
-        subprocess.run(
-            [
-                sys.executable,
-                "-m",
-                "jupyter",
-                "nbconvert",
-                "--to",
-                "markdown",
-                str(nb_path),
-                "--output-dir",
-                str(static_dir),
-            ],
-            check=True,
-        )
+            with nb_path.open("w", encoding="utf-8") as f:
+                write(nb, f)
+        else:
+            print(f"[DEV] Skipping execution of {nb_path}")
 
-        print(f"Stripping outputs from {nb_path}...")
-        subprocess.run([sys.executable, "-m", "nbstripout", str(nb_path)], check=True)
+        if not args.no_nbconvert:
+            if args.dev and md_path.exists():
+                print(f"[DEV] Using existing markdown for {nb_path}")
+            else:
+                print(f"Exporting {nb_path} to markdown...")
+                run(
+                    [
+                        sys.executable,
+                        "-m",
+                        "jupyter",
+                        "nbconvert",
+                        "--to",
+                        "markdown",
+                        str(nb_path),
+                        "--output-dir",
+                        str(static_dir),
+                    ]
+                )
+        else:
+            print(f"[DEV] Skipping nbconvert for {nb_path}")
 
-    print("Building Sphinx docs...")
-    subprocess.run(
+        if not args.dev:
+            print(f"Stripping outputs from {nb_path}...")
+            run([sys.executable, "-m", "nbstripout", str(nb_path)])
+
+        print("Building Sphinx docs...")
+    run(
         [
             sys.executable,
             "-m",
@@ -59,7 +90,7 @@ def main():
             "html",
             str(project_root / "docs"),
             str(project_root / "docs/_build/html"),
-        ],
-        check=True,
+        ]
     )
+
     print("Done!")
