+++ src/paradigma/feature_extraction.py
@@ -146,9 +146,10 @@ def compute_power_in_bandwidth(
     spectral_resolution : float, optional
         The spectral resolution of the PSD in Hz (default: 1).
     cumulative_sum_method : str, optional
-        The method used to integrate the PSD over the frequency band. Supported values are:
+        The method used to integrate the PSD over the frequency band (default: 'trapz').
+        Supported values are:
         - 'trapz': Trapezoidal rule.
-        - 'sum': Simple summation (default: 'trapz').
+        - 'sum': Simple summation.
 
     Returns
     -------
@@ -169,7 +170,7 @@ def compute_power_in_bandwidth(
         masked_psd = psd[:, band_mask, :]
 
     if cumulative_sum_method == "trapz":
-        band_power = spectral_resolution * np.trapz(
+        band_power = spectral_resolution * np.trapezoid(
             masked_psd, freqs[band_mask], axis=1
         )
     elif cumulative_sum_method == "sum":
@@ -410,7 +411,7 @@ def compute_relative_power(
         for peak_freq in peak_freqs
     ]
     rel_power = [
-        np.trapz(psd[j, idx], freqs[idx]) / np.trapz(psd[j, :], freqs)
+        np.trapezoid(psd[j, idx], freqs[idx]) / np.trapezoid(psd[j, :], freqs)
         for j, idx in enumerate(dom_band_idx)
     ]
     return rel_power
