+++ docs/tutorials/index.md
@@ -1,0 +1,14 @@
+## Tutorials
+
+This section contains step-by-step tutorials demonstrating how to use
+the ParaDigMa toolbox to extract digital measures from wrist sensor data.
+
+---
+
+**Available tutorials**
+
+- {doc}`Device-specific data loading <_static/device_specific_data_loading>`
+- {doc}`Data preparation <_static/data_preparation>`
+- {doc}`Gait analysis <_static/gait_analysis>`
+- {doc}`Tremor analysis <_static/tremor_analysis>`
+- {doc}`Pulse rate analysis <_static/pulse_rate_analysis>`
