### 2026-01-09 09543a2
Remove redundant text
