### 2026-01-09 0a562aa
Rewrite data loading
