### 2026-01-09 1901870
Update readme with relative urls
