+++ CONTRIBUTING.md
@@ -102,6 +102,10 @@ poetry run build-docs
 poetry run serve-docs
 ```
 
+`build-docs` accepts arguments `--dev` to skip execution of notebooks and `--no-nbconvert` to skip
+conversion of notebooks to markdown. To speeds up development significantly if no changes are
+made to notebooks.
+
 6. Commit your changes and open a pull request.
 
 ## Pull Request Guidelines
