### 2026-01-09 e0cf685
Change relative paths to absolute for anchored paths
