+++ docs/index.md
@@ -28,9 +28,3 @@ license.md
 
 autoapi/index
 ```
-
-```{toctree}
-:maxdepth: 1
-:hidden:
-autoapi/paradigma/util/index
-```
