+++ docs/tutorials/index.md
@@ -7,8 +7,8 @@ the ParaDigMa toolbox to extract digital measures from wrist sensor data.
 
 **Available tutorials**
 
-- {doc}`Device-specific data loading <_static/device_specific_data_loading>`
-- {doc}`Data preparation <_static/data_preparation>`
-- {doc}`Gait analysis <_static/gait_analysis>`
-- {doc}`Tremor analysis <_static/tremor_analysis>`
-- {doc}`Pulse rate analysis <_static/pulse_rate_analysis>`
+- [Device-specific data loading](device_specific_data_loading)
+- [Data preparation](data_preparation)
+- [Gait analysis](gait_analysis)
+- [Tremor analysis](tremor_analysis)
+- [Pulse rate analysis](pulse_rate_analysis)
