### 2026-01-09 216f07c
Process co-pilot suggestions
