### 2026-01-09 dfbabca
Remove inaccurate claims
