+++ paper/paper.md
@@ -72,7 +72,7 @@ bibliography: paper.bib
 
 # Summary
 
-PARkinson's disease DIGital bioMArkers (ParaDigMa) is an open-source Python package designed to extract digital biomarkers of Parkinson's disease (PD) from wrist-worn sensor data collected during passive monitoring in daily life [@postParaDigMaToolboxDeriving2025a]. It processes raw accelerometer, gyroscope, and photoplethysmography (PPG) signals using modular pipelines which address common challenges in real-world wearable sensor data analysis. The package currently targets three clinically relevant symptom domains: (1) tremor, (2) arm swing during gait, and (3) pulse rate (to derive digital biomarkers related to autonomic dysfunction). ParaDigMa enables researchers to derive interpretable and clinically meaningful digital biomarkers to monitor PD in daily life.
+PARkinson's disease DIGital bioMArkers (ParaDigMa) is an open-source Python package designed to extract digital biomarkers of Parkinson's disease (PD) from wrist-worn sensor data collected during passive monitoring in daily life [@postParaDigMaToolboxDeriving2025a]. It processes accelerometer, gyroscope, and photoplethysmography (PPG) signals using modular pipelines which address common challenges in real-world wearable sensor data analysis. The package currently targets three clinically relevant symptom domains: (1) tremor, (2) arm swing during gait, and (3) pulse rate (to derive digital biomarkers related to autonomic dysfunction). ParaDigMa enables researchers to derive interpretable and clinically meaningful digital biomarkers to monitor PD in daily life.
 
 # Statement of need
 
@@ -90,11 +90,11 @@ Beyond PD-specific toolboxes , ParaDigMa also differs from general-purpose toolb
 
 # Architecture
 
-ParaDigMa is a Python toolbox for processing and analyzing wrist-worn sensor data. Its architecture, illustrated in Figure 1, is designed for flexibility and ease of use: each processing step is implemented as a standalone function, enabling users to run complete pipelines through a small number of sequential function calls. The toolbox supports data from inertial measurement units (IMUs) and/or photoplethysmography (PPG) sensors, and is device-agnostic, accepting input with minimal requirements (specified in its documentation).
+ParaDigMa is a Python toolbox for processing and analyzing wrist-worn sensor data. Its architecture, illustrated in Figure 1, is designed for flexibility and ease of use: each processing step is implemented as a standalone function, enabling users to run complete pipelines through a small number of sequential function calls. The toolbox supports data from inertial measurement units (IMUs) and/or photoplethysmography (PPG) sensors, accepting input with minimal requirements (specified in its documentation).
 
 ![A schematic overview of the architecture of ParaDigMa. Horizontal bars indicate categories of input or output data, with rectangles representing specific data objects. Data processing steps, shown as blue ovals, follow a consistent structure across domains. Arrows represent the sequential flow of operations, and colored columns distinguish the individual processing domains. ](imgs/pipeline-architecture.png)
 
-From top to bottom, data are processed as follows. First, raw sensor data are preprocessed through resampling, interpolation, and filtering (1. Preprocessing). The resulting time series are segmented into fixed-size windows, from which temporal and spectral features are extracted (2. Feature extraction). These features are then passed to domain-specific classifiers for the detection of relevant behavior (e.g., gait) or symptoms (e.g., tremor) (3. Classification). These predictions are combined with the preprocessed data to quantify the detected behavior (e.g., arm swing range of motion during gait) or symptoms (e.g., tremor power) (4. Quantification). The last step (5. Aggregation) contains functionalities to aggregate the output from steps 3 and 4 across relevant time scales to derive digital biomarkers (e.g., weekly typical tremor power).
[…]
