### 2026-01-09 1d39492
Change str trapz to trapezoid for interpretability
