### 2026-01-10 biomarkersParkinson/paradigma#250 @ntolley
BUG: JSON decode error in tutorial documents
**Describe the bug**
A fresh install of paradigma using either `pip` or `uv` produces a JSON decode error. This was done on two different linux operating systems with python 3.11 and 3.12.

**To Reproduce**
Steps to reproduce the behavior:
Install package on a clean environment and run the first code block of the `pulse_rate_analysis.ipynb` notebook (or any of the other analysis notebooks)

**Expected behavior**
The code block should run without error

**Screenshots**
```
JSONDecodeError                           Traceback (most recent call last)
Cell In[1], line 13
      9 imu_prefix = 'imu'
     11 segment_nr = '0001'
---> 13 df_ppg, metadata_time_ppg, metadata_values_ppg = load_tsdf_dataframe(
     14     path_to_data=path_to_prepared_data [/](https://vscode-remote+ood-002eccv-002ebrown-002eedu.vscode-resource.vscode-cdn.net/) ppg_prefix,
     15     prefix=f'PPG_segment{segment_nr}'
     16 )
     18 # Only relevant if you have IMU data available
     19 df_imu, metadata_time_imu, metadata_values_imu = load_tsdf_dataframe(
     20     path_to_data=path_to_prepared_data [/](https://vscode-remote+ood-002eccv-002ebrown-002eedu.vscode-resource.vscode-cdn.net/) imu_prefix,
     21     prefix=f'IMU_segment{segment_nr}'
     22 )

File [~/.conda/envs/para_test/lib/python3.12/site-packages/paradigma/util.py:178](https://vscode-remote+ood-002eccv-002ebrown-002eedu.vscode-resource.vscode-cdn.net/users/ntolley/grad_school/paradigma/docs/tutorials/~/.conda/envs/para_test/lib/python3.12/site-packages/paradigma/util.py:178), in load_tsdf_dataframe(path_to_data, prefix, meta_suffix, time_suffix, values_suffix)
    175 time_filename = f"{prefix}_{time_suffix}"
    176 values_filename = f"{prefix}_{values_suffix}"
--> 178 metadata_time, metadata_values = read_metadata(
    179     path_to_data, meta_filename, time_filename, values_filename
    180 )
    181 df = tsdf.load_dataframe_from_binaries(
    182     [metadata_time, metadata_values], tsdf.constants.ConcatenationType.columns
    183 )
    185 return df, metadata_time, metadata_values

File [~/.conda/envs/para_test/lib/python3.12/site-packages/paradigma/util.py:159](https://vscode-remote+ood-002eccv-002ebrown-002eedu.vscode-resource.vscode-cdn.net/users/ntolley/grad_school/paradigma/docs/tutorials/~/.conda/envs/para_test/lib/python3.12/site-packages/paradigma/util.py:159), in read_metadata(input_path, meta_filename, time_filename, values_filename)
    156 def read_metadata(
    157     input_path: str, meta_filename: str, time_filename: str, values_filename: str
    158 ) -> Tuple[TSDFMetadata, TSDFMetadata]:
--> 159     metadata_dict = tsdf.load_metadata_from_path(
    160         os.path.join(input_path, meta_filename)
    161     )
    162     metadata_time = metadata_dict[time_filename]
    163     metadata_values = metadata_dict[values_filename]

File [~/.conda/envs/para_test/lib/python3.12/site-packages/tsdf/read_tsdf.py:84](https://vscode-remote+ood-002eccv-002ebrown-002eedu.vscode-resource.vscode-cdn.net/users/ntolley/grad_school/paradigma/docs/tutorials/~/.conda/envs/para_test/lib/python3.12/site-packages/tsdf/read_tsdf.py:84), in load_metadata_from_path(path)
     82 # The data is isomorphic to a JSON
     83 with open(path, "r") as file:
---> 84     data = json.load(file)
     86 abs_path = os.path.realpath(path)
     87 # Parse the data and verify that it complies with TSDF requirements

File [~/.conda/envs/para_test/lib/python3.12/json/__init__.py:293](https://vscode-remote+ood-002eccv-002ebrown-002eedu.vscode-resource.vscode-cdn.net/users/ntolley/grad_school/paradigma/docs/tutorials/~/.conda/envs/para_test/lib/python3.12/json/__init__.py:293), in load(fp, cls, object_hook, parse_float, parse_int, parse_constant, object_pairs_hook, **kw)
    274 def load(fp, *, cls=None, object_hook=None, parse_float=None,
    275         parse_int=None, parse_constant=None, object_pairs_hook=None, **kw):
    276     """Deserialize ``fp`` (a ``.read()``-supporting file-like object containing
    277     a JSON document) to a Python object.
    278
   (...)    291     kwarg; otherwise ``JSONDecoder`` is used.
    292     """
--> 293     return loads(fp.read(),
    294         cls=cls, object_hook=object_hook,
    295         parse_float=parse_float, parse_int=parse_int,
    296         parse_constant=parse_constant, object_pairs_hook=object_pairs_hook, **kw)

File [~/.conda/envs/para_test/lib/python3.12/json/__init__.py:346](https://vscode-remote+ood-002eccv-002ebrown-002eedu.vscode-resource.vscode-cdn.net/users/ntolley/grad_school/paradigma/docs/tutorials/~/.conda/envs/para_test/lib/python3.12/json/__init__.py:346), in loads(s, cls, object_hook, parse_float, parse_int, parse_constant, object_pairs_hook, **kw)
    341     s = s.decode(detect_encoding(s), 'surrogatepass')
    343 if (cls is None and object_hook is None and
    344         parse_int is None and parse_float is None and
    345         parse_constant is None and object_pairs_hook is None and not kw):
--> 346     return _default_decoder.decode(s)
    347 if cls is None:
    348     cls = JSONDecoder

File [~/.conda/envs/para_test/lib/python3.12/json/decoder.py:338](https://vscode-remote+ood-002eccv-002ebrown-002eedu.vscode-resource.vscode-cdn.net/users/ntolley/grad_school/paradigma/docs/tutorials/~/.conda/envs/para_test/lib/python3.12/json/decoder.py:338), in JSONDecoder.decode(self, s, _w)
    333 def decode(self, s, _w=WHITESPACE.match):
    334     """Return the Python representation of ``s`` (a ``str`` instance
    335     containing a JSON document).
    336
    337     """
--> 338     obj, end = self.raw_decode(s, idx=_w(s, 0).end())
    339     end = _w(s, end).end()
    340     if end != len(s):

File [~/.conda/envs/para_test/lib/python3.12/json/decoder.py:356](https://vscode-remote+ood-002eccv-002ebrown-002eedu.vscode-resource.vscode-cdn.net/users/ntolley/grad_school/paradigma/docs/tutorials/~/.conda/envs/para_test/lib/python3.12/json/decoder.py:356), in JSONDecoder.raw_decode(self, s, idx)
    354     obj, end = self.scan_once(s, idx)
    355 except StopIteration as err:
--> 356     raise JSONDecodeError("Expecting value", s, err.value) from None
    357 return obj, end

JSONDecodeError: Expecting value: line 1 column 1 (char 0)
```
