### 2026-01-11 4caeb1e
Add readme segments, create getting started guide
