### 2026-01-11 7a0a83f
fix: Final black formatting
