+++ src/paradigma/pipeline.py
@@ -1,0 +1,793 @@
+"""
+High-level pipeline runner for ParaDigMa toolbox.
+
+This module provides a unified interface for running different ParaDigMa pipelines
+(gait, tremor, pulse_rate) on sensor data.
+"""
+
+import logging
+from pathlib import Path
+from typing import Dict, List, Optional, Union
+
+import pandas as pd
+
+from paradigma.classification import ClassifierPackage
+from paradigma.config import GaitConfig, PulseRateConfig, TremorConfig
+from paradigma.constants import DataColumns
+from paradigma.pipelines import gait_pipeline, pulse_rate_pipeline, tremor_pipeline
+from paradigma.util import load_tsdf_dataframe
+
+try:
+    from openmovement.load import CwaData
+except ImportError:
+    CwaData = None
+
+try:
+    from avro.datafile import DataFileReader
+    from avro.io import DatumReader
+except ImportError:
+    DataFileReader = None
+    DatumReader = None
+
+logger = logging.getLogger(__name__)
+
+# Pipeline mapping
+AVAILABLE_PIPELINES = {
+    "gait": {
+        "config_class": GaitConfig,
+        "extract_features": gait_pipeline.extract_gait_features,
+        "detect": gait_pipeline.detect_gait,
+        "quantify": gait_pipeline.quantify_arm_swing,
+        "classifier_path": "gait/classifier_gait.pkl",
+        "required_sensors": ["accelerometer", "gyroscope"],
+    },
+    "tremor": {
+        "config_class": TremorConfig,
+        "extract_features": tremor_pipeline.extract_tremor_features,
+        "detect": tremor_pipeline.detect_tremor,
+        "classifier_path": "tremor/classifier_tremor.pkl",
+        "required_sensors": ["gyroscope"],
+    },
+    "pulse_rate": {
+        "config_class": PulseRateConfig,
+        "estimate": pulse_rate_pipeline.estimate_pulse_rate,
+        "aggregate": pulse_rate_pipeline.aggregate_pulse_rate,
+        "classifier_path": "pulse_rate/classifier_ppg_quality.pkl",
+        "required_sensors": ["ppg", "accelerometer"],
+    },
+}
+
+DEFAULT_CONFIGS = {
+    "gait": lambda: GaitConfig(step="gait"),
+    "tremor": lambda: TremorConfig(step="features"),
+    "pulse_rate": lambda: PulseRateConfig(),
+}
+
+
+def run_pipeline(
+    data_path: Union[str, Path],
+    pipelines: List[str],
+    config: Union[str, Dict, None] = "default",
+    output_dir: Optional[Union[str, Path]] = None,
+    data_format: Optional[str] = None,
+    file_pattern: Optional[str] = None,
+    column_mapping: Optional[Dict[str, str]] = None,
+    parallel: bool = False,
+    verbose: bool = False,
+) -> Dict[str, pd.DataFrame]:
+    """
+    Run one or more ParaDigMa pipelines on sensor data.
+
+    This function provides a high-level interface to run ParaDigMa pipelines on raw
+    sensor data stored in TSDF format.
+
+    Parameters
+    ----------
+    data_path : str or Path
+        Path to directory or file containing sensor data.
+    pipelines : List[str]
+        List of pipeline names to run. Valid options: ['gait', 'tremor', 'pulse_rate']
+    config : str, dict, or None, optional
+        Pipeline configuration. Options:
+        - "default": Use default configurations
+        - dict: Dictionary mapping pipeline names to config objects
+        - None: Use default configurations
+    output_dir : str or Path, optional
+        Directory to save pipeline results. If None, results are only returned.
+    data_format : str, optional
[…]
