+++ src/paradigma/__init__.py
@@ -3,4 +3,7 @@ from importlib.metadata import version
 
 __version__ = version("paradigma")
 
-__all__ = []
+# Import main pipeline runner function
+from paradigma.pipeline import list_available_pipelines, run_pipeline
+
+__all__ = ["run_pipeline", "list_available_pipelines"]
