### 2026-01-12 3e3c75f
Fix syntax errors and code formatting in pipeline implementation
