+++ src/paradigma/pipeline.py
@@ -9,6 +9,7 @@ import logging
 from pathlib import Path
 from typing import Dict, List, Optional, Union
 
+import numpy as np
 import pandas as pd
 
 from paradigma.classification import ClassifierPackage
@@ -67,11 +68,15 @@ DEFAULT_CONFIGS = {
 def run_pipeline(
     data_path: Union[str, Path],
     pipelines: List[str],
+    data_format: str,
     config: Union[str, Dict, None] = "default",
     output_dir: Optional[Union[str, Path]] = None,
-    data_format: Optional[str] = None,
     file_pattern: Optional[str] = None,
     column_mapping: Optional[Dict[str, str]] = None,
+    sampling_frequency: Optional[float] = None,
+    steps: Optional[List[str]] = None,
+    save_intermediate: bool = False,
+    multi_subject: bool = False,
     parallel: bool = False,
     verbose: bool = False,
 ) -> Dict[str, pd.DataFrame]:
@@ -84,9 +89,12 @@ def run_pipeline(
     Parameters
     ----------
     data_path : str or Path
-        Path to directory or file containing sensor data.
+        Path to directory or file containing sensor data. For multi_subject=True,
+        path should contain subdirectories, one per subject.
     pipelines : List[str]
         List of pipeline names to run. Valid options: ['gait', 'tremor', 'pulse_rate']
+    data_format : str
+        Data format. Required parameter. Options: 'tsdf', 'empatica', 'axivity', 'prepared'.
     config : str, dict, or None, optional
         Pipeline configuration. Options:
         - "default": Use default configurations
@@ -94,15 +102,22 @@ def run_pipeline(
         - None: Use default configurations
     output_dir : str or Path, optional
         Directory to save pipeline results. If None, results are only returned.
-    data_format : str, optional
-        Data format. Options: 'tsdf', 'empatica', 'axivity', 'prepared'.
-        If None, format will be auto-detected from file extensions.
     file_pattern : str, optional
         File pattern or extension to filter files (e.g., '*.parquet', '*.pkl').
         Used for prepared dataframes format.
     column_mapping : dict, optional
         Dictionary mapping old column names to new ones. Use this to rename
         columns to match pipeline expectations (e.g., {'acceleration_x': 'accelerometer_x'})
+    sampling_frequency : float, optional
+        Sampling frequency in Hz. Required for raw data formats (tsdf, empatica, axivity)
+        if not specified in data metadata.
+    steps : List[str], optional
+        Pipeline steps to run. Options: ['preprocessing', 'detection', 'quantification', 'aggregation'].
+        If None, runs all steps. Different pipelines have different step sequences.
+    save_intermediate : bool, optional
+        Save intermediate results at each step (default: False)
+    multi_subject : bool, optional
+        Process multiple subjects in separate subdirectories (default: False)
     parallel : bool, optional
         Enable parallel processing where supported (default: False)
     verbose : bool, optional
@@ -154,10 +169,13 @@ def run_pipeline(
             f"Available: {list(AVAILABLE_PIPELINES.keys())}"
         )
 
-    # Auto-detect data format if not specified
-    if data_format is None:
-        data_format = _detect_data_format(data_path, file_pattern)
-        logger.info(f"Auto-detected data format: {data_format}")
+    # Validate data format is provided
+    if data_format not in ["tsdf", "empatica", "axivity", "prepared"]:
+        raise ValueError(
+            f"Invalid data_format: {data_format}. Must be one of: tsdf, empatica, axivity, prepared"
+        )
+
+    logger.info(f"Processing data format: {data_format}")
 
     # Setup configurations
     configs = _setup_configs(pipelines, config)
@@ -166,6 +184,23 @@ def run_pipeline(
[…]
