### 2026-01-12 af972dc
Fix code formatting
