+++ src/paradigma/pipeline.py
@@ -65,10 +65,56 @@ DEFAULT_CONFIGS = {
 }
 
 
+def _detect_sampling_frequency(df: pd.DataFrame, time_column: str = "time") -> float:
+    """Detect sampling frequency from time series data.
+
+    Args:
+        df: DataFrame with time series data
+        time_column: Name of time column
+
+    Returns:
+        Detected sampling frequency in Hz
+
+    Raises:
+        ValueError: If sampling frequency cannot be determined
+    """
+    try:
+        time_data = df[time_column].values
+        if len(time_data) < 2:
+            raise ValueError("Need at least 2 time points to detect sampling frequency")
+
+        # Calculate time differences
+        time_diffs = np.diff(time_data)
+
+        # Remove any outliers that might be due to data gaps
+        median_diff = np.median(time_diffs)
+        filtered_diffs = time_diffs[
+            np.abs(time_diffs - median_diff) < median_diff * 0.5
+        ]
+
+        if len(filtered_diffs) == 0:
+            filtered_diffs = time_diffs
+
+        # Calculate mean interval
+        mean_interval = np.mean(filtered_diffs)
+
+        if mean_interval <= 0:
+            raise ValueError("Invalid time intervals detected")
+
+        # Convert to frequency
+        frequency = 1.0 / mean_interval
+
+        logger.info(f"Auto-detected sampling frequency: {frequency:.2f} Hz")
+        return frequency
+
+    except Exception as e:
+        raise ValueError(f"Could not detect sampling frequency: {e}")
+
+
 def run_pipeline(
     data_path: Union[str, Path],
     pipelines: List[str],
-    data_format: str,
+    data_format: Optional[str] = None,
     config: Union[str, Dict, None] = "default",
     output_dir: Optional[Union[str, Path]] = None,
     file_pattern: Optional[str] = None,
@@ -93,8 +139,9 @@ def run_pipeline(
         path should contain subdirectories, one per subject.
     pipelines : List[str]
         List of pipeline names to run. Valid options: ['gait', 'tremor', 'pulse_rate']
-    data_format : str
-        Data format. Required parameter. Options: 'tsdf', 'empatica', 'axivity', 'prepared'.
+    data_format : str, optional
+        Data format. If not provided, will be auto-detected from file extensions.
+        Options: 'tsdf', 'empatica', 'axivity', 'prepared'.
     config : str, dict, or None, optional
         Pipeline configuration. Options:
         - "default": Use default configurations
@@ -109,8 +156,8 @@ def run_pipeline(
         Dictionary mapping old column names to new ones. Use this to rename
         columns to match pipeline expectations (e.g., {'acceleration_x': 'accelerometer_x'})
     sampling_frequency : float, optional
-        Sampling frequency in Hz. Required for raw data formats (tsdf, empatica, axivity)
-        if not specified in data metadata.
+        Sampling frequency in Hz. If not provided, will be auto-detected from time series data.
+        Only needed if auto-detection fails for raw data formats.
     steps : List[str], optional
         Pipeline steps to run. Options: ['preprocessing', 'detection', 'quantification', 'aggregation'].
         If None, runs all steps. Different pipelines have different step sequences.
@@ -130,7 +177,7 @@ def run_pipeline(
 
     Examples
     --------
-    >>> # Run gait pipeline on TSDF data
+    >>> # Run gait pipeline (data format auto-detected)
     >>> results = run_pipeline(
     ...     data_path="data/tsdf/",
     ...     pipelines=["gait"],
@@ -145,7 +192,7 @@ def run_pipeline(
     ...     file_pattern="*.parquet"
     ... )
 
-    >>> # Run on Axivity CWA files
+    >>> # Run on Axivity CWA files with explicit parameters
     >>> results = run_pipeline(
     ...     data_path="data/axivity/",
[…]
