### 2026-01-12 6573782
Add tutorial urls to readme (#245)
* Reframe index ToC, add dev option for building docs, add new preprints to readme
