+++ docs/tutorials/index.md
@@ -1,0 +1,14 @@
+## Tutorials
+
+This section contains step-by-step tutorials demonstrating how to use
+the ParaDigMa toolbox to extract digital measures from wrist sensor data.
+
+---
+
+**Available tutorials**
+
+- [Device-specific data loading](device_specific_data_loading)
+- [Data preparation](data_preparation)
+- [Gait analysis](gait_analysis)
+- [Tremor analysis](tremor_analysis)
+- [Pulse rate analysis](pulse_rate_analysis)
