### 2026-01-14 biomarkersParkinson/paradigma#250 @Erikpostt
Your approach is correct, so let me take some time to reproduce this. I need to double-check recent changes to LFS in Paradigma, as we haven't seen this issue before.
