### 2026-01-14 355aa92
Update tutorials, shorten readme, add guides, expand orchestrator
