+++ src/paradigma/pipeline.py
@@ -1,1484 +1,0 @@
-"""
-High-level pipeline runner for ParaDigMa toolbox.
-
-This module provides a unified interface for running different ParaDigMa pipelines
-(gait, tremor, pulse_rate) on sensor data.
-"""
-
-import logging
-from pathlib import Path
-from typing import Dict, List, Optional, Union
-
-import numpy as np
-import pandas as pd
-
-from paradigma.classification import ClassifierPackage
-from paradigma.config import GaitConfig, PulseRateConfig, TremorConfig
-from paradigma.constants import DataColumns
-from paradigma.pipelines import gait_pipeline, pulse_rate_pipeline, tremor_pipeline
-from paradigma.util import load_tsdf_dataframe
-
-try:
-    from openmovement.load import CwaData
-except ImportError:
-    CwaData = None
-
-try:
-    from avro.datafile import DataFileReader
-    from avro.io import DatumReader
-except ImportError:
-    DataFileReader = None
-    DatumReader = None
-
-logger = logging.getLogger(__name__)
-
-# Pipeline mapping
-AVAILABLE_PIPELINES = {
-    "gait": {
-        "config_class": GaitConfig,
-        "extract_features": gait_pipeline.extract_gait_features,
-        "detect": gait_pipeline.detect_gait,
-        "quantify": gait_pipeline.quantify_arm_swing,
-        "classifier_path": "gait/classifier_gait.pkl",
-        "required_sensors": ["accelerometer", "gyroscope"],
-    },
-    "tremor": {
-        "config_class": TremorConfig,
-        "extract_features": tremor_pipeline.extract_tremor_features,
-        "detect": tremor_pipeline.detect_tremor,
-        "classifier_path": "tremor/classifier_tremor.pkl",
-        "required_sensors": ["gyroscope"],
-    },
-    "pulse_rate": {
-        "config_class": PulseRateConfig,
-        "estimate": pulse_rate_pipeline.estimate_pulse_rate,
-        "aggregate": pulse_rate_pipeline.aggregate_pulse_rate,
-        "classifier_path": "pulse_rate/classifier_ppg_quality.pkl",
-        "required_sensors": ["ppg", "accelerometer"],
-    },
-}
-
-DEFAULT_CONFIGS = {
-    "gait": lambda: GaitConfig(step="gait"),
-    "tremor": lambda: TremorConfig(step="features"),
-    "pulse_rate": lambda: PulseRateConfig(),
-}
-
-
-def _detect_sampling_frequency(df: pd.DataFrame, time_column: str = "time") -> float:
-    """Detect sampling frequency from time series data.
-
-    Args:
-        df: DataFrame with time series data
-        time_column: Name of time column
-
-    Returns:
-        Detected sampling frequency in Hz
-
-    Raises:
-        ValueError: If sampling frequency cannot be determined
-    """
-    try:
-        time_data = df[time_column].values
-        if len(time_data) < 2:
-            raise ValueError("Need at least 2 time points to detect sampling frequency")
-
-        # Calculate time differences
-        time_diffs = np.diff(time_data)
-
-        # Remove any outliers that might be due to data gaps
-        median_diff = np.median(time_diffs)
-        filtered_diffs = time_diffs[
-            np.abs(time_diffs - median_diff) < median_diff * 0.5
-        ]
-
-        if len(filtered_diffs) == 0:
-            filtered_diffs = time_diffs
-
-        # Calculate mean interval
-        mean_interval = np.mean(filtered_diffs)
-
-        if mean_interval <= 0:
-            raise ValueError("Invalid time intervals detected")
-
-        # Convert to frequency
-        frequency = 1.0 / mean_interval
-
[…]
