+++ src/paradigma/cli.py
@@ -1,214 +1,0 @@
-"""
-Command Line Interface for ParaDigMa toolbox.
-
-This module provides CLI commands for running ParaDigMa pipelines on sensor data.
-"""
-
-import argparse
-import json
-import sys
-from pathlib import Path
-
-from paradigma.pipeline import list_available_pipelines, run_pipeline
-
-
-def main():
-    """Main entry point for the ParaDigMa CLI."""
-    parser = argparse.ArgumentParser(
-        description="ParaDigMa - Parkinson's Disease Digital Markers Toolbox",
-        formatter_class=argparse.RawDescriptionHelpFormatter,
-        epilog="""
-Examples:
-  # Run gait pipeline on TSDF data (data format required)
-  paradigma run data/tsdf/ --pipelines gait --data-format tsdf --sampling-frequency 100 --output results/
-
-  # Run on prepared dataframes (parquet files)
-  paradigma run data/prepared/ --pipelines gait tremor --data-format prepared --file-pattern "*.parquet" --output results/
-
-  # Run on Axivity CWA files with verbose output
-  paradigma run data/axivity/ --pipelines gait --data-format axivity --sampling-frequency 100 --verbose --output results/
-
-  # Use column mapping for different naming conventions
-  paradigma run data/tsdf/ --pipelines gait --data-format tsdf --sampling-frequency 100 --column-mapping '{"acceleration_x": "accelerometer_x", "rotation_x": "gyroscope_x"}'
-
-  # Multi-subject processing
-  paradigma run subjects_data/ --pipelines gait --data-format prepared --multi-subject --output results/
-
-  # Save intermediate results
-  paradigma run data/prepared/ --pipelines gait --data-format prepared --save-intermediate --output results/
-  paradigma run data/mixed/ --pipelines gait tremor --output results/
-
-  # List available pipelines
-  paradigma list-pipelines
-        """,
-    )
-
-    subparsers = parser.add_subparsers(dest="command", help="Available commands")
-
-    # Run command
-    run_parser = subparsers.add_parser(
-        "run", help="Run ParaDigMa pipelines on sensor data"
-    )
-    run_parser.add_argument(
-        "data_path", type=str, help="Path to directory or file containing sensor data"
-    )
-    run_parser.add_argument(
-        "--pipelines",
-        nargs="+",
-        required=True,
-        choices=list_available_pipelines(),
-        help="Pipeline(s) to run",
-    )
-    run_parser.add_argument(
-        "--output", "-o", type=str, help="Output directory for results (optional)"
-    )
-    run_parser.add_argument(
-        "--data-format",
-        type=str,
-        choices=["tsdf", "empatica", "axivity", "prepared"],
-        help="Data format. If not provided, will be auto-detected from file extensions. Options: tsdf, empatica, axivity, prepared",
-    )
-    run_parser.add_argument(
-        "--file-pattern",
-        type=str,
-        help='File pattern for prepared dataframes (e.g., "*.parquet", "*.pkl")',
-    )
-    run_parser.add_argument(
-        "--config", type=str, help="Path to JSON configuration file (optional)"
-    )
-    run_parser.add_argument(
-        "--column-mapping",
-        type=str,
-        help='Column mapping as JSON string or path to JSON file (e.g., \'{"acceleration_x": "accelerometer_x"}\')',
-    )
-    run_parser.add_argument(
-        "--sampling-frequency",
-        type=float,
-        help="Sampling frequency in Hz. If not provided, will be auto-detected from time series data.",
-    )
-    run_parser.add_argument(
-        "--steps",
-        nargs="+",
-        help="Pipeline steps to run (e.g., detection quantification aggregation)",
-    )
-    run_parser.add_argument(
[…]
