+++ src/paradigma/preprocessing.py
@@ -7,72 +7,257 @@ from scipy import signal
 from scipy.interpolate import interp1d
 
 from paradigma.config import IMUConfig, PPGConfig
+from paradigma.segmenting import create_segments, discard_segments
 from paradigma.util import invert_watch_side
 
 
 def resample_data(
     df: pd.DataFrame,
-    time_column: str,
-    values_column_names: List[str],
-    sampling_frequency: int,
-    resampling_frequency: int,
+    time_column: str = "time",
+    values_column_names: List[str] | None = None,
+    sampling_frequency: int | None = None,
+    resampling_frequency: int | None = None,
     tolerance: float | None = None,
+    validate_contiguous: bool = True,
+    auto_segment: bool = False,
+    max_segment_gap_s: float | None = None,
+    min_segment_length_s: float | None = None,
+    verbosity: int = 2,
 ) -> pd.DataFrame:
     """
-    Resamples sensor data to a specified frequency using cubic interpolation.
+    Unified resampling function with optional auto-segmentation for non-contiguous data.
+
+    This function combines the functionality of both prepare_data.resample_data (auto-detection)
+    and preprocessing.resample_data (explicit validation). It supports:
+    - Automatic frequency detection or explicit specification
+    - Contiguity validation with configurable tolerance
+    - Automatic segmentation of non-contiguous data
+    - Preservation of non-numeric columns
 
     Parameters
     ----------
     df : pd.DataFrame
         The input DataFrame containing the sensor data.
-    time_column : str
+    time_column : str, default 'time'
         The name of the column containing the time data.
-    values_column_names : List[str]
-        A list of column names that should be resampled.
-    sampling_frequency : int
-        The original sampling frequency of the data (in Hz).
-    resampling_frequency : int
-        The frequency to which the data should be resampled (in Hz).
+    values_column_names : List[str], optional
+        Column names to resample. If None, auto-detects all numeric columns except time.
+    sampling_frequency : int, optional
+        Original sampling frequency (Hz). If None, auto-detected from data.
+    resampling_frequency : int, optional
+        Target sampling frequency in Hz.
     tolerance : float, optional
-        The tolerance added to the expected difference when checking
-        for contiguous timestamps. If not provided, it defaults to the tolerance specified in IMUConfig.
+        Tolerance for contiguity checking (seconds). Defaults to IMUConfig tolerance.
+    validate_contiguous : bool, default True
+        Whether to validate data contiguity. If False, gaps are silently interpolated.
+    auto_segment : bool, default False
+        If True, automatically split non-contiguous data into segments and process each.
+        Adds 'data_segment_nr' column to output. If False and data is non-contiguous with
+        validate_contiguous=True, raises ValueError.
+    max_segment_gap_s : float, optional
+        Maximum gap (seconds) before starting new segment. Used when auto_segment=True.
+        Defaults to IMUConfig.max_segment_gap_s (1.5s).
+    min_segment_length_s : float, optional
+        Minimum segment length (seconds) to keep. Used when auto_segment=True.
+        Defaults to IMUConfig.min_segment_length_s (1.5s).
+    verbosity : int, default 1
+        Logging verbosity: 0=errors only, 1=basic info, 2+=detailed info.
 
     Returns
     -------
     pd.DataFrame
-        A DataFrame with the resampled data, where each column contains resampled values.
-        The time column will reflect the new resampling frequency.
+        Resampled DataFrame. If auto_segment=True and multiple segments found,
+        includes 'data_segment_nr' column identifying each contiguous data segment.
 
     Raises
     ------
     ValueError
[…]
