+++ src/paradigma/prepare_data.py
@@ -14,11 +14,10 @@ Based on data_preparation tutorial.
 import logging
 from typing import Dict, List
 
-import numpy as np
 import pandas as pd
-from scipy.interpolate import interp1d
 
 from paradigma.constants import DataColumns, TimeUnit
+from paradigma.preprocessing import resample_data
 from paradigma.util import (
     convert_units_accelerometer,
     convert_units_gyroscope,
@@ -252,108 +251,6 @@ def correct_watch_orientation(
     return out
 
 
-def resample_data(
-    df: pd.DataFrame,
-    target_frequency: float = 100.0,
-    time_column: str = DataColumns.TIME,
-    verbosity: int = 1,
-) -> pd.DataFrame:
-    """
-    Resample data to target frequency using cubic interpolation.
-
-    Parameters
-    ----------
-    df : pd.DataFrame
-        Input DataFrame
-    target_frequency : float, default 100.0
-        Target sampling frequency in Hz
-    time_column : str, default DataColumns.TIME
-        Name of time column
-
-    Returns
-    -------
-    pd.DataFrame
-        Resampled DataFrame
-    """
-    df = df.copy()
-
-    if time_column not in df.columns:
-        raise ValueError(f"Time column '{time_column}' not found in DataFrame")
-
-    # Check current sampling frequency
-    time_diff = df[time_column].diff().dropna()
-    current_dt = time_diff.median()
-    current_frequency = 1.0 / current_dt
-
-    if verbosity >= 2:
-        logger.info(f"Current sampling frequency: {current_frequency:.2f} Hz")
-
-    if abs(current_frequency - target_frequency) < 0.1:
-        if verbosity >= 2:
-            logger.info("Data already at target frequency")
-        return df
-
-    # Extract time and values
-    time_abs_array = np.array(df[time_column])
-
-    # Get all numeric columns except time
-    numeric_columns = df.select_dtypes(include=[np.number]).columns
-    values_column_names = [col for col in numeric_columns if col != time_column]
-
-    if not values_column_names:
-        logger.warning("No numeric columns found for resampling")
-        return df
-
-    values_array = np.array(df[values_column_names])
-
-    # Ensure the time array is strictly increasing
-    if not np.all(np.diff(time_abs_array) > 0):
-        raise ValueError("Time array is not strictly increasing")
-
-    # Resample the time data using the specified frequency
-    t_resampled = np.arange(time_abs_array[0], time_abs_array[-1], 1 / target_frequency)
-
-    # Choose interpolation method
-    interpolation_kind = "cubic" if len(time_abs_array) > 3 else "linear"
-    interpolator = interp1d(
-        time_abs_array,
-        values_array,
-        axis=0,
-        kind=interpolation_kind,
-        fill_value="extrapolate",
-    )
-
-    # Interpolate
-    resampled_values = interpolator(t_resampled)
-
-    # Create a DataFrame with the resampled data
-    df_resampled = pd.DataFrame(resampled_values, columns=values_column_names)
-    df_resampled[time_column] = t_resampled
-
-    # Copy non-numeric columns (if any) - take first value
-    for column in df.columns:
-        if column not in df_resampled.columns:
-            df_resampled[column] = df[column].iloc[0]
-
-    # Return the DataFrame with columns in the correct order
-    resampled_columns = (
-        [time_column]
-        + values_column_names
-        + [
-            col
-            for col in df_resampled.columns
-            if col not in [time_column] + values_column_names
-        ]
-    )
-    df_resampled = df_resampled[resampled_columns]
-
-    if verbosity >= 1:
-        logger.info(
[…]
