### 2026-01-15 2ce56a0
Resolve raised issues by copilot in PR
