### 2026-01-15 87f352e
Re-build docs with verbose added
