+++ tests/test_prepare_data.py
@@ -40,7 +40,7 @@ class TestPrepareRawData:
             time_input_unit=TimeUnit.RELATIVE_S,
             resampling_frequency=100.0,
             validate=True,
-            verbosity=0,
+            verbose=0,
         )
 
         # Verify output
@@ -89,7 +89,7 @@ class TestPrepareRawData:
             max_segment_gap_s=2.0,  # Gaps > 2s trigger segmentation
             min_segment_length_s=1.0,
             validate=False,  # Skip validation since we have segments
-            verbosity=0,
+            verbose=0,
         )
 
         # Verify segmentation occurred
