### 2026-01-15 #9502 @Asanto32
Hi @Erikpostt,

Just going over some of these now:

> **5. Linting could be quite useful for improving readability of the codebase.** We appreciate the suggestion. Our pre-commit hooks already include Black, isort, and Ruff, which enforce consistent formatting and linting across the codebase. We believe major readability guidelines are covered automatically. Are there specific formatting issues you noticed that are not currently handled?
>

I recall some sections of code where the number of characters in a line was very long, for example: https://github.com/biomarkersParkinson/paradigma/blob/6573782c4f0b1e8444e5dbf57ee9046b5c91b309/src/paradigma/feature_extraction.py#L979
Usually this would be caught in Ruff, but perhaps you have removed the "line too long" format check?

> **7. I feel like the claims about using this toolbox for raw sensor data/device agnostic processing are not accurate. I tried feeding in data I have from GENEActiv and Actigraph watches, however this would require me to extract data, separate the data from the timestamps, and create specific metadata json files. This requires raw data to be heavily modified into, you can’t simply provide a path to a “.bin” or “.gt3x” file, for example. I understand that in theory I just need a pandas dataframe to run the analysis, but I feel like some of the claims about only needing raw sensor data should be revisited.** Thank you for testing Paradigma with your own data. The requirement is indeed a pandas DataFrame; the tutorial examples using `tsdf` files are just one format. Device-agnostic claims refer to the fact that results are consistent across different devices, not that every raw file type is automatically supported. Do you think that clarifying these claims is sufficient? With "... but I feel like some of the claims about only needing raw sensor data should be revisited", do you mean that in fact _Paradigma_ requires a pandas DataFrame in memory?

Yes, I think the text of the paper and the documentation (ReadMe, tutorials) should cover that you cannot simply supply raw sensor data (ex: `.gt3x`, `bin`, fileypes that are commonly used). But rather it must first be converted/processed into a pandas DataFrame/readable format.

> **8. My biggest issue is that there is no Python or CLI entrypoint. If this was truly a modular pipeline, we should be able to easily point to raw sensor data, and call any specific or combination of analysis pipelines. As it stands, unless I am missing something, a user would need to follow what seems to be quite a long series of notebooks in the tutorial. Perhaps I am not the target audience, but I feel this would make the toolbox much more usable.** Great point! We agree this would improve usability. Currently, the target audience is researchers with minor Python experience, following tutorials to run analyses. Developing Python/CLI entrypoints is on our roadmap, and of high priority. Would you consider this a required feature for publication, or acceptable to implement shortly afterward?
>

I look forward to testing the package again once this has been implemented!
