### 2026-01-23 eef4bc3
Modify codebase to reviewer suggestions
