+++ tests/test_gait_analysis.py
@@ -50,6 +50,42 @@ def create_test_gait_data(duration_minutes=2):
     )
 
 
+def create_test_gait_data_with_gaps(n_segments=3):
+    """Create test data with gaps (multiple data segments)."""
+    segments = []
+    for i in range(n_segments):
+        # Create 30 second segments with large gaps between them
+        duration_seconds = 30
+        n_samples = int(duration_seconds * 100.0)  # 100 Hz
+        time_start = i * 100  # 100 second gaps between segments
+        time = np.linspace(time_start, time_start + duration_seconds, n_samples)
+
+        # Simple IMU data
+        acc_x = 0.2 * np.sin(2 * np.pi * 1.5 * time) + 0.1 * np.random.random(n_samples)
+        acc_y = 0.2 * np.cos(2 * np.pi * 1.5 * time) + 0.1 * np.random.random(n_samples)
+        acc_z = 9.8 + 0.2 * np.random.random(n_samples)
+
+        gyro_x = 0.1 * np.sin(2 * np.pi * 2 * time) + 0.05 * np.random.random(n_samples)
+        gyro_y = 0.1 * np.cos(2 * np.pi * 2 * time) + 0.05 * np.random.random(n_samples)
+        gyro_z = 0.05 * np.random.random(n_samples)
+
+        segment_df = pd.DataFrame(
+            {
+                DataColumns.TIME: time,
+                DataColumns.ACCELEROMETER_X: acc_x,
+                DataColumns.ACCELEROMETER_Y: acc_y,
+                DataColumns.ACCELEROMETER_Z: acc_z,
+                DataColumns.GYROSCOPE_X: gyro_x,
+                DataColumns.GYROSCOPE_Y: gyro_y,
+                DataColumns.GYROSCOPE_Z: gyro_z,
+                DataColumns.DATA_SEGMENT_NR: i + 1,  # Add data segment number
+            }
+        )
+        segments.append(segment_df)
+
+    return pd.concat(segments, ignore_index=True)
+
+
 def test_gait_pipeline_basic():
     """Test basic gait pipeline functionality."""
     df_test = create_test_gait_data()
@@ -86,6 +122,125 @@ def test_gait_pipeline_integration():
         assert "quantifications" in results
         assert "metadata" in results
         assert "aggregations" in results
+
+
+def test_gait_segment_nr_column_name():
+    """
+    Test that gait_segment_nr column is present and backward compatibility is maintained.
+
+    This test verifies:
+    1. The new GAIT_SEGMENT_NR constant exists
+    2. The old SEGMENT_NR constant still works (backward compatibility)
+    3. Both constants reference the same column name
+    4. Output DataFrame contains the gait_segment_nr column when using constants
+    """
+    # Test 1: Verify both constants exist and point to same value
+    assert hasattr(
+        DataColumns, "GAIT_SEGMENT_NR"
+    ), "GAIT_SEGMENT_NR constant should exist"
+    assert hasattr(
+        DataColumns, "SEGMENT_NR"
+    ), "SEGMENT_NR constant should exist for backward compatibility"
+    assert (
+        DataColumns.GAIT_SEGMENT_NR == "gait_segment_nr"
+    ), "GAIT_SEGMENT_NR should be 'gait_segment_nr'"
+    assert (
+        DataColumns.SEGMENT_NR == DataColumns.GAIT_SEGMENT_NR
+    ), "SEGMENT_NR should alias to GAIT_SEGMENT_NR for backward compatibility"
+
+    # Test 2: Create a mock DataFrame to verify column access works with both constants
+    mock_gait_results = pd.DataFrame(
+        {
+            "gait_segment_nr": [1, 1, 2, 2, 3],
+            "range_of_motion": [0.5, 0.6, 0.7, 0.5, 0.8],
[…]
