### 2026-01-26 5e4083d
Update readme and docs based on internal feedback
