+++ docs/guides/supported_devices.md
@@ -1,132 +1,247 @@
 # Supported Devices
 
-This guide documents the devices that have been tested with ParaDigMa and provides device-specific loading and usage guidance.
+This guide documents devices that have been tested with ParaDigMa and provides device-specific guidance for data loading and usage.
 
-## Validated Devices
+ParaDigMa is designed to work with wrist sensor data from any device, as long as the data meets the [Sensor Requirements](sensor_requirements.md). However, validation has only been performed for specific devices listed below.
 
-### Gait-up Physilog 4
+---
 
-**Validation Status**: Arm Swing & Tremor ✓
+## Scientifically Validated Devices
 
-A research-grade inertial measurement unit (IMU) worn as a wrist-mounted sensor band.
-
-- **Sensors**: 3-axis accelerometer, 3-axis gyroscope, 3-axis magnetometer
-- **Sampling Rate**: 200 Hz
-
-#### Validation Details
-
-- **Arm Swing**: Validated on Parkinson@Home dataset
-- **Tremor**: Validated on Parkinson@Home dataset
-- **Known Limitations**: No PPG sensor on this device
+These devices have undergone rigorous scientific validation with ParaDigMa pipelines.
 
 ### Verily Study Watch
 
-**Validation Status**: Arm Swing, Tremor & Pulse Rate ✓
-
 A research-grade smartwatch developed by Verily Life Sciences for clinical research.
 
+#### Specifications
+
 - **Sensors**: 3-axis accelerometer, 3-axis gyroscope, photoplethysmography (PPG)
-- **Sampling Rate**: 100 Hz (IMU), 30 Hz (PPG)
+- **Sampling Rates**: 100 Hz (IMU), 30 Hz (PPG)
+- **Validated Pipelines**: Gait, Tremor, Pulse Rate
 
-#### Validation Details
+---
 
-- **Arm Swing**: Validated on PPP dataset
-- **Tremor**: Validated on PPP dataset
-- **Pulse Rate**: Validated on PPP dataset
+### Gait-up Physilog 4
 
-## Commonly Used Devices with Community Support
+A research-grade IMU worn as a wrist-mounted sensor band.
 
-### Empatica E4
+#### Specifications
 
-**Validation Status**: Community support (not formally validated by ParaDigMa team)
+- **Sensors**: 3-axis accelerometer, 3-axis gyroscope, 3-axis magnetometer
+- **Sampling Rate**: 200 Hz
+- **Validated Pipelines**: Gait, Tremor
 
-A wrist-worn research device with PPG and accelerometer sensors.
+---
 
-- **Sensors**: 3-axis accelerometer, PPG
-- **Data Format**: AVRO format (native)
-- **Supported by ParaDigMa**: Partial (data loading supported; not validated)
+## Empirically Validated Devices
 
-#### Loading Empatica Data
+These devices have been tested with ParaDigMa and show promising results, but have not undergone full scientific validation.
 
-See the [Device-Specific Data Loading Tutorial](https://biomarkersparkinson.github.io/paradigma/tutorials/device_specific_data_loading.html) for Empatica E4 data loading examples.
-
-#### Known Considerations
-
-- Empatica E4 lacks gyroscope data; tremor and gait analysis require external gyroscope
-- Community examples available but not validated by ParaDigMa team
-
-### Axivity AX3/6
-
-**Validation Status**: Community support (not formally validated by ParaDigMa team)
+### Axivity AX6
 
 A compact wrist-worn IMU designed for long-term monitoring.
 
-- **Sensors**: 3-axis accelerometer (AX3), or 3-axis accelerometer and gyroscope (AX6)
-- **Data Format**: CWA format (native)
[…]
