+++ README.md
@@ -45,6 +45,7 @@ from paradigma.orchestrator import run_paradigma
 results = run_paradigma(
     dfs=df,
     pipelines=['gait', 'tremor'],
+    watch_side='left',  # Required for gait pipeline
     save_intermediate=['quantification', 'aggregation']  # Saves to ./output by default
 )
 
@@ -53,6 +54,7 @@ results = run_paradigma(
     dfs=[df1, df2, df3],
     pipelines=['gait', 'tremor'],
     output_dir="./results",  # Custom output directory
+    watch_side='left',
     save_intermediate=['quantification', 'aggregation']
 )
 
@@ -60,12 +62,24 @@ results = run_paradigma(
 results = run_paradigma(
     dfs={'morning_session': df1, 'evening_session': df2},
     pipelines=['gait', 'tremor'],
+    watch_side='right',
     save_intermediate=[]  # No files saved - results only in memory
 )
 
-# Access results
-quantified_measures = results['quantifications']
-aggregated_measures = results['aggregations']
+# Example 4: Load from data directory
+results = run_paradigma(
+    data_path='./my_data',
+    pipelines=['gait', 'tremor'],
+    watch_side='left',
+    file_pattern='*.parquet',
+    save_intermediate=['quantification', 'aggregation']
+)
+
+# Access results (nested by pipeline)
+gait_measures = results['quantifications']['gait']
+tremor_measures = results['quantifications']['tremor']
+gait_aggregates = results['aggregations']['gait']
+tremor_aggregates = results['aggregations']['tremor']
 ```
 
 **See our [tutorials](https://biomarkersparkinson.github.io/paradigma/tutorials/index.html) for complete examples.**
@@ -76,52 +90,70 @@ aggregated_measures = results['aggregations']
   <img src="https://raw.githubusercontent.com/biomarkersParkinson/paradigma/main/docs/source/_static/img/pipeline-architecture.png" alt="Pipeline architeecture"/>
 </p>
 
+### Validated Processing Pipelines
+
+ParaDigMa contains the following validated processing pipelines:
 | Pipeline | Input sensors | Output week-level aggregation | Tutorial |
 | ---- | ---- | ------- | ---- |
 | **Arm swing during gait** | Accelerometer + Gyroscope | Typical, maximum & variability of arm swing range of motion | [Guide](https://biomarkersparkinson.github.io/paradigma/tutorials/gait_analysis) |
 | **Tremor** | Gyroscope | % tremor time, typical & maximum tremor power | [Guide](https://biomarkersparkinson.github.io/paradigma/tutorials/tremor_analysis) |
 | **Pulse rate** | PPG (+ Accelerometer) | Resting & maximum pulse rate | [Guide](https://biomarkersparkinson.github.io/paradigma/tutorials/pulse_rate_analysis) |
 
+### Pipeline Architecture
+
+ParaDigMa can best be understood by categorizing the sequential processes:
 | Process | Description |
 | ---- | ---- |
-| Preprocessing | Preparing raw sensor signals for further processing |
-| Feature extraction | Extracting features based on windowed sensor signals |
-| Classification | Detecting segments of interest using validated classifiers (e.g., gait segments) |
-| Quantification | Extracting specific measures from the detected segments (e.g., arm swing measures) |
-| Aggregation | Aggregating the measures over a specific time period (e.g., week-level aggregates)
+| **Preprocessing** | Preparing raw sensor signals for further processing |
+| **Feature extraction** | Extracting features based on windowed sensor signals |
+| **Classification** | Detecting segments of interest using validated classifiers (e.g., gait segments) |
[…]

[…]
