### 2026-01-28 98c3fce
Always create output dir for logging
