### 2026-01-28 d58f3a1
Ensure single file to preserve memory efficiency
