+++ src/paradigma/orchestrator.py
@@ -37,7 +37,11 @@ from paradigma.config import (
     TremorConfig,
 )
 from paradigma.constants import DataColumns, TimeUnit
-from paradigma.load import load_data_files, save_prepared_data
+from paradigma.load import (
+    get_data_file_paths,
+    load_single_data_file,
+    save_prepared_data,
+)
 from paradigma.pipelines.gait_pipeline import (
     aggregate_arm_swing_params,
     run_gait_pipeline,
@@ -521,72 +525,103 @@ def run_paradigma(
     logger.addHandler(file_handler)
     logger.info(f"Logging to {log_file}")
 
-    # Step 1: Load data files
+    # Step 1: Get file paths or convert provided DataFrames
+    file_paths = None  # Will hold list of file paths if loading from directory
+    dfs_dict = None  # Will hold dict of DataFrames if provided directly
+
     if data_path is not None:
         if verbose >= 1:
-            logger.info("Step 1: Loading data files")
+            logger.info("Step 1: Finding data files")
         try:
-            dfs = load_data_files(
+            file_paths = get_data_file_paths(
                 data_path=data_path, file_patterns=file_pattern, verbose=verbose
             )
             if verbose >= 1:
-                logger.info(f"Loaded {len(dfs)} data files")
+                logger.info(f"Found {len(file_paths)} data files")
         except Exception as e:
-            logger.error(f"Failed to load data files: {e}")
+            logger.error(f"Failed to find data files: {e}")
             raise
 
-        if not dfs:
+        if not file_paths:
             raise ValueError(f"No data files found in {data_path}")
     else:
         if verbose >= 1:
             logger.info("Step 1: Using provided DataFrame(s) as input")
 
-    if isinstance(dfs, list):
-        # Convert list to dict with unique dataframe IDs
-        dfs = {f"df_{i}": df for i, df in enumerate(dfs, start=1)}
-    elif isinstance(dfs, pd.DataFrame):
-        # Convert single DataFrame to dict with default name
-        dfs = {"df_1": dfs}
+        # Convert provided dfs to dict format
+        if isinstance(dfs, list):
+            dfs_dict = {f"df_{i}": df for i, df in enumerate(dfs, start=1)}
+        elif isinstance(dfs, pd.DataFrame):
+            dfs_dict = {"df_1": dfs}
+        else:
+            dfs_dict = dfs
 
-    # Step 2: Prepare data if needed
-    if not skip_preparation:
-        if verbose >= 1:
-            logger.info("Step 2: Preparing raw data")
+    # Determine number of files to process
+    num_files = len(file_paths) if file_paths else len(dfs_dict)
 
-        prepare_params = {
-            "time_input_unit": time_input_unit,
-            "resampling_frequency": target_frequency,
-            "column_mapping": column_mapping,
-            "auto_segment": split_by_gaps,
-            "max_segment_gap_s": max_gap_seconds,
-            "min_segment_length_s": min_segment_seconds,
-            "verbose": verbose,
-        }
+    # Initialize results storage for each pipeline
+    all_results = {
+        "quantifications": {p: [] for p in pipelines},
+        "aggregations": {p: {} for p in pipelines},
+        "metadata": {p: {} for p in pipelines},
+    }
 
-        # Add pipeline-specific preparation parameters
-        if "gait" in pipelines or "tremor" in pipelines:
-            prepare_params["gyroscope_units"] = gyroscope_units
+    # Step 2 & 3: Process each file individually
+    if verbose >= 1:
+        logger.info(f"Steps 2-3: Processing {num_files} files individually")
 
-        if "gait" in pipelines:
-            prepare_params.update(
-                {
-                    "accelerometer_units": accelerometer_units,
[…]
