+++ src/paradigma/orchestrator.py
@@ -3,13 +3,11 @@ High-level pipeline orchestrator for ParaDigMa toolbox.
 
 This module provides the main entry point for running analysis pipelines:
 
-Main Functions
---------------
+Main Function
+-------------
 - run_paradigma(): Complete pipeline from data loading/preparation to aggregated results.
   Main entry point for end-to-end analysis supporting multiple pipelines (gait, tremor, pulse_rate).
-
-- run_pipeline_batch(): Runs a specific pipeline on multiple data files and aggregates results.
-  Used internally by run_paradigma() for processing batches of prepared data.
+  Can process raw data from disk or already-prepared DataFrames.
 
 The orchestrator coordinates:
 1. Data loading and preparation (unit conversion, resampling, orientation correction)
@@ -56,295 +54,6 @@ from paradigma.prepare_data import prepare_raw_data
 logger = logging.getLogger(__name__)
 
 
-def run_pipeline_batch(
-    pipeline_name: str,
-    dfs: pd.DataFrame | List[pd.DataFrame] | Dict[str, pd.DataFrame],
-    output_dir: str | Path = "./output",
-    watch_side: str | None = None,
-    store_intermediate: List[str] | None = [],
-    aggregates: List[str] | None = None,
-    imu_config: IMUConfig | None = None,
-    ppg_config: PPGConfig | None = None,
-    gait_config: GaitConfig | None = None,
-    arm_activity_config: GaitConfig | None = None,
-    tremor_config: TremorConfig | None = None,
-    pulse_rate_config: PulseRateConfig | None = None,
-    gait_segment_categories: List[str] | None = None,
-    verbose: int = 1,
-) -> Dict[str, pd.DataFrame | Dict]:
-    """
-    Run specific pipeline on multiple data files and aggregate results.
-
-    Parameters
-    ----------
-    dfs : DataFrame, list of DataFrames, or dict of DataFrames
-        Data files to process. Can be:
-        - Single DataFrame: Will be processed as one file with key 'df_1'.
-        - List[DataFrame]: Multiple dataframes assigned IDs as 'df_1', 'df_2', etc.
-        - Dict[str, DataFrame]: Keys are file names, values are dataframes.
-        Note: The 'file_key' column is only added to quantification results when
-        len(dfs) > 1, allowing cleaner output for single-file processing.
-        See input_formats guide for details.
-    pipeline_name: str
-        Name of the pipeline to run: 'gait', 'tremor', or 'pulse_rate'.
-    watch_side : str, optional
-        Watch side: 'left' or 'right' (required for gait pipeline).
-    store_intermediate : list of str, default []
-        Steps of which output is to be stored. Valid values:
-        ['preparation', 'preprocessing', 'classification', 'quantification', 'aggregation']
-        If empty, no files are saved (results are only returned).
-    output_dir : str or Path, optional
-        Output directory for results (required only if store_intermediate is not empty).
-    aggregates : list of str, optional
-        Aggregation methods for quantification.
-    imu_config : IMUConfig, optional
-        IMU preprocessing configuration.
-    ppg_config : PPGConfig, optional
-        PPG preprocessing configuration.
-    gait_config : GaitConfig, optional
-        Gait analysis configuration.
-    arm_activity_config : GaitConfig, optional
-        Arm activity analysis configuration.
-    tremor_config : TremorConfig, optional
-        Tremor analysis configuration.
-    pulse_rate_config : PulseRateConfig, optional
-        Pulse rate analysis configuration.
-    gait_segment_categories : list of str, optional
-        Categories for gait segment aggregation.
-    verbose : int, default 1
-        Logging verbose level.
-
-    Returns
-    -------
-    dict
-        Results dictionary containing:
[…]
