### 2026-01-28 f994df4
Ensure single file to preserve memory efficiency
