### 2026-01-29 7652587
Remove validation tutorial, integrate in readme
