+++ README.md
@@ -11,14 +11,14 @@
 
 ## Overview
 
-ParaDigMa (Parkinson's disease Digital Markers) is a Python toolbox for extracting validated digital biomarkers from wrist-worn sensor data in Parkinson's disease. It processes accelerometer, gyroscope, and PPG signals collected during passive monitoring in daily life.
+ParaDigMa (Parkinson's disease Digital Markers) is a Python toolbox for extracting validated digital biomarkers from wrist sensor data in Parkinson's disease. It processes accelerometer, gyroscope, and PPG signals collected during passive monitoring in daily life.
 
 **Key Features:**
 - Arm swing during gait analysis
 - Tremor analysis
-- Pulse Rate analysis
+- Pulse rate analysis
 - Scientifically validated in peer-reviewed publications
-- Extensible architecture for custom analyses
+- Modular, extensible architecture for custom analyses
 
 ## Quick Start
 
@@ -92,11 +92,13 @@ tremor_aggregates = results['aggregations']['tremor']
 
 ### Validated Processing Pipelines
 
-| Pipeline | Input sensors | Output week-level aggregation | Tutorial |
-| ---- | ---- | ------- | ---- |
-| **Arm swing during gait** | Accelerometer + Gyroscope | Typical, maximum & variability of arm swing range of motion | [Guide](https://biomarkersparkinson.github.io/paradigma/tutorials/gait_analysis) |
-| **Tremor** | Gyroscope | % tremor time, typical & maximum tremor power | [Guide](https://biomarkersparkinson.github.io/paradigma/tutorials/tremor_analysis) |
-| **Pulse rate** | PPG (+ Accelerometer) | Resting & maximum pulse rate | [Guide](https://biomarkersparkinson.github.io/paradigma/tutorials/pulse_rate_analysis) |
+| Pipeline | Input sensors | Output week-level aggregation | Publications | Tutorial |
+| ---- | ---- | ------- | ---- | ---- |
+| **Arm swing during gait** | Accelerometer + Gyroscope | Typical, maximum & variability of arm swing range of motion | [Post 2025](https://doi.org/10.1186/s12984-025-01578-z), [Post 2026*](https://doi.org/10.64898/2026.01.06.26343500) | [Guide](https://biomarkersparkinson.github.io/paradigma/tutorials/gait_analysis) |
+| **Tremor** | Gyroscope | % tremor time, typical & maximum tremor power | [Timmermans 2025a](https://doi.org/10.1038/s41531-025-01056-2), [Timmermans 2025b*](https://www.medrxiv.org/content/10.64898/2025.12.23.25342892v1) | [Guide](https://biomarkersparkinson.github.io/paradigma/tutorials/tremor_analysis) |
+| **Pulse rate** | PPG (+ Accelerometer) | Resting & maximum pulse rate | [Veldkamp 2025*](https://doi.org/10.1101/2025.08.15.25333751) | [Guide](https://biomarkersparkinson.github.io/paradigma/tutorials/pulse_rate_analysis) |
+
+*\* Indicates pre-print*
 
 ### Pipeline Architecture
 
@@ -121,38 +123,29 @@ ParaDigMa can best be understood by categorizing the sequential processes:
 - **[Scientific Validation](https://biomarkersparkinson.github.io/paradigma/guides/validation.html)** - Validation studies and publications
 - **[API Reference](https://biomarkersparkinson.github.io/paradigma/autoapi/paradigma/index.html)** - Complete API documentation
 
-### Data Formats
-
-ParaDigMa supports the following data formats:
-
-- In-memory (recommend): **Pandas DataFrames** (see examples above)
-- Data loading file extensions: **TSDF, Parquet, CSV, Pickle** and **several Device-specific formats** (AVRO (Empatica), CWA (Axivity))
-
-### Troubleshooting
-
[…]
