### 2026-01-29 ca9b8b7
Rename all uppercase variables to comply with PEP 8 (N806/N803)
