+++ src/paradigma/pipelines/pulse_rate_utils.py
@@ -1,5 +1,3 @@
-from typing import Tuple
-
 import numpy as np
 from scipy import signal
 
@@ -42,9 +40,8 @@ def assign_sqa_label(
 
     for i in range(n_samples):
         # Start and end indices for current epoch
-        start_idx = max(
-            0, int((i - (samples_per_epoch - samples_shift)) // fs)
-        )  # max to handle first epochs
+        # Max to handle first epochs
+        start_idx = max(0, int((i - (samples_per_epoch - samples_shift)) // fs))
         end_idx = min(int(i // fs), len(ppg_prob))  # min to handle last epochs
 
         # Extract probabilities and labels for the current epoch
@@ -64,7 +61,7 @@ def assign_sqa_label(
 
 def extract_pr_segments(
     sqa_label: np.ndarray, min_pr_samples: int
-) -> Tuple[np.ndarray, np.ndarray]:
+) -> tuple[np.ndarray, np.ndarray]:
     """
     Extracts pulse rate segments based on the SQA label.
 
@@ -106,16 +103,19 @@ def extract_pr_from_segment(
     Parameters
     ----------
     ppg : np.ndarray
-        The preprocessed PPG segment with 2 seconds of padding on both sides to reduce boundary effects.
+        The preprocessed PPG segment with 2 seconds of padding on both sides
+        to reduce boundary effects.
     tfd_length : int
-        Length of each segment (in seconds) to calculate the time-frequency distribution.
+        Length of each segment (in seconds) to calculate the time-frequency
+        distribution.
     fs : int
         The sampling frequency of the PPG signal.
     kern_type : str
-        Type of TFD kernel to use (e.g., 'wvd' for Wigner-Ville distribution).
+        Type of TFD kernel to use (e.g., 'wvd' for Wigner-Ville
+        distribution).
     kern_params : dict
-        Parameters for the specified kernel. Not required for 'wvd', but relevant for other
-        kernels like 'spwvd' or 'swvd'. Default is None.
+        Parameters for the specified kernel. Not required for 'wvd', but
+        relevant for other kernels like 'spwvd' or 'swvd'. Default is None.
 
     Returns
     -------
@@ -157,8 +157,8 @@ def extract_pr_with_tfd(
     ppg: np.ndarray, fs: int, kern_type: str, kern_params: dict
 ) -> np.ndarray:
     """
-    Estimate pulse rate (PR) from a PPG segment using a TFD method with optional
-    moving average filtering.
+    Estimate pulse rate (PR) from a PPG segment using a TFD method with
+    optional moving average filtering.
 
     Parameters
     ----------
@@ -174,7 +174,8 @@ def extract_pr_with_tfd(
     Returns
     -------
     pr_smooth_tfd : np.ndarray
-        Estimated pr values (in beats per minute) for each 2-second segment of the PPG signal.
+        Estimated pr values (in beats per minute) for each 2-second segment
+        of the PPG signal.
     """
     # Generate the TFD matrix using the specified kernel
     tfd_obj = TimeFreqDistr()
@@ -204,22 +205,34 @@ def extract_pr_with_tfd(
 class TimeFreqDistr:
     def __init__(self):
         """
-        This module contains the implementation of the Generalized Time-Frequency Distribution (TFD) computation using non-separable kernels.
-        This is a Python implementation of the MATLAB code provided by John O Toole in the following repository: https://github.com/otoolej/memeff_TFDs
+        This module contains the implementation of the Generalized
+        Time-Frequency Distribution (TFD) computation using non-separable
+        kernels. This is a Python implementation of the MATLAB code provided
+        by John O Toole in the following repository:
+        https://github.com/otoolej/memeff_TFDs
 
-        The following functions are implemented for the computation of the TFD:
[…]
