+++ src/paradigma/orchestrator.py
@@ -5,8 +5,9 @@ This module provides the main entry point for running analysis pipelines:
 
 Main Function
 -------------
-- run_paradigma(): Complete pipeline from data loading/preparation to aggregated results.
-  Main entry point for end-to-end analysis supporting multiple pipelines (gait, tremor, pulse_rate).
+- run_paradigma(): Complete pipeline from data loading/preparation
+  to aggregated results. Main entry point for end-to-end analysis
+  supporting multiple pipelines (gait, tremor, pulse_rate).
   Can process raw data from disk or already-prepared DataFrames.
 
 The orchestrator coordinates:
@@ -22,7 +23,6 @@ import json
 import logging
 from datetime import datetime
 from pathlib import Path
-from typing import Dict, List
 
 import numpy as np
 import pandas as pd
@@ -57,20 +57,20 @@ logger = logging.getLogger(__name__)
 def run_paradigma(
     output_dir: str | Path = "./output",
     data_path: str | Path | None = None,
-    dfs: pd.DataFrame | List[pd.DataFrame] | Dict[str, pd.DataFrame] | None = None,
+    dfs: pd.DataFrame | list[pd.DataFrame] | dict[str, pd.DataFrame] | None = None,
     skip_preparation: bool = False,
-    pipelines: List[str] | str | None = None,
+    pipelines: list[str] | str | None = None,
     watch_side: str | None = None,
     accelerometer_units: str = "g",
     gyroscope_units: str = "deg/s",
     time_input_unit: TimeUnit = TimeUnit.RELATIVE_S,
     target_frequency: float = 100.0,
-    column_mapping: Dict[str, str] | None = None,
-    device_orientation: List[str] | None = ["x", "y", "z"],
-    save_intermediate: List[str] = [],
-    file_pattern: str | List[str] | None = None,
-    aggregates: List[str] | None = None,
-    segment_length_bins: List[str] | None = None,
+    column_mapping: dict[str, str] | None = None,
+    device_orientation: list[str] | None = ["x", "y", "z"],
+    save_intermediate: list[str] = [],
+    file_pattern: str | list[str] | None = None,
+    aggregates: list[str] | None = None,
+    segment_length_bins: list[str] | None = None,
     split_by_gaps: bool = False,
     max_gap_seconds: float | None = None,
     min_segment_seconds: float | None = None,
@@ -81,11 +81,12 @@ def run_paradigma(
     tremor_config: TremorConfig | None = None,
     pulse_rate_config: PulseRateConfig | None = None,
     verbose: int = 1,
-) -> Dict[str, pd.DataFrame | Dict]:
+) -> dict[str, pd.DataFrame | dict]:
     """
     Complete ParaDigMa analysis pipeline from data loading to aggregated results.
 
-    This is the main entry point for ParaDigMa analysis. It supports multiple pipeline types:
+    This is the main entry point for ParaDigMa analysis. It supports
+    multiple pipeline types:
     - gait: Arm swing during gait analysis
     - tremor: Tremor detection and quantification
     - pulse_rate: Pulse rate estimation from PPG signals
@@ -109,11 +110,13 @@ def run_paradigma(
         len(dfs) > 1, allowing cleaner output for single-file processing.
         See input_formats guide for details.
     skip_preparation : bool, default False
-        Whether data is already prepared. If False, data will be prepared (unit conversion,
-        resampling, etc.). If True, assumes data is already in the required format.
+        Whether data is already prepared. If False, data will be
+        prepared (unit conversion, resampling, etc.). If True,
+        assumes data is already in the required format.
     pipelines : list of str or str, optional
-        Pipelines to run: 'gait', 'tremor', and/or 'pulse_rate'. If providing a list, currently
[…]

[…]
