### 2026-02-03 76e752e
Modify logging, solve NA bug in interpolation, expand building docs
