+++ src/paradigma/load.py
@@ -22,7 +22,8 @@ logger = logging.getLogger(__name__)
 
 
 def load_tsdf_data(
-    data_path: str | Path, prefix: str = "IMU", verbose: int = 1
+    data_path: str | Path,
+    prefix: str = "IMU",
 ) -> tuple[pd.DataFrame, dict, dict]:
     """
     Load TSDF data from .meta and .bin files.
@@ -33,8 +34,6 @@ def load_tsdf_data(
         Path to directory containing TSDF files.
     prefix : str, default "IMU"
         Prefix for TSDF files (e.g., "IMU_segment0001").
-    verbose : int, default 1
-        Logging verbose level.
 
     Returns
     -------
@@ -43,19 +42,19 @@ def load_tsdf_data(
         dict, values metadata dict)
     """
     data_path = Path(data_path)
-    if verbose >= 1:
-        logger.info(f"Loading TSDF data from {data_path} with prefix '{prefix}'")
+    logger.info(f"Loading TSDF data from {data_path} with prefix '{prefix}'")
 
     df, time_meta, values_meta = load_tsdf_dataframe(
         path_to_data=data_path, prefix=prefix
     )
 
-    if verbose >= 1:
-        logger.info(f"Loaded TSDF data: {df.shape[0]} rows, {df.shape[1]} columns")
+    logger.info(f"Loaded TSDF data: {df.shape[0]} rows, {df.shape[1]} columns")
     return df, time_meta, values_meta
 
 
-def load_empatica_data(file_path: str | Path, verbose: int = 1) -> pd.DataFrame:
+def load_empatica_data(
+    file_path: str | Path,
+) -> pd.DataFrame:
     """
     Load Empatica .avro file.
 
@@ -63,8 +62,6 @@ def load_empatica_data(file_path: str | Path, verbose: int = 1) -> pd.DataFrame:
     ----------
     file_path : str or Path
         Path to .avro file.
-    verbose : int, default 1
-        Logging verbose level.
 
     Returns
     -------
@@ -73,8 +70,7 @@ def load_empatica_data(file_path: str | Path, verbose: int = 1) -> pd.DataFrame:
         gyroscope_x/y/z (if available).
     """
     file_path = Path(file_path)
-    if verbose >= 1:
-        logger.info(f"Loading Empatica data from {file_path}")
+    logger.info(f"Loading Empatica data from {file_path}")
 
     with open(file_path, "rb") as f:
         reader = DataFileReader(f, DatumReader())
@@ -151,16 +147,16 @@ def load_empatica_data(file_path: str | Path, verbose: int = 1) -> pd.DataFrame:
 
     df = pd.DataFrame(df_data)
 
-    if verbose >= 1:
-        logger.info(f"Loaded Empatica data: {nrows} rows at {sampling_frequency} Hz")
-    if verbose >= 2:
-        logger.info(f"Start time: {pd.to_datetime(t_start, unit='us')}")
-        logger.info(f"Columns: {list(df.columns)}")
+    logger.info(f"Loaded Empatica data: {nrows} rows at {sampling_frequency} Hz")
+    logger.debug(f"Start time: {pd.to_datetime(t_start, unit='us')}")
+    logger.debug(f"Columns: {list(df.columns)}")
 
     return df
 
 
-def load_axivity_data(file_path: str | Path, verbose: int = 1) -> pd.DataFrame:
+def load_axivity_data(
+    file_path: str | Path,
+) -> pd.DataFrame:
     """
     Load Axivity .CWA file.
 
@@ -168,8 +164,6 @@ def load_axivity_data(file_path: str | Path, verbose: int = 1) -> pd.DataFrame:
     ----------
     file_path : str or Path
[…]
