+++ src/paradigma/orchestrator.py
@@ -53,11 +53,17 @@ from paradigma.prepare_data import prepare_raw_data
 
 logger = logging.getLogger(__name__)
 
+# Custom logging level for detailed info (between INFO=20 and DEBUG=10)
+DETAILED_INFO = 15
+logging.addLevelName(DETAILED_INFO, "DETAILED")
+
 
 def run_paradigma(
-    output_dir: str | Path = "./output",
+    *,
     data_path: str | Path | None = None,
     dfs: pd.DataFrame | list[pd.DataFrame] | dict[str, pd.DataFrame] | None = None,
+    save_intermediate: list[str] = [],
+    output_dir: str | Path = "./output",
     skip_preparation: bool = False,
     pipelines: list[str] | str | None = None,
     watch_side: str | None = None,
@@ -67,7 +73,6 @@ def run_paradigma(
     target_frequency: float = 100.0,
     column_mapping: dict[str, str] | None = None,
     device_orientation: list[str] | None = ["x", "y", "z"],
-    save_intermediate: list[str] = [],
     file_pattern: str | list[str] | None = None,
     aggregates: list[str] | None = None,
     segment_length_bins: list[str] | None = None,
@@ -80,7 +85,8 @@ def run_paradigma(
     arm_activity_config: GaitConfig | None = None,
     tremor_config: TremorConfig | None = None,
     pulse_rate_config: PulseRateConfig | None = None,
-    verbose: int = 1,
+    logging_level: int = logging.INFO,
+    custom_logger: logging.Logger | None = None,
 ) -> dict[str, pd.DataFrame | dict]:
     """
     Complete ParaDigMa analysis pipeline from data loading to aggregated results.
@@ -109,6 +115,17 @@ def run_paradigma(
         Note: The 'file_key' column is only added to quantification results when
         len(dfs) > 1, allowing cleaner output for single-file processing.
         See input_formats guide for details.
+    save_intermediate : list of str, default []
+        Which intermediate results to store. Valid values:
+        - 'preparation': Save prepared data
+        - 'preprocessing': Save preprocessed signals
+        - 'classification': Save classification results
+        - 'quantification': Save quantified measures
+        - 'aggregation': Save aggregated results
+        If empty, no files are saved (results are only returned).
+    output_dir : str or Path, default './output'
+        Output directory for all results. Files are only saved if
+        save_intermediate is not empty.
     skip_preparation : bool, default False
         Whether data is already prepared. If False, data will be
         prepared (unit conversion, resampling, etc.). If True,
@@ -131,17 +148,6 @@ def run_paradigma(
         Custom column name mapping.
     device_orientation : list of str, optional
         Custom device orientation corrections.
-    output_dir : str or Path, default './output'
-        Output directory for all results. Files are only saved if
-        save_intermediate is not empty.
-    save_intermediate : list of str, default []
-        Which intermediate results to store. Valid values:
-        - 'preparation': Save prepared data
-        - 'preprocessing': Save preprocessed signals
-        - 'classification': Save classification results
-        - 'quantification': Save quantified measures
-        - 'aggregation': Save aggregated results
-        If empty, no files are saved (results are only returned).
     file_pattern : str or list of str, optional
         File pattern(s) to match when loading data (e.g., 'parquet', '*.csv').
     aggregates : list of str, optional
@@ -173,12 +179,16 @@ def run_paradigma(
         Tremor analysis configuration.
     pulse_rate_config : PulseRateConfig, optional
         Pulse rate analysis configuration.
-    verbose : int, default 1
-        Logging verbose level:
-        - 0: Only errors and warnings
-        - 1: Basic info (default)
-        - 2: Detailed info with progress
-        - 3: Debug level with all details
[…]
