### 2026-02-04 8c2a115
Feature/pipeline runner with column mapping (#251)
feat: add pipeline orchestrator with centralized logging and data preparation (#251)

Implement end-to-end ParaDigMa analysis pipeline with automatic data preparation,
centralized logging system, and improved documentation.

Major changes:
- Add orchestrator.py with run_paradigma() for complete analysis pipeline
- Add prepare_data.py for automatic data preparation with column mapping and validation
- Add load.py to centralize data loading capabilities
- Replace verbose parameter with standard logging across all pipeline functions
- Add GAIT_SE
[…]
