+++ tests/test_pulse_rate_analysis.py
@@ -1,36 +1,95 @@
+import tempfile
 from pathlib import Path
 
-from test_notebooks import compare_data
+import numpy as np
+import pandas as pd
+import pytest
 
-from paradigma.config import IMUConfig, PPGConfig
-from paradigma.testing import preprocess_ppg_data_io
+from paradigma.config import IMUConfig, PPGConfig, PulseRateConfig
+from paradigma.constants import DataColumns
+from paradigma.orchestrator import run_paradigma
+from paradigma.pipelines.pulse_rate_pipeline import run_pulse_rate_pipeline
 
-# Mappings between the metadata and binary files
+try:
+    from test_notebooks import compare_data
 
-accelerometer_preproc_pairs: list[tuple[str, str]] = [
-    ("accelerometer_meta.json", "accelerometer_values.bin"),
-    ("accelerometer_meta.json", "accelerometer_time.bin"),
-]
-ppg_preproc_pairs: list[tuple[str, str]] = [
-    ("PPG_meta.json", "PPG_values.bin"),
-    ("PPG_meta.json", "PPG_time.bin"),
-]
-ppg_features_pairs: list[tuple[str, str]] = [
-    ("features_ppg_meta.json", "features_ppg_values.bin"),
-    ("features_ppg_meta.json", "features_ppg_time.bin"),
-]
+    from paradigma.testing import preprocess_ppg_data_io
 
-accelerometer_features_pairs: list[tuple[str, str]] = [
-    ("features_acc_meta.json", "features_acc_values.bin"),
-    ("features_acc_meta.json", "features_acc_time.bin"),
-]
+    HAS_TESTING_UTILS = True
+except ImportError:
+    HAS_TESTING_UTILS = False
 
 
+def create_test_pulse_rate_data(duration_minutes=2):
+    """Create simple test data for pulse rate analysis."""
+    duration_seconds = duration_minutes * 60
+    n_samples = int(duration_seconds * 50.0)  # 50 Hz
+    time = np.linspace(0, duration_seconds, n_samples)
+
+    # Simple PPG signal
+    heart_rate_hz = 70.0 / 60.0  # 70 BPM
+    ppg_signal = np.sin(2 * np.pi * heart_rate_hz * time) + 0.1 * np.random.random(
+        n_samples
+    )
+
+    # Simple accelerometer data
+    acc_x = 0.1 * np.random.random(n_samples)
+    acc_y = 0.1 * np.random.random(n_samples)
+    acc_z = 9.8 + 0.1 * np.random.random(n_samples)
+
+    return pd.DataFrame(
+        {
+            DataColumns.TIME: time,
+            DataColumns.PPG: ppg_signal,
+            DataColumns.ACCELEROMETER_X: acc_x,
+            DataColumns.ACCELEROMETER_Y: acc_y,
+            DataColumns.ACCELEROMETER_Z: acc_z,
+        }
+    )
+
+
+def test_pulse_rate_pipeline_basic():
+    """Test basic pulse rate pipeline functionality."""
+    df_test = create_test_pulse_rate_data()
+
+    with tempfile.TemporaryDirectory() as temp_dir:
+        result = run_pulse_rate_pipeline(
+            df_ppg_prepared=df_test,
+            output_dir=temp_dir,
+            pulse_rate_config=PulseRateConfig(),
+            ppg_config=PPGConfig(),
+        )
+
+        assert isinstance(result, pd.DataFrame)
+        assert DataColumns.TIME in result.columns
+
+
+def test_pulse_rate_pipeline_integration():
+    """Test pulse rate pipeline with orchestrator."""
+    dfs = {"test": create_test_pulse_rate_data()}
+
+    with tempfile.TemporaryDirectory() as temp_dir:
[…]
