+++ tests/test_paradigma.py
@@ -3,7 +3,8 @@ import tsdf
 
 def test_transform_time_array(shared_datadir):
     """
-    The initial test to check if the preprocessing function works as expected. It checks the output dimensions and the type of the output.
+    The initial test to check if the preprocessing function works as expected.
+    It checks the output dimensions and the type of the output.
     """
     metadata_dict = tsdf.load_metadata_from_path(
         shared_datadir / "1.prepared_data/ppg/PPG_meta.json"
