+++ src/paradigma/segmenting.py
@@ -1,5 +1,3 @@
-from typing import List
-
 import numpy as np
 import pandas as pd
 
@@ -9,19 +7,21 @@ from paradigma.util import deprecated
 
 def tabulate_windows(
     df: pd.DataFrame,
-    columns: List[str],
+    columns: list[str],
     window_length_s: float,
     window_step_length_s: float,
     fs: int,
 ) -> np.ndarray:
     """
-    Split the given DataFrame into overlapping windows of specified length and step size.
+    Split the given DataFrame into overlapping windows of specified length
+    and step size.
 
-    This function extracts windows of data from the specified columns of the DataFrame, based on
-    the window length and step size provided in the configuration. The windows are returned in
-    a 3D NumPy array, where the first dimension represents the window index, the second dimension
-    represents the time steps within the window, and the third dimension represents the columns
-    of the data.
+    This function extracts windows of data from the specified columns of the
+    DataFrame, based on the window length and step size provided in the
+    configuration. The windows are returned in a 3D NumPy array, where the
+    first dimension represents the window index, the second dimension
+    represents the time steps within the window, and the third dimension
+    represents the columns of the data.
 
     Parameters
     ----------
@@ -40,17 +40,22 @@ def tabulate_windows(
     -------
     np.ndarray
         A 3D NumPy array of shape (n_windows, window_size, n_columns), where:
-        - `n_windows` is the number of windows that can be formed from the data.
-        - `window_size` is the length of each window in terms of the number of time steps.
-        - `n_columns` is the number of columns in the input DataFrame specified by `columns`.
+        - `n_windows` is the number of windows that can be formed from the
+          data.
+        - `window_size` is the length of each window in terms of the number
+          of time steps.
+        - `n_columns` is the number of columns in the input DataFrame
+          specified by `columns`.
 
-        If the length of the data is shorter than the specified window size, an empty array is returned.
+        If the length of the data is shorter than the specified window size,
+        an empty array is returned.
 
     Notes
     -----
-    This function uses `np.lib.stride_tricks.sliding_window_view` to generate sliding windows of data.
-    The step size is applied to extract windows at intervals.
-    If the data is insufficient for at least one window, an empty array will be returned.
+    This function uses `np.lib.stride_tricks.sliding_window_view` to
+    generate sliding windows of data. The step size is applied to extract
+    windows at intervals. If the data is insufficient for at least one
+    window, an empty array will be returned.
 
     Example
     -------
@@ -84,7 +89,8 @@ def tabulate_windows(
 
 def tabulate_windows_legacy(config, df, agg_func="first"):
     """
-    Efficiently creates a windowed dataframe from the input dataframe using vectorized operations.
+    Efficiently creates a windowed dataframe from the input dataframe using
+    vectorized operations.
 
     Parameters
     ----------
@@ -93,11 +99,13 @@ def tabulate_windows_legacy(config, df, agg_func="first"):
         - `window_length_s`: The number of seconds per window.
         - `window_step_length_s`: The number of seconds to shift between windows.
         - `sampling_frequency`: The sampling frequency in Hz.
-        - `single_value_colnames`: List of column names where a single value (e.g., mean) is needed.
-        - `list_value_colnames`: List of column names where all 600 values should be stored in a list.
+        - `single_value_colnames`: List of column names where a single value
+          (e.g., mean) is needed.
+        - `list_value_colnames`: List of column names where all 600 values
[…]
