### 2026-02-04 a47384f
Update dependencies, bump to v1.0.5
