### 2026-02-04 ce7e61e
Update dependencies, bump to v1.0.5
