### 2026-02-04 a63c5df
Update citation and changelog
