### 2026-02-04 991af94
Update citation and changelog
