### 2026-02-04 4d4d0c9
Update dependencies, bump to v1.1.0 (#252)
* Update dependencies, bump to v1.1.0

* Update citation and changelog
