### 2026-02-05 b5de9d5
Added instructions on how to use kernels in tutorials
