### 2026-02-05 biomarkersParkinson/paradigma#250 @Erikpostt
Hey @ntolley, the changes should be working on `main` now. I'm currently awaiting approval for merging into `release` all other review-related changes. Feel free to pull `main`, or wait a day and pull `release`. I'm posted a more elaborate explanation in the JOSS review after merging into `release`.

For instructions on `git-lfs` issues, please see [installation guide](https://github.com/biomarkersParkinson/paradigma/blob/main/docs/guides/installation.md).
