### 2026-02-05 f3485a1
Release v1.1.0 (#255)
* Merge dev branch into release
