+++ src/paradigma/util.py
@@ -2,7 +2,6 @@ import functools
 import os
 import warnings
 from datetime import datetime, timedelta
-from typing import List, Optional, Tuple
 
 import numpy as np
 import pandas as pd
@@ -16,12 +15,14 @@ from paradigma.constants import DataColumns, TimeUnit
 
 def deprecated(reason: str = ""):
     """
-    Decorator to mark functions as deprecated. It will show a warning when the function is used.
+    Decorator to mark functions as deprecated. It will show a warning when the
+    function is used.
 
     Parameters
     ----------
     reason : str, optional
-        Additional message to explain why it is deprecated and what to use instead.
+        Additional message to explain why it is deprecated and what to use
+        instead.
     """
 
     def decorator(func):
@@ -155,7 +156,7 @@ def write_df_data(
 
 def read_metadata(
     input_path: str, meta_filename: str, time_filename: str, values_filename: str
-) -> Tuple[TSDFMetadata, TSDFMetadata]:
+) -> tuple[TSDFMetadata, TSDFMetadata]:
     metadata_dict = tsdf.load_metadata_from_path(
         os.path.join(input_path, meta_filename)
     )
@@ -186,8 +187,8 @@ def load_tsdf_dataframe(
 
 
 def load_metadata_list(
-    dir_path: str, meta_filename: str, filenames: List[str]
-) -> List[TSDFMetadata]:
+    dir_path: str, meta_filename: str, filenames: list[str]
+) -> list[TSDFMetadata]:
     """
     Load the metadata objects from a metadata file according to the specified binaries.
 
@@ -216,7 +217,8 @@ def transform_time_array(
     start_time: float = 0.0,
 ) -> np.ndarray:
     """
-    Transforms the time array to relative time (when defined in delta time) and scales the values.
+    Transforms the time array to relative time (when defined in delta time)
+    and scales the values.
 
     Parameters
     ----------
@@ -225,7 +227,8 @@ def transform_time_array(
     input_unit_type : str
         The time unit type of the input time array.
     output_unit_type : str
-        The time unit type of the output time array. ParaDigMa expects `TimeUnit.RELATIVE_S`.
+        The time unit type of the output time array. ParaDigMa expects
+        `TimeUnit.RELATIVE_S`.
     start_time : float, optional
         The start time of the time array in UNIX seconds (default is 0.0)
 
@@ -236,9 +239,13 @@ def transform_time_array(
 
     Notes
     -----
-    - The function handles different time units (`TimeUnit.RELATIVE_MS`, `TimeUnit.RELATIVE_S`, `TimeUnit.ABSOLUTE_MS`, `TimeUnit.ABSOLUTE_S`, `TimeUnit.DIFFERENCE_MS`, `TimeUnit.DIFFERENCE_S`).
-    - The transformation allows for scaling of the time array, converting between time unit types (e.g., relative, absolute, or difference).
-    - When converting to `TimeUnit.RELATIVE_MS`, the function calculates the relative time starting from the provided or default start time.
+    - The function handles different time units (`TimeUnit.RELATIVE_MS`,
+      `TimeUnit.RELATIVE_S`, `TimeUnit.ABSOLUTE_MS`, `TimeUnit.ABSOLUTE_S`,
+      `TimeUnit.DIFFERENCE_MS`, `TimeUnit.DIFFERENCE_S`).
+    - The transformation allows for scaling of the time array, converting
+      between time unit types (e.g., relative, absolute, or difference).
+    - When converting to `TimeUnit.RELATIVE_MS`, the function calculates the
+      relative time starting from the provided or default start time.
     """
     input_units = input_unit_type.split("_")[-1].lower()
[…]
