+++ src/paradigma/load.py
@@ -1,0 +1,476 @@
+"""
+Data loading module for ParaDigMa toolbox.
+
+This module provides functions to load sensor data from various formats:
+- Raw data: TSDF (.meta/.bin), Empatica (.avro), Axivity (.CWA)
+- Prepared data: parquet, pickle, csv
+
+Based on device_specific_data_loading tutorial.
+"""
+
+import logging
+import pickle
+from pathlib import Path
+
+import pandas as pd
+from avro.datafile import DataFileReader
+from avro.io import DatumReader
+
+from paradigma.util import load_tsdf_dataframe
+
+logger = logging.getLogger(__name__)
+
+
+def load_tsdf_data(
+    data_path: str | Path,
+    prefix: str = "IMU",
+) -> tuple[pd.DataFrame, dict, dict]:
+    """
+    Load TSDF data from .meta and .bin files.
+
+    Parameters
+    ----------
+    data_path : str or Path
+        Path to directory containing TSDF files.
+    prefix : str, default "IMU"
+        Prefix for TSDF files (e.g., "IMU_segment0001").
+
+    Returns
+    -------
+    tuple
+        Tuple containing (DataFrame with loaded data, time metadata
+        dict, values metadata dict)
+    """
+    data_path = Path(data_path)
+    logger.info(f"Loading TSDF data from {data_path} with prefix '{prefix}'")
+
+    df, time_meta, values_meta = load_tsdf_dataframe(
+        path_to_data=data_path, prefix=prefix
+    )
+
+    logger.info(f"Loaded TSDF data: {df.shape[0]} rows, {df.shape[1]} columns")
+    return df, time_meta, values_meta
+
+
+def load_empatica_data(
+    file_path: str | Path,
+) -> pd.DataFrame:
+    """
+    Load Empatica .avro file.
+
+    Parameters
+    ----------
+    file_path : str or Path
+        Path to .avro file.
+
+    Returns
+    -------
+    pd.DataFrame
+        DataFrame with columns: time, time_dt, accelerometer_x/y/z,
+        gyroscope_x/y/z (if available).
+    """
+    file_path = Path(file_path)
+    logger.info(f"Loading Empatica data from {file_path}")
+
+    with open(file_path, "rb") as f:
+        reader = DataFileReader(f, DatumReader())
+        empatica_data = next(reader)
+
+    accel_data = empatica_data["rawData"]["accelerometer"]
+
+    # Check for gyroscope data
+    gyro_data = None
+    if (
+        "gyroscope" in empatica_data["rawData"]
+        and len(empatica_data["rawData"]["gyroscope"]["x"]) > 0
+    ):
+        gyro_data = empatica_data["rawData"]["gyroscope"]
+    else:
+        raise ValueError("Gyroscope data not found in Empatica file.")
+
+    # Check Avro schema version for conversion
+    avro_version = (
+        empatica_data["schemaVersion"]["major"],
+        empatica_data["schemaVersion"]["minor"],
+        empatica_data["schemaVersion"]["patch"],
+    )
+
+    # Convert accelerometer data based on schema version
+    if avro_version < (6, 5, 0):
+        physical_range = (
+            accel_data["imuParams"]["physicalMax"]
+            - accel_data["imuParams"]["physicalMin"]
+        )
+        digital_range = (
+            accel_data["imuParams"]["digitalMax"]
+            - accel_data["imuParams"]["digitalMin"]
+        )
+        conversion_factor = physical_range / digital_range
+    else:
+        conversion_factor = accel_data["imuParams"]["conversionFactor"]
+
+    accel_x = [val * conversion_factor for val in accel_data["x"]]
[…]
