+++ _downloads/cb20967773053f16d9849c6fb63e4159/segmenting.py
@@ -1,0 +1,439 @@
+import numpy as np
+import pandas as pd
+
+from paradigma.constants import DataColumns
+from paradigma.util import deprecated
+
+
+def tabulate_windows(
+    df: pd.DataFrame,
+    columns: list[str],
+    window_length_s: float,
+    window_step_length_s: float,
+    fs: int,
+) -> np.ndarray:
+    """
+    Split the given DataFrame into overlapping windows of specified length
+    and step size.
+
+    This function extracts windows of data from the specified columns of the
+    DataFrame, based on the window length and step size provided in the
+    configuration. The windows are returned in a 3D NumPy array, where the
+    first dimension represents the window index, the second dimension
+    represents the time steps within the window, and the third dimension
+    represents the columns of the data.
+
+    Parameters
+    ----------
+    df : pd.DataFrame
+        The input DataFrame containing the data to be windowed.
+    columns : list of str
+        A list of column names from the DataFrame that will be used for windowing.
+    window_length_s : float
+        The length of each window in seconds.
+    window_step_length_s : float
+        The step size between consecutive windows in seconds.
+    fs : int
+        The sampling frequency of the data in Hz.
+
+    Returns
+    -------
+    np.ndarray
+        A 3D NumPy array of shape (n_windows, window_size, n_columns), where:
+        - `n_windows` is the number of windows that can be formed from the
+          data.
+        - `window_size` is the length of each window in terms of the number
+          of time steps.
+        - `n_columns` is the number of columns in the input DataFrame
+          specified by `columns`.
+
+        If the length of the data is shorter than the specified window size,
+        an empty array is returned.
+
+    Notes
+    -----
+    This function uses `np.lib.stride_tricks.sliding_window_view` to
+    generate sliding windows of data. The step size is applied to extract
+    windows at intervals. If the data is insufficient for at least one
+    window, an empty array will be returned.
+
+    Example
+    -------
+    config = Config(window_length_s=5, window_step_length_s=1, sampling_frequency=100)
+    df = pd.DataFrame({'col1': np.random.randn(100), 'col2': np.random.randn(100)})
+    columns = ['col1', 'col2']
+    windows = tabulate_windows(config, df, columns)
+    """
+    window_size = int(window_length_s * fs)
+    window_step_size = int(window_step_length_s * fs)
+    n_columns = len(columns)
+
+    data = df[columns].values
+
+    # Check if data length is sufficient
+    if len(data) < window_size:
+        return np.empty(
+            (0, window_size, n_columns)
+        )  # Return an empty array if insufficient data
+
+    windows = np.lib.stride_tricks.sliding_window_view(
+        data, window_shape=(window_size, n_columns)
+    )[::window_step_size].squeeze()
+
+    # Ensure 3D shape (n_windows, window_size, n_columns)
+    if windows.ndim == 2:  # Single window case
+        windows = windows[np.newaxis, :, :]  # Add a new axis at the start
+
+    return windows
+
+
+def tabulate_windows_legacy(config, df, agg_func="first"):
+    """
+    Efficiently creates a windowed dataframe from the input dataframe using
+    vectorized operations.
+
+    Parameters
+    ----------
+    config : object
+        A configuration object containing:
+        - `window_length_s`: The number of seconds per window.
+        - `window_step_length_s`: The number of seconds to shift between windows.
+        - `sampling_frequency`: The sampling frequency in Hz.
+        - `single_value_colnames`: List of column names where a single value
+          (e.g., mean) is needed.
+        - `list_value_colnames`: List of column names where all 600 values
+          should be stored in a list.
+    agg_func : str or callable, optional
[…]
