+++ _sources/autoapi/paradigma/index.rst.txt
@@ -3,6 +3,11 @@ paradigma
 
 .. py:module:: paradigma
 
+.. autoapi-nested-parse::
+
+   ParaDigMa: Parkinson Digital Biomarker Analysis Toolbox
+
+
 
 Submodules
 ----------
@@ -14,7 +19,10 @@ Submodules
    /autoapi/paradigma/config/index
    /autoapi/paradigma/constants/index
    /autoapi/paradigma/feature_extraction/index
+   /autoapi/paradigma/load/index
+   /autoapi/paradigma/orchestrator/index
    /autoapi/paradigma/pipelines/index
+   /autoapi/paradigma/prepare_data/index
    /autoapi/paradigma/preprocessing/index
    /autoapi/paradigma/segmenting/index
    /autoapi/paradigma/testing/index
