+++ _sources/autoapi/paradigma/pipelines/pulse_rate_utils/index.rst.txt
@@ -41,7 +41,7 @@ Module Contents
    :rtype: np.ndarray
 
 
-.. py:function:: extract_pr_segments(sqa_label: numpy.ndarray, min_pr_samples: int) -> Tuple[numpy.ndarray, numpy.ndarray]
+.. py:function:: extract_pr_segments(sqa_label: numpy.ndarray, min_pr_samples: int) -> tuple[numpy.ndarray, numpy.ndarray]
 
    Extracts pulse rate segments based on the SQA label.
 
@@ -58,16 +58,19 @@ Module Contents
 
    Extracts pulse rate from the time-frequency distribution of the PPG signal.
 
-   :param ppg: The preprocessed PPG segment with 2 seconds of padding on both sides to reduce boundary effects.
+   :param ppg: The preprocessed PPG segment with 2 seconds of padding on both sides
+               to reduce boundary effects.
    :type ppg: np.ndarray
-   :param tfd_length: Length of each segment (in seconds) to calculate the time-frequency distribution.
+   :param tfd_length: Length of each segment (in seconds) to calculate the time-frequency
+                      distribution.
    :type tfd_length: int
    :param fs: The sampling frequency of the PPG signal.
    :type fs: int
-   :param kern_type: Type of TFD kernel to use (e.g., 'wvd' for Wigner-Ville distribution).
+   :param kern_type: Type of TFD kernel to use (e.g., 'wvd' for Wigner-Ville
+                     distribution).
    :type kern_type: str
-   :param kern_params: Parameters for the specified kernel. Not required for 'wvd', but relevant for other
-                       kernels like 'spwvd' or 'swvd'. Default is None.
+   :param kern_params: Parameters for the specified kernel. Not required for 'wvd', but
+                       relevant for other kernels like 'spwvd' or 'swvd'. Default is None.
    :type kern_params: dict
 
    :returns: The estimated pulse rate.
@@ -76,8 +79,8 @@ Module Contents
 
 .. py:function:: extract_pr_with_tfd(ppg: numpy.ndarray, fs: int, kern_type: str, kern_params: dict) -> numpy.ndarray
 
-   Estimate pulse rate (PR) from a PPG segment using a TFD method with optional
-   moving average filtering.
+   Estimate pulse rate (PR) from a PPG segment using a TFD method with
+   optional moving average filtering.
 
    :param ppg_segment: Segment of the PPG signal to process.
    :type ppg_segment: array-like
@@ -88,7 +91,8 @@ Module Contents
    :param kern_params: Parameters for the specified kernel.
    :type kern_params: dict
 
-   :returns: **pr_smooth_tfd** -- Estimated pr values (in beats per minute) for each 2-second segment of the PPG signal.
+   :returns: **pr_smooth_tfd** -- Estimated pr values (in beats per minute) for each 2-second segment
+             of the PPG signal.
    :rtype: np.ndarray
 
 
@@ -96,7 +100,8 @@ Module Contents
 
    .. py:method:: nonsep_gdtfd(x: numpy.ndarray, kern_type: None | str = None, kern_params: None | dict = None)
 
-      Computes the generalized time-frequency distribution (TFD) using a non-separable kernel.
+      Computes the generalized time-frequency distribution (TFD) using a
+      non-separable kernel.
 
       Parameters:
       -----------
@@ -113,16 +118,20 @@ Module Contents
               sep  - kernel for separable kernel (combintation of SWVD and PWVD)
 
       kern_params : dict, optional
-          Dictionary of parameters specific to the kernel type. Default is None.
-          The structure of the dictionary depends on the selected kernel type:
+          Dictionary of parameters specific to the kernel type. Default is
+          None. The structure of the dictionary depends on the selected
+          kernel type:
               - wvd:
               An empty dictionary, as no additional parameters are required.
               - swvd:
[…]
