### 2026-02-05 e409e47
Solve openmovement pypi conflict (#257)
* Solve dependency conflicts and PyPi issue
