### 2026-02-05 #9502 @Erikpostt
Hi @ntolley @Asanto32 @richrobe @Kevin-Mattheus-Moerman,

Thank you for your patience; it took a while for all changes to be reviewed. Your suggestions required substantial additions to the codebase, which we had to validate carefully. We have created a new release `v1.1.0` based on these changes and are looking forward to your response.

This is a high-level overview of the most relevant changes:

---

* We have created an **orchestrator entry point** `run_paradigma()` that can be used as follows:

```python
from paradigma import run_paradigma

# Run complete pipeline on your data
results = run_paradigma(
    data_path="path/to/your/data",
    pipelines=["gait", "tremor"],
)
```

This can further be customized. See the [Pipeline Orchestrator tutorial](https://biomarkersparkinson.github.io/paradigma/tutorials/pipeline_orchestrator.html) for details.

**Features**
* **Pipeline Orchestrator** ([tutorial](https://biomarkersparkinson.github.io/paradigma/tutorials/pipeline_orchestrator.html)): Added `run_paradigma()` function for complete end-to-end analysis from data loading to aggregated results
* **Automatic Data Preparation**: New [`prepare_data.py`](https://github.com/biomarkersParkinson/paradigma/blob/main/src/paradigma/prepare_data.py) module with flexible column mapping
* **Centralized Data Loading**: New [`load.py`](https://github.com/biomarkersParkinson/paradigma/blob/main/src/paradigma/load.py) module supporting multiple formats (TSDF, Parquet, CSV, Pickle, CWA, AVRO)
* **High-Level Pipeline Functions**: Added `run_gait_pipeline()`, `run_tremor_pipeline()`, and `run_pulse_rate_pipeline()` for simplified usage
* **Device-Specific Support**: [Tutorial](https://biomarkersparkinson.github.io/paradigma/tutorials/device_specific_data_loading.html) and sample data for Axivity (.CWA) and Empatica (.AVRO) parsing

Improvements
* **Documentation**:
    * Added tutorials landing page with improved navigation
    * New guides: data_input, installation, sensor_requirements, supported_devices
    * Added pipeline orchestrator tutorial for end-to-end processing
    * Updated README with direct links to tutorials and new publications

* **Code Quality**:
    * Limited line length to 88 characters per PEP8 standards
    * Updated all docstrings to PEP standards
---

Now let me connect changes made to your individual suggestions.

@ntolley
* Can you check whether using LFS is now working on your side (refers to [this issue](https://github.com/biomarkersParkinson/paradigma/issues/250))? We have migrated large data files to Git LFS. If cloning still fails, please follow the [installation guide](https://biomarkersparkinson.github.io/paradigma/guides/installation.html) for LFS setup instructions.
* We believe we've addressed all major points raised, but given the scope of changes, we'd appreciate your feedback on whether anything needs further adjustment.

@Asanto32
* Indeed `ruff` did not catch these long lines, so we've fixed line length to 88 characters in `pyproject.toml`.
* We've removed claims from the paper and documentation that ParaDigMa works with any raw data file format directly.
* We have added `prepare_data()` functionality and integrated it into `run_paradigma()` for minimal preprocessing support.
* **Currently supported formats**: TSDF, Parquet, CSV, Pickle, CWA (Axivity), AVRO (Empatica), and Verily JSON/.bin
* **Not yet supported**: GENEActiv (.bin), Actigraph (.gt3x), and other proprietary formats
* We're willing to add support for additional formats. If you can share sample data (even synthetic/anonymized), we'd be happy to implement support for GENEActiv and Actigraph.
* Regarding the long tutorial series: We've added `run_paradigma()` as a single-function entry point that combines all pipelines. The individual tutorials remain available for users who want to understand the details or customize specific steps, but are no longer required for basic usage.

@richrobe
* We have added `run_paradigma` as an entrypoint to ParaDigMa, with instructions in the readme and tutorials.
* We have added instructions on how to install `git-lfs` in `CONTRIBUTING.md` and the [installation guide](https://biomarkersparkinson.github.io/paradigma/guides/installation.html),.
* Links to the tutorials have been added to the Readme.
* How to start using notebooks and registering a kernel is explained in the tutorials landing page.
* We have added visualizations to the `paper` branch by adding generated data and notebooks with weekly aggregations. These shall be added to Paradigma once the paper is published in JOSS.
* Instructions on how to run tests has been added to the `CONTRIBUTING.md`.

I hope all of your suggestions have been implemented correctly. I'm looking forward to hearing your feedback. Please let me know if further adjustments are necessary.
