+++ src/paradigma/orchestrator.py
@@ -283,9 +283,15 @@ def run_paradigma(
 
     # Initialize results storage for each pipeline
     all_results = {
-        "quantifications": {p: [] for p in pipelines},
+        "quantifications": {
+            p: [] if p != "gait" else {"filtered": [], "unfiltered": []}
+            for p in pipelines
+        },
         "aggregations": {p: {} for p in pipelines},
-        "metadata": {p: {} for p in pipelines},
+        "metadata": {
+            p: {} if p != "gait" else {"filtered": {}, "unfiltered": {}}
+            for p in pipelines
+        },
         "errors": [],
     }
 
@@ -374,53 +380,61 @@ def run_paradigma(
 
                 try:
                     if pipeline_name == "gait":
-                        quantification_data, quantification_metadata = (
-                            run_gait_pipeline(
-                                df_prepared=df_prepared,
-                                watch_side=watch_side,
-                                imu_config=imu_config,
-                                gait_config=gait_config,
-                                arm_activity_config=arm_activity_config,
-                                store_intermediate=store_intermediate_per_file,
-                                output_dir=file_output_dir,
-                                segment_number_offset=max_gait_segment_nr,
-                                logging_level=logging_level,
-                                custom_logger=active_logger,
-                            )
+                        quantifications_dict, metadata_dict = run_gait_pipeline(
+                            df_prepared=df_prepared,
+                            watch_side=watch_side,
+                            imu_config=imu_config,
+                            gait_config=gait_config,
+                            arm_activity_config=arm_activity_config,
+                            store_intermediate=store_intermediate_per_file,
+                            output_dir=file_output_dir,
+                            segment_number_offset=max_gait_segment_nr,
+                            logging_level=logging_level,
+                            custom_logger=active_logger,
                         )
 
-                        if len(quantification_data) > 0:
-                            # Add file identifier if processing multiple files
-                            quantification_data = quantification_data.copy()
-                            if num_files > 1:
-                                quantification_data["file_key"] = file_name
-                            all_results["quantifications"][pipeline_name].append(
-                                quantification_data
-                            )
+                        # Process both filtered and unfiltered quantifications
+                        for quant_type in ["filtered", "unfiltered"]:
+                            quantification_data = quantifications_dict[quant_type]
+                            quantification_metadata = metadata_dict[quant_type]
 
-                            # Update max segment number for next file
-                            max_gait_segment_nr = int(
-                                quantification_data["gait_segment_nr"].max()
-                            )
+                            if len(quantification_data) > 0:
+                                # Add file identifier if processing multiple files
+                                quantification_data = quantification_data.copy()
+                                if num_files > 1:
+                                    quantification_data["file_key"] = file_name
+                                all_results["quantifications"][pipeline_name][
+                                    quant_type
+                                ].append(quantification_data)
 
-                        # Store metadata and update offset even if no quantifications
-                        if (
-                            quantification_metadata
-                            and "per_segment" in quantification_metadata
-                        ):
-                            all_results["metadata"][pipeline_name].update(
-                                quantification_metadata["per_segment"]
-                            )
+                                # Update max segment number for next file (use filtered)
+                                if quant_type == "filtered":
+                                    max_gait_segment_nr = int(
+                                        quantification_data["gait_segment_nr"].max()
+                                    )
 
-                            # Update max segment number based on metadata to prevent
-                            # overwrites
[…]
