### 2026-02-06 3a2f9cd
Fix orchestrator to pass required segment_meta and segment_cats to aggregate_arm_swing_params
- Added parsing of segment_length_bins to segment_cats tuples
- Pass segment_meta from metadata dictionary to aggregation function
- Applies to both filtered and unfiltered gait aggregations
