### 2026-02-08 #9502 @Erikpostt
@Kevin-Mattheus-Moerman can you see why the editorialbot is unable to generate a pdf from the markdown file?
