### 2026-02-09 c2ca043
Reduce size tutorials, update docs
