+++ pyproject.toml
@@ -22,6 +22,7 @@ python-dateutil = "^2.9.0.post0"
 nbconvert = "^7.16.6"
 pyarrow = "^22.0.0"
 avro = "^1.12.1"
+numpy = "^2.0.0"
 
 [tool.poetry.group.testing.dependencies]
 ipykernel = "^6.27.1"
