### 2026-02-09 06dff26
Implement code review suggestions: fix early return, normalize DataFrames, support tuple bins, update docs
Co-authored-by: Erikpostt <57133568+Erikpostt@users.noreply.github.com>
