### 2026-02-10 ede90e9
Fix code review issues: unfiltered gait quantification, DataFrame normalization, and segment_length_bins parsing (#265)
* Implement code review suggestions: fix early return, normalize DataFrames, support tuple bins, update docs

Co-authored-by: Erikpostt <57133568+Erikpostt@users.noreply.github.com>

* Solve pre-commit hook issues

---------

Co-authored-by: copilot-swe-agent[bot] <198982749+Copilot@users.noreply.github.com>
Co-authored-by: Erikpostt <57133568+Erikpostt@users.noreply.github.com>
