### 2026-02-10 f932ec3
Resolve copilot issues
