### 2026-02-10 9d2b0ba
Resolve second round of copilot suggestions
