### 2026-02-10 5052f85
Limit pyarrow version for stability
