+++ pyproject.toml
@@ -20,7 +20,7 @@ scikit-learn = ">=1.3.2,<1.6.1"
 tsdf = "^0.6.1"
 python-dateutil = "^2.9.0.post0"
 nbconvert = "^7.16.6"
-pyarrow = ">=7.0"
+pyarrow = ">=14.0.1,<23.0.0"
 avro = "^1.12.1"
 numpy = "^2.0.0"
 
