### 2026-02-12 165abe3
Ensure proper handling of quantification errors; adjust tests accordingly
