### 2026-02-17 b5674d8
Resolve new issue with automatic path creation: PPG -> ppg
