+++ tests/test_paradigma.py
@@ -7,10 +7,10 @@ def test_transform_time_array(shared_datadir):
     It checks the output dimensions and the type of the output.
     """
     metadata_dict = tsdf.load_metadata_from_path(
-        shared_datadir / "1.prepared_data/ppg/PPG_meta.json"
+        shared_datadir / "1.prepared_data/ppg/ppg_meta.json"
     )
 
     # Retrieve the metadata object we want, using the name of the binary as key
-    metadata_values = metadata_dict["PPG_time.bin"]
+    metadata_values = metadata_dict["ppg_time.bin"]
     data = tsdf.load_ndarray_from_binary(metadata_values)
     assert data.shape == (64775,)
