### 2026-02-19 71acf96
Add LFS installation to workflow for modifying test data
