+++ .github/workflows/build-and-test.yml
@@ -25,6 +25,14 @@ jobs:
         with:
           fetch-depth: 0
 
+      - name: Install Git LFS
+        run: |
+          sudo apt-get install git-lfs
+          git lfs install
+
+      - name: Pull LFS files
+        run: git lfs pull
+
       - name: Set up Python ${{ matrix.python-version }}
         uses: actions/setup-python@v5
         with:
