### 2026-02-19 7e53b9a
Add LFS installation to workflow for modifying test data
