+++ .github/workflows/build-and-test.yml
@@ -25,13 +25,10 @@ jobs:
         with:
           fetch-depth: 0
 
-      - name: Install Git LFS
-        run: |
-          sudo apt-get install git-lfs
-          git lfs install
-
-      - name: Pull LFS files
-        run: git lfs pull
+      - name: Git LFS Pull
+        uses: actions/checkout@v4
+        with:
+          lfs: true
 
       - name: Set up Python ${{ matrix.python-version }}
         uses: actions/setup-python@v5
