### 2026-02-19 f6a8c5b
Add LFS installation to workflow for modifying test data
