### 2026-02-24 c2a71a9
Change setting sampling frequency to automatic deduction in resampling data
