+++ src/paradigma/preprocessing.py
@@ -1,3 +1,4 @@
+import logging
 from datetime import datetime
 
 import numpy as np
@@ -8,6 +9,8 @@ from scipy.interpolate import interp1d
 from paradigma.config import IMUConfig, PPGConfig
 from paradigma.segmenting import create_segments, discard_segments
 from paradigma.util import invert_watch_side
+
+logger = logging.getLogger(__name__)
 
 
 def resample_data(
@@ -21,7 +24,6 @@ def resample_data(
     auto_segment: bool = False,
     max_segment_gap_s: float | None = None,
     min_segment_length_s: float | None = None,
-    verbose: int = 2,
 ) -> pd.DataFrame:
     """
     Unified resampling function with optional auto-segmentation for non-contiguous data.
@@ -59,10 +61,6 @@ def resample_data(
     min_segment_length_s : float, optional
         Minimum segment length (seconds) to keep. Used when auto_segment=True.
         Defaults to IMUConfig.min_segment_length_s (1.5s).
-    verbose : int, default 1
-        Logging verbose: 0=errors only, 1=basic info, 2+=detailed info.
-        Note: This function still uses verbose for backward compatibility
-        with existing code that calls it directly.
 
     Returns
     -------
@@ -110,7 +108,8 @@ def resample_data(
 
     # Validate resampling frequency
     if resampling_frequency is None:
-        raise ValueError("resampling_frequency must be provided")
+        resampling_frequency = 100
+        logger.warning("resampling_frequency automatically set to 100 Hz")
 
     resampling_frequency = float(resampling_frequency)
 
@@ -124,8 +123,7 @@ def resample_data(
         ]
         if not values_column_names:
             raise ValueError("No numeric columns found for resampling")
-        if verbose >= 2:
-            print(f"Auto-detected {len(values_column_names)} columns for resampling")
+        logger.debug(f"Auto-detected {len(values_column_names)} columns for resampling")
 
     # Auto-detect or use provided sampling frequency
     time_abs_array = np.array(df[time_column])
@@ -133,8 +131,7 @@ def resample_data(
         time_diff = df[time_column].diff().dropna()
         current_dt = time_diff.median()
         sampling_frequency = 1.0 / current_dt
-        if verbose >= 2:
-            print(f"Auto-detected sampling frequency: {sampling_frequency:.2f} Hz")
+        logger.debug(f"Auto-detected sampling frequency: {sampling_frequency:.2f} Hz")
     else:
         sampling_frequency = float(sampling_frequency)
 
@@ -167,13 +164,13 @@ def resample_data(
             )
         elif auto_segment:
             # Split into segments
-            if verbose >= 1:
-                print("Non-contiguous data detected. Auto-segmenting...")
+            logger.info("Non-contiguous data detected. Auto-segmenting...")
 
             # Create segments based on gaps
             segment_array = create_segments(
                 time_array=time_abs_array,
                 max_segment_gap_s=max_segment_gap_s,
+                fs=int(sampling_frequency),
             )
             df["data_segment_nr"] = segment_array
 
@@ -187,15 +184,14 @@ def resample_data(
             )
 
             n_segments = df["data_segment_nr"].nunique()
-            if verbose >= 1:
-                segment_durations = []
-                for seg_nr in df["data_segment_nr"].unique():
-                    seg_df = df[df["data_segment_nr"] == seg_nr]
-                    duration = (
-                        seg_df[time_column].iloc[-1] - seg_df[time_column].iloc[0]
-                    )
-                    segment_durations.append(f"{duration:.1f}s")
-                print(f"Created {n_segments} segments: {', '.join(segment_durations)}")
[…]
