+++ src/paradigma/pipelines/gait_pipeline.py
@@ -282,8 +282,8 @@ def extract_arm_activity_features(
         df=df,
         segment_nr_colname=DataColumns.GAIT_SEGMENT_NR,
         min_segment_length_s=config.min_segment_length_s,
-        fs=config.sampling_frequency,
         format="timestamps",
+        fs=config.sampling_frequency,
     )
 
     # Create windows of fixed length and step size from the time series per segment
@@ -402,7 +402,7 @@ def filter_gait(
 
 def quantify_arm_swing(
     df: pd.DataFrame,
-    fs: int,
+    fs: int | None = None,
     filtered: bool = False,
     max_segment_gap_s: float = 1.5,
     min_segment_length_s: float = 1.5,
@@ -442,6 +442,9 @@ def quantify_arm_swing(
     df["unfiltered_segment_nr"] = create_segments(
         time_array=df[DataColumns.TIME], max_segment_gap_s=max_segment_gap_s
     )
+
+    # fs is deprecated
+    fs = 1 / df[DataColumns.TIME].diff().median()
 
     # Remove segments that do not meet predetermined criteria
     df = discard_segments(
@@ -783,6 +786,7 @@ def extract_temporal_domain_features(
 def extract_spectral_domain_features(
     windowed_data: np.ndarray,
     config,
+    fs,
     sensor: str,
 ) -> pd.DataFrame:
     """
@@ -799,8 +803,8 @@ def extract_spectral_domain_features(
         A 2D numpy array where each row corresponds to a window of sensor data.
 
     config : object
-        Configuration object containing settings such as sampling frequency,
-        window type, frequency bands, and MFCC parameters.
+        Configuration object containing settings such as window type,
+        frequency bands, and MFCC parameters.
 
     sensor : str
         The name of the sensor (e.g., 'accelerometer', 'gyroscope').
@@ -816,9 +820,7 @@ def extract_spectral_domain_features(
     feature_dict = {}
 
     # Compute periodogram (power spectral density)
-    freqs, psd = periodogram(
-        x=windowed_data, fs=config.sampling_frequency, window=config.window_type, axis=1
-    )
+    freqs, psd = periodogram(x=windowed_data, fs=fs, window=config.window_type, axis=1)
 
     # Compute power in specified frequency bands
     for band_name, band_freqs in config.d_frequency_bandwidths.items():
@@ -996,7 +998,6 @@ def run_gait_pipeline(
         config=imu_config,
         sensor="both",
         watch_side=watch_side,
-        verbose=1 if logging_level <= logging.INFO else 0,
     )
 
     if "preprocessing" in store_intermediate:
@@ -1040,7 +1041,6 @@ def run_gait_pipeline(
         df_predictions=df_gait,
         pred_proba_colname=DataColumns.PRED_GAIT_PROBA,
         window_length_s=gait_config.window_length_s,
-        fs=gait_config.sampling_frequency,
     )
 
     # Add binary prediction column
@@ -1112,7 +1112,6 @@ def run_gait_pipeline(
         df_predictions=df_arm_activity,
         pred_proba_colname=DataColumns.PRED_NO_OTHER_ARM_ACTIVITY_PROBA,
         window_length_s=arm_activity_config.window_length_s,
-        fs=arm_activity_config.sampling_frequency,
     )
 
     # Add binary prediction column
@@ -1136,7 +1135,6 @@ def run_gait_pipeline(
         quantified_arm_swing_unfiltered, gait_segment_meta_unfiltered = (
             quantify_arm_swing(
                 df=df_filtered,
-                fs=arm_activity_config.sampling_frequency,
                 filtered=False,  # Quantify all gait
                 max_segment_gap_s=arm_activity_config.max_segment_gap_s,
[…]
