### 2026-02-24 02c9cd1
Change setting sampling frequency to automatic deduction in resampling data
