### 2026-02-24 b589bb1
Solve pytest conflicts
