### 2026-02-24 b5917bb
Resolve errors in notebooks and pytest
