### 2026-02-24 01e5c8a
Add unfiltered gait quantification (#264)
* Add unfiltered gait quantification to store both clean and all gait data

BREAKING CHANGE: run_gait_pipeline() now returns nested dicts instead of flat tuple

Previously, run_gait_pipeline() returned:
  (quantified_df: DataFrame, metadata: dict)

Now returns:
  ({'filtered': df, 'unfiltered': df}, {'filtered': meta, 'unfiltered': meta})

This change allows analysis of both:
- Filtered (clean gait only): Arm swing during walking without other activities
- Unfiltered (all gait): All arm swing during any gait, regardless of quality

Changes:
- gait_pipe
[…]
