+++ tests/test_gait_analysis.py
@@ -5,10 +5,10 @@ import numpy as np
 import pandas as pd
 import pytest
 
+import paradigma.pipelines.gait_pipeline as gait_pipeline
 from paradigma.config import GaitConfig, IMUConfig
 from paradigma.constants import DataColumns
 from paradigma.orchestrator import run_paradigma
-from paradigma.pipelines.gait_pipeline import run_gait_pipeline
 
 try:
     from test_notebooks import compare_data
@@ -91,7 +91,7 @@ def test_gait_pipeline_basic():
     df_test = create_test_gait_data()
 
     with tempfile.TemporaryDirectory() as temp_dir:
-        result = run_gait_pipeline(
+        result = gait_pipeline.run_gait_pipeline(
             df_prepared=df_test,
             watch_side="right",
             output_dir=temp_dir,
@@ -102,8 +102,29 @@ def test_gait_pipeline_basic():
 
         assert isinstance(result, tuple)
         assert len(result) == 2
-        assert isinstance(result[0], pd.DataFrame)  # Gait results DataFrame
-        assert isinstance(result[1], dict)  # Gait metadata dictionary
+
+        # First element should be a dict with 'filtered' and
+        # 'unfiltered' quantifications
+        assert isinstance(result[0], dict)
+        assert "filtered" in result[0]
+        assert "unfiltered" in result[0]
+        assert isinstance(result[0]["filtered"], pd.DataFrame)
+        assert isinstance(result[0]["unfiltered"], pd.DataFrame)
+
+        expected_columns = {
+            DataColumns.GAIT_SEGMENT_NR,
+            DataColumns.RANGE_OF_MOTION,
+            DataColumns.PEAK_VELOCITY,
+        }
+        assert expected_columns.issubset(result[0]["filtered"].columns)
+        assert expected_columns.issubset(result[0]["unfiltered"].columns)
+
+        # Second element should be a dict with 'filtered' and 'unfiltered' metadata
+        assert isinstance(result[1], dict)
+        assert "filtered" in result[1]
+        assert "unfiltered" in result[1]
+        assert isinstance(result[1]["filtered"], dict)
+        assert isinstance(result[1]["unfiltered"], dict)
 
 
 def test_gait_pipeline_integration():
@@ -125,6 +146,108 @@ def test_gait_pipeline_integration():
         assert "quantifications" in results
         assert "metadata" in results
         assert "aggregations" in results
+
+        gait_quantifications = results["quantifications"]["gait"]
+        assert "filtered" in gait_quantifications
+        assert "unfiltered" in gait_quantifications
+        assert isinstance(gait_quantifications["filtered"], pd.DataFrame)
+        assert isinstance(gait_quantifications["unfiltered"], pd.DataFrame)
+
+        expected_columns = {
+            DataColumns.GAIT_SEGMENT_NR,
+            DataColumns.RANGE_OF_MOTION,
+            DataColumns.PEAK_VELOCITY,
+        }
+        assert expected_columns.issubset(gait_quantifications["filtered"].columns)
+        assert expected_columns.issubset(gait_quantifications["unfiltered"].columns)
+
+
+def test_multi_file_unfiltered_only_offsets(monkeypatch):
+    """Ensure unfiltered-only gait results keep unique segment ids across files."""
+
+    class DummyClassifierPackage:
+        def __init__(self, threshold: float = 0.5) -> None:
+            self.threshold = threshold
+
+    def fake_load(_path):
+        return DummyClassifierPackage()
+
+    def fake_detect_gait(df, _clf_package, parallel=False):
+        return np.ones(len(df))
+
[…]
