+++ src/paradigma/orchestrator.py
@@ -58,6 +58,16 @@ DETAILED_INFO = 15
 logging.addLevelName(DETAILED_INFO, "DETAILED")
 
 
+def _empty_gait_quantification_df() -> pd.DataFrame:
+    return pd.DataFrame(
+        columns=[
+            DataColumns.GAIT_SEGMENT_NR,
+            DataColumns.RANGE_OF_MOTION,
+            DataColumns.PEAK_VELOCITY,
+        ]
+    )
+
+
 def run_paradigma(
     *,
     data_path: str | Path | None = None,
@@ -75,7 +85,7 @@ def run_paradigma(
     device_orientation: list[str] | None = ["x", "y", "z"],
     file_pattern: str | list[str] | None = None,
     aggregates: list[str] | None = None,
-    segment_length_bins: list[str] | None = None,
+    segment_length_bins: list[tuple[float, float] | list[float] | str] | None = None,
     split_by_gaps: bool = False,
     max_gap_seconds: float | None = None,
     min_segment_seconds: float | None = None,
@@ -152,9 +162,10 @@ def run_paradigma(
         File pattern(s) to match when loading data (e.g., 'parquet', '*.csv').
     aggregates : list of str, optional
         Aggregation methods for quantification.
-    segment_length_bins : list of str, optional
+    segment_length_bins : list of str or list of tuple/list, optional
         Duration bins for gait segment aggregation (gait pipeline only).
-        Example: ['(0, 10)', '(10, 20)'] for segments 0-10s and 10-20s.
+        Accepts either string bins like ['(0, 10)', '(10, 20)'] or 2-element
+        tuples/lists like [(0, 10), (10, 20)] for segments 0-10s and 10-20s.
     split_by_gaps : bool, default False
         If True, automatically split non-contiguous data into segments
         during preparation.
@@ -194,9 +205,15 @@ def run_paradigma(
     -------
     dict
         Complete analysis results with nested structure for multiple pipelines:
-        - 'quantifications': dict with pipeline names as keys and DataFrames as values
-        - 'aggregations': dict with pipeline names as keys and result dicts as values
-        - 'metadata': dict with pipeline names as keys and metadata dicts as values
+                - 'quantifications': dict with pipeline names as keys and DataFrames
+                    as values. For gait, values are nested dicts with 'filtered' and
+                    'unfiltered' DataFrames.
+                - 'aggregations': dict with pipeline names as keys and result dicts
+                    as values. For gait, values are nested dicts with 'filtered' and
+                    'unfiltered' aggregation dictionaries.
+                - 'metadata': dict with pipeline names as keys and metadata dicts
+                    as values. For gait, values are nested dicts with 'filtered' and
+                    'unfiltered' metadata dictionaries.
         - 'errors': list of dicts tracking any errors that occurred during processing.
           Each error dict contains 'stage', 'error', and optionally 'file' and
           'pipeline'.
@@ -283,17 +300,28 @@ def run_paradigma(
 
     # Initialize results storage for each pipeline
     all_results = {
-        "quantifications": {p: [] for p in pipelines},
-        "aggregations": {p: {} for p in pipelines},
-        "metadata": {p: {} for p in pipelines},
+        "quantifications": {
+            p: [] if p != "gait" else {"filtered": [], "unfiltered": []}
+            for p in pipelines
+        },
+        "aggregations": {
+            p: {} if p != "gait" else {"filtered": {}, "unfiltered": {}}
+            for p in pipelines
+        },
+        "metadata": {
+            p: {} if p != "gait" else {"filtered": {}, "unfiltered": {}}
+            for p in pipelines
+        },
         "errors": [],
     }
 
     # Steps 2-3: Process each file individually
     active_logger.info(f"Steps 2-3: Processing {num_files} files individually")
 
[…]
