+++ docs/index.md
@@ -11,7 +11,7 @@ guides/installation.md
 guides/data_input.md
 guides/sensor_requirements.md
 guides/supported_devices.md
-guides/validation.md
+guides/input_formats.md
 guides/config.md
 guides/coordinate_system.md
 ```
@@ -23,6 +23,7 @@ guides/coordinate_system.md
 
 contributing.md
 conduct.md
+changelog.md
 license.md
 ```
 
