### 2026-02-24 #9502 @Erikpostt
Thank you @Asanto32.

@ntolley @Asanto32 @richrobe if you were planning on testing using data of the `tests` subdirectory, please pull the `main` branch. The current release had a minor bug when trying to use this data.
