### 2026-02-26 #9502 @Asanto32
Hi @Erikpostt ,

Thanks for all the changes. I think it looks and runs much better. I am just awaiting that final version of the pdf (I see above there was an error generating it?), but otherwise I think I am happy with these improvements.

One thing about this comment "Not yet supported: GENEActiv (.bin), Actigraph (.gt3x), and other proprietary formats
We're willing to add support for additional formats. If you can share sample data (even synthetic/anonymized), we'd be happy to implement support for GENEActiv and Actigraph."

I can point you to our reader [here](https://github.com/childmindresearch/actfast), very fast reader built in rust specifically for Actigraph and Geneactiv files as we were tired of dealing with the other slow readers. Very easy to transform the output from `actfast` (`Dict`) into a polars or pandas DataFrame. You can also get short sample data from our processing toolbox [`wristpy`](https://github.com/childmindresearch/wristpy/tree/main/tests/sample_data).
