### 2026-02-27 #9502 @Erikpostt
@Asanto32, thank you for reviewing the changes and providing your feedback once again. @Kevin-Mattheus-Moerman any chance you could look into the Editorialbot error when generating the pdf?

The sample data for GENEActiv and Actigraph look useful, I'll have to look into that in due time. We'll also be sure to test data loading with `actfast` and do some quick performance checks.
