### 2026-03-06 biomarkersParkinson/paradigma#250 @Erikpostt
Hi @ntolley,

Apologies, I figured it was about the `pulse_rate_analysis.ipynb` tutorial, but you're talking about the `tests` directory. These notebooks are in fact outdated and are no longer used for testing purposes (e.g., we use `test_pulse_rate_analysis.py` for testing the pulse rate pipeline). Illustrating this is the fact that it's still called _heart_rate_analysis_.

Per your recommendation, I could indeed change the prefixes, but preferably I would clean up the `tests` directory by removing redundant data folders (as you demonstrated) and discard the testing notebooks. Would you be okay if I clean this up?

For trying out the core functionalities, the most logical place is `docs/tutorials`, as is now clearly documented (hopefully) [here](https://biomarkersparkinson.github.io/paradigma/). Let me know if these tutorials tackle the questions you face.
