### 2026-03-17 #9502 @ntolley
Hi @Kevin-Mattheus-Moerman apologies for the delay, there was one last [issue](https://github.com/biomarkersParkinson/paradigma/issues/250) which I just confirmed that @Erikpostt has resolved

My review is now complete and I have confirmed everything is working and is well-documented, great work @Erikpostt! This package is a great contribution to the digital health community
