### 2026-03-20 864b3b7
Fix save_intermediate parameter: use 'classification' for all prediction outputs and store predictions instead of features
