### 2026-03-20 c317334
Also update for arm activity predictions
