### 2026-03-20 de5222a
Store window-level predictions of gait pipeline, consistent with other pipelines
