### 2026-03-20 276a422
Store window-level predictions of gait pipeline, consistent with other pipelines
