### 2026-03-20 5421204
feat: Add adaptive frequency bounds for variable sampling rates
- Add sampling_frequency property with setter to IMUConfig
- Implement _update_frequency_dependent_params() virtual method
- Override in GaitConfig: auto-clamp spectrum_high_frequency and mfcc_high_frequency to Nyquist
- Override in TremorConfig: auto-clamp fmax_peak_search and fmax_mfcc to Nyquist
- Add warnings when frequency bounds are clamped for safety
- Fixes issue where 50Hz/64Hz data caused 'out of bounds' errors in feature extraction

This allows users to set sampling_frequency at any point and freq-dependent params
adapt
[…]
