### 2026-03-20 f989ee4
fix: Defer _update_frequency_dependent_params() until after child class initialization
This fixes the initialization order bug where the update method was being called
in IMUConfig.__init__() before subclass attributes like mfcc_high_frequency were
set. Solution: defer calls until the end of GaitConfig and TremorConfig __init__.

The property setter still works for post-initialization frequency changes, and
the virtual method pattern correctly updates all bounds when sampling_frequency
changes at any point in the config lifecycle.

Verified with test_adaptive_frequencies_demo.py - all 3 tests 
[…]
