### 2026-03-20 dae1451
chore: Remove development test files and implementation notes
- Remove test_adaptive_frequencies_demo.py (development validation only)
- Remove implementation notes from IMUConfig (internal architectural detail)
- Keep core adaptive frequency bounds implementation clean for production

Commit validates that all 3 frequency test cases (50/64/100 Hz) pass successfully.
