### 2026-03-20 2cae6b7
fix: Only clamp MFCC bounds when they exceed Nyquist, preserve 25 Hz band
