### 2026-03-20 f6c1700
fix: Clamp MFCC bounds to exactly Nyquist, not 95% of it
