### 2026-03-21 82be156
Enforce resampling below 50 Hz
