### 2026-03-23 44b88b4
Fix save_intermediate parameter: use 'classification' for all predict… (#269)
* Fix save_intermediate parameter: use 'classification' for all prediction outputs and store predictions instead of features
