### 2026-03-23 #9502 @Asanto32
> [@ntolley](https://github.com/ntolley) thank you for your time and effort! [@Asanto32](https://github.com/Asanto32) did you have a chance to read the updated manuscript?

Yes! All looks good to me. Thanks for all the greatk work!
