### 2026-03-24 cd3ccd7
Bump version to 1.1.1 and update changelog
