### 2026-03-24 5696bda
Restore tests/ folder from main
