+++ tests/test_gait_analysis.py
@@ -1,0 +1,418 @@
+import tempfile
+from pathlib import Path
+
+import numpy as np
+import pandas as pd
+import pytest
+
+import paradigma.pipelines.gait_pipeline as gait_pipeline
+from paradigma.config import GaitConfig, IMUConfig
+from paradigma.constants import DataColumns
+from paradigma.orchestrator import run_paradigma
+
+try:
+    from conftest import compare_data
+
+    from paradigma.testing import (
+        extract_gait_features_io,
+    )
+
+    HAS_TESTING_UTILS = True
+except ImportError:
+    HAS_TESTING_UTILS = False
+
+
+def create_test_gait_data(duration_minutes=2):
+    """Create simple test data for gait analysis."""
+    duration_seconds = duration_minutes * 60
+    n_samples = int(duration_seconds * 100.0)  # 100 Hz
+    time = np.linspace(0, duration_seconds, n_samples)
+
+    # Simple IMU data
+    acc_x = 0.2 * np.sin(2 * np.pi * 1.5 * time) + 0.1 * np.random.random(n_samples)
+    acc_y = 0.2 * np.cos(2 * np.pi * 1.5 * time) + 0.1 * np.random.random(n_samples)
+    acc_z = 9.8 + 0.2 * np.random.random(n_samples)
+
+    gyro_x = 0.1 * np.sin(2 * np.pi * 2 * time) + 0.05 * np.random.random(n_samples)
+    gyro_y = 0.1 * np.cos(2 * np.pi * 2 * time) + 0.05 * np.random.random(n_samples)
+    gyro_z = 0.05 * np.random.random(n_samples)
+
+    return pd.DataFrame(
+        {
+            DataColumns.TIME: time,
+            DataColumns.ACCELEROMETER_X: acc_x,
+            DataColumns.ACCELEROMETER_Y: acc_y,
+            DataColumns.ACCELEROMETER_Z: acc_z,
+            DataColumns.GYROSCOPE_X: gyro_x,
+            DataColumns.GYROSCOPE_Y: gyro_y,
+            DataColumns.GYROSCOPE_Z: gyro_z,
+        }
+    )
+
+
+def create_test_gait_data_with_gaps(n_segments=3):
+    """Create test data with gaps (multiple data segments)."""
+    segments = []
+    for i in range(n_segments):
+        # Create 30 second segments with large gaps between them
+        duration_seconds = 30
+        n_samples = int(duration_seconds * 100.0)  # 100 Hz
+        time_start = i * 100  # 100 second gaps between segments
+        time = np.linspace(time_start, time_start + duration_seconds, n_samples)
+
+        # Simple IMU data
+        acc_x = 0.2 * np.sin(2 * np.pi * 1.5 * time) + 0.1 * np.random.random(n_samples)
+        acc_y = 0.2 * np.cos(2 * np.pi * 1.5 * time) + 0.1 * np.random.random(n_samples)
+        acc_z = 9.8 + 0.2 * np.random.random(n_samples)
+
+        gyro_x = 0.1 * np.sin(2 * np.pi * 2 * time) + 0.05 * np.random.random(n_samples)
+        gyro_y = 0.1 * np.cos(2 * np.pi * 2 * time) + 0.05 * np.random.random(n_samples)
+        gyro_z = 0.05 * np.random.random(n_samples)
+
+        segment_df = pd.DataFrame(
+            {
+                DataColumns.TIME: time,
+                DataColumns.ACCELEROMETER_X: acc_x,
+                DataColumns.ACCELEROMETER_Y: acc_y,
+                DataColumns.ACCELEROMETER_Z: acc_z,
+                DataColumns.GYROSCOPE_X: gyro_x,
[…]
