### 2026-03-24 cc72daa
Sync/main to release (#270)
* Bump version to 1.1.1 and update changelog
