### 2026-03-24 e7c891e
Update changelog to v1.1.1
