### 2026-03-24 38f49df
Release v1.1.1 (#272)
* Add unfiltered gait quantification to store both clean and all gait data

* Fix orchestrator to pass required segment_meta and segment_cats to aggregate_arm_swing_params

* Clean up test infrastructure: remove unused test files, notebooks, and consolidate helpers to conftest.py

* Fix save_intermediate parameter: use 'classification' for all prediction outputs and store predictions instead of features

* Bump version to 1.1.1 and update changelog

* Remove tests/notebooks/ folder
