### 2026-03-24 fc39a8f
Update paper.md
Attempt to fix paper error. Also changed UK to United Kingdom
