### 2026-03-24 #9502 @Kevin-Mattheus-Moerman
@Erikpostt I tried to re-build the paper but ran into an error. Could not see yet what the error is exactly. Perhaps this PR fixes it: https://github.com/biomarkersParkinson/paradigma/pull/275
