### 2026-03-25 #9502 @Erikpostt
> 👉📄 [Download article proof](https://raw.githubusercontent.com/openjournals/joss-papers/joss.09502/joss.09502/10.21105.joss.09502.pdf) 📄 [View article proof on GitHub](https://github.com/openjournals/joss-papers/blob/joss.09502/joss.09502/10.21105.joss.09502.pdf) 📄 👈

The latest article is here, so it did generate before (no more changes were made afterwards). I'll look into the PR.
