### 2026-03-25 708e6cc
Fix title formatting in paper.md (double quotes to block "parkinson's" from closing single quotes title
