### 2026-03-25 41c7fb7
Fix title formatting in paper.md (block scalar to resolve conflict with "parkinson's" single quote usage) (#276)
* Remove quotes to examine where this caused the issue
