### 2026-03-25 #9502 @Erikpostt
@Kevin-Mattheus-Moerman double quotes in title were required to enable "parkinson's" in title (note single quote in that word). I've changed it to a block scalar, local tests seemed to work.
