### 2026-03-25 #9502 @Erikpostt
@Kevin-Mattheus-Moerman would you allow us to add 2 more references of our recent pre-prints to the final paper? These would be added to L102-103.
