### 2026-03-25 #9502 @Kevin-Mattheus-Moerman
> @Kevin-Mattheus-Moerman would you allow us to add 2 more references of our recent pre-prints to the final paper? These would be added to L102-103.

Sure, that's fine. Let me know when you are done and I'll review your paper.
