+++ src/paradigma/orchestrator.py
@@ -72,7 +72,9 @@ def run_paradigma(
     *,
     data_path: str | Path | None = None,
     dfs: pd.DataFrame | list[pd.DataFrame] | dict[str, pd.DataFrame] | None = None,
+    run_steps: list[str] | str = "all",
     save_intermediate: list[str] = [],
+    return_intermediate: list[str] = [],
     output_dir: str | Path = "./output",
     skip_preparation: bool = False,
     pipelines: list[str] | str | None = None,
@@ -125,15 +127,31 @@ def run_paradigma(
         Note: The 'file_key' column is only added to quantification results when
         len(dfs) > 1, allowing cleaner output for single-file processing.
         See input_formats guide for details.
+    run_steps : list of str or str, default 'all'
+        Which steps of the pipeline(s) to run. When constructing a list, the following
+        steps are valid: 'preprocessing', 'classification', 'quantification',
+        'aggregation'. Steps must be in this order; skipping intermediate steps is not
+        allowed (e.g., cannot do ['preprocessing', 'quantification'] because
+        'classification' is required). Use this parameter when running ParaDigMa in
+        parallel across multiple directories or data chunks, allowing you to split
+        the pipeline into separate runs and aggregate results later by only running
+        aggregation on combined quantifications.
     save_intermediate : list of str, default []
-        Which intermediate results to store. Valid values:
+        Which intermediate results to store to disk. Valid values:
         - 'preparation': Save prepared data
         - 'preprocessing': Save preprocessed signals
         - 'classification': Save classification/prediction results (e.g., gait
         predictions, tremor predictions, PPG signal quality predictions)
         - 'quantification': Save quantified measures
         - 'aggregation': Save aggregated results
-        If empty, no files are saved (results are only returned).
+        Must be subset of run_steps. If empty, no files are saved.
+    return_intermediate : list of str, default []
+        Which intermediate results to include in return dictionary. Valid values:
+        same as save_intermediate. By default, only quantifications/aggregations
+        are returned to minimize memory usage. Set to include preprocessing/
+        classification results for debugging or inspection. Must be subset of
+        run_steps. Results can be saved to disk without being returned (and vice
+        versa) by configuring these parameters independently.
     output_dir : str or Path, default './output'
         Output directory for all results. Files are only saved if
         save_intermediate is not empty.
@@ -237,6 +255,62 @@ def run_paradigma(
             f"Supported pipelines: 'gait', 'tremor', 'pulse_rate'"
         )
 
+    # Parse and validate run_steps
+    valid_steps = ["preprocessing", "classification", "quantification", "aggregation"]
+    if isinstance(run_steps, str):
+        if run_steps == "all":
+            run_steps = valid_steps.copy()
+        else:
+            raise ValueError(
+                f"Invalid run_steps value: '{run_steps}'. Must be 'all' or a list."
+            )
+
+    # Validate all steps are valid
+    invalid_steps = [s for s in run_steps if s not in valid_steps]
+    if invalid_steps:
+        raise ValueError(
+            f"Invalid steps in run_steps: {invalid_steps}. "
+            f"Valid steps are: {valid_steps}"
+        )
+
+    # Check that steps are in correct order and no intermediate steps are skipped
+    step_indices = {step: i for i, step in enumerate(valid_steps)}
+    sorted_steps = sorted(run_steps, key=lambda s: step_indices[s])
+    if list(run_steps) != sorted_steps:
+        raise ValueError(
+            f"Steps must be in order: {sorted_steps} (received: {run_steps})"
+        )
+
+    # Check for skipped intermediate steps
+    run_steps_set = set(run_steps)
+    for i, step in enumerate(valid_steps):
+        if step in run_steps_set:
[…]
