+++ tests/test_tremor_analysis.py
@@ -53,8 +53,13 @@ def test_tremor_pipeline_basic():
             imu_config=IMUConfig(),
         )
 
-        assert isinstance(result, pd.DataFrame)
-        assert DataColumns.TIME in result.columns
+        # New dict return format
+        assert isinstance(result, dict)
+        assert "quantification" in result
+        assert "_steps_executed" in result
+        quantification = result.get("quantification")
+        assert isinstance(quantification, pd.DataFrame)
+        assert DataColumns.TIME in quantification.columns
 
 
 def test_tremor_pipeline_integration():
