+++ src/paradigma/pipelines/gait_pipeline.py
@@ -874,7 +874,9 @@ def run_gait_pipeline(
     segment_number_offset_unfiltered: int = 0,
     logging_level: int = logging.INFO,
     custom_logger: logging.Logger | None = None,
-) -> tuple[dict[str, pd.DataFrame], dict[str, dict]]:
+    run_steps: list[str] | None = None,
+    return_intermediate: list[str] | None = None,
+) -> dict:
     """
     Run the complete gait analysis pipeline on prepared data (steps 1-6).
 
@@ -927,22 +929,32 @@ def run_gait_pipeline(
         etc.)
     custom_logger : logging.Logger, optional
         Custom logger instance. If provided, logging_level is ignored.
+    run_steps : list[str], optional
+        Which steps to execute: 'preprocessing', 'classification', 'quantification'.
+        Must be a contiguous sequence starting from 'preprocessing'.
+        If None, defaults to running all steps.
+    return_intermediate : list[str], optional
+        Which intermediate results to include in the return dict.
+        Must be a subset of run_steps.
+        If None, defaults to only returning quantification.
 
     Returns
     -------
-    tuple[dict[str, pd.DataFrame], dict[str, dict]]
-        A tuple containing two dictionaries:
-        - First dict contains quantified arm swing parameters with keys:
-            - 'filtered': DataFrame with arm swings from clean gait only
-            - 'unfiltered': DataFrame with arm swings from all gait
-          Each DataFrame has columns:
+    dict
+        A dictionary containing pipeline results with the following possible keys:
+        - 'preprocessing': DataFrame with preprocessed IMU data
+        (if 'preprocessing' in return_intermediate)
+        - 'classification': Dict with classification results
+        (if 'classification' in return_intermediate)
+        - 'quantification': Dict with keys 'filtered' and 'unfiltered', each containing
+        a DataFrame
+          with arm swing parameters. Each DataFrame has columns:
             - gait_segment_nr: Gait segment number within this data segment
             - Various arm swing metrics (range of motion, peak angular velocity, etc.)
             - Additional metadata columns
-        - Second dict contains gait segment metadata with keys:
-            - 'filtered': Metadata for filtered quantification
-            - 'unfiltered': Metadata for unfiltered quantification
-          Each metadata dict contains information about each detected gait segment
+        - 'metadata': Dict with keys 'filtered' and 'unfiltered' containing metadata
+        - '_steps_executed': List of steps that were actually executed
+        - '_error': Error message if an error occurred (None if successful)
 
     Notes
     -----
@@ -961,6 +973,32 @@ def run_gait_pipeline(
 
     if store_intermediate is None:
         store_intermediate = []
+
+    # Initialize run_steps and return_intermediate
+    if run_steps is None:
+        run_steps = ["preprocessing", "classification", "quantification"]
+    if return_intermediate is None:
+        return_intermediate = ["quantification"]
+
+    # Validate run_steps and return_intermediate
+    valid_steps = ["preprocessing", "classification", "quantification"]
+    if not all(step in valid_steps for step in run_steps):
+        invalid = [s for s in run_steps if s not in valid_steps]
+        raise ValueError(f"Invalid steps in run_steps: {invalid}. Valid: {valid_steps}")
+
+    if not all(step in run_steps for step in return_intermediate):
+        invalid = [s for s in return_intermediate if s not in run_steps]
+        raise ValueError(f"return_intermediate steps {invalid} are not in run_steps")
+
+    steps_executed = []
+    result_dict = {
+        "preprocessing": None,
[…]
