+++ paper/paper.bib
@@ -212,6 +212,20 @@
   langid = {english},
 }
 
+@misc{postLongitudinalProgressionDigital2026,
+  title = {Longitudinal Progression of Digital Arm Swing Measures during Free-Living Gait in Early {{Parkinson}}'s Disease},
+  author = {Post, Erik and van Laarhoven, Twan and Timmermans, Nienke A. and Raykov, Yordan P. and Little, Max A. and Nonnekes, Jorik and Ho, King Chung and Heskes, Tom M. and Bloem, Bastiaan R. and Evers, Luc J. W.},
+  year = 2026,
+  month = jan,
+  pages = {2026.01.06.26343500},
+  publisher = {medRxiv},
+  issn = {3067-2007},
+  doi = {10.64898/2026.01.06.26343500},
+  archiveprefix = {medRxiv},
+  copyright = {\copyright{} 2026, Posted by openRxiv. This pre-print is available under a Creative Commons License (Attribution-NonCommercial 4.0 International), CC BY-NC 4.0, as described at http://creativecommons.org/licenses/by-nc/4.0/},
+  langid = {english},
+}
+
 @techreport{radbouduniversitymedicalcenterSlowSPEEDNLSlowingParkinsons2025,
   type = {Clinical Trial Registration},
   title = {Slow-{{SPEED-NL}}: {{Slowing Parkinson}}'s {{Early Through Exercise Dosage-Netherlands}}},
@@ -284,6 +298,21 @@
   doi = {10.1038/s41531-025-01056-2},
   urldate = {2025-08-11},
   langid = {english}
+}
+
+@misc{timmermansProgressionDailylifeTremor2025a,
+  title = {Progression of Daily-Life Tremor Measures in Early {{Parkinson}} Disease: A Longitudinal Continuous Monitoring Study},
+  shorttitle = {Progression of Daily-Life Tremor Measures in Early {{Parkinson}} Disease},
+  author = {Timmermans, Nienke A. and Bucur, Ioan Gabriel and Soriano, Diogo C. and Post, Erik and Cagnan, Hayriye and Shin, Sooyoon and Little, Max A. and Raykov, Yordan P. and Bloem, Bastiaan R. and Helmich, Rick C. and Evers, Luc J. W.},
+  year = 2025,
+  month = dec,
+  pages = {2025.12.23.25342892},
+  publisher = {medRxiv},
+  issn = {3067-2007},
+  doi = {10.64898/2025.12.23.25342892},
+  archiveprefix = {medRxiv},
+  copyright = {\copyright{} 2025, Posted by openRxiv. This pre-print is available under a Creative Commons License (Attribution-NonCommercial 4.0 International), CC BY-NC 4.0, as described at http://creativecommons.org/licenses/by-nc/4.0/},
+  langid = {english},
 }
 
 @techreport{ucbbiopharmasrlDoubleBlindPlaceboControlledRandomized2024,
