+++ paper/paper.md
@@ -71,13 +71,13 @@ bibliography: paper.bib
 
 # Summary
 
-PARkinson's disease DIGital bioMArkers (ParaDigMa) is an open-source Python package designed to extract digital biomarkers of Parkinson's disease (PD) from wrist-worn sensor data collected during passive monitoring in daily life [@postParaDigMaToolboxDeriving2025a]. It processes accelerometer, gyroscope, and photoplethysmography (PPG) signals using modular pipelines which address common challenges in real-world wearable sensor data analysis. The package currently targets three clinically relevant symptom domains: (1) tremor, (2) arm swing during gait, and (3) pulse rate (to derive digital biomarkers related to autonomic dysfunction). ParaDigMa enables researchers to derive interpretable and clinically meaningful digital biomarkers to monitor PD in daily life.
+PARkinson's disease DIGital bioMArkers (ParaDigMa) is an open-source Python package designed to extract digital biomarkers of Parkinson's disease (PD) from wrist sensor data collected during passive monitoring in daily life [@postParaDigMaToolboxDeriving2025a]. It processes accelerometer, gyroscope, and photoplethysmography (PPG) signals using modular pipelines which address common challenges in real-world wearable sensor data analysis. The package currently targets three clinically relevant symptom domains: (1) tremor, (2) arm swing during gait, and (3) pulse rate (to derive digital biomarkers related to autonomic dysfunction). ParaDigMa enables researchers to derive interpretable and clinically meaningful digital biomarkers to monitor PD in daily life.
 
 # Statement of need
 
 The use of wearable sensors to monitor PD in daily life has grown exponentially over the past two decades [@debSystematicSurveyResearch2022], with mounting evidence supporting their utility for diagnosis, monitoring disease progression, and evaluating treatment efficacy [@sunDigitalBiomarkersPrecision2024]. However, the extraction of digital biomarkers from raw sensor signals often relies on proprietary or non-replicable pipelines applied to individual cohorts. To support their adoption in clinical and research practice, and to enable meaningful comparisons across studies, reproducible and standardized data processing workflows are essential [@sunDigitalBiomarkersPrecision2024; @boyleActivityBehavioralRecognition2025].
 
-ParaDigMa addresses this need by providing pipelines for extracting digital biomarkers of PD from wrist-worn sensor data obtained from passive monitoring in daily life. Instead of requiring researchers to build end-to-end pipelines from scratch, ParaDigMa streamlines the transformation of raw wrist sensor data into clinically meaningful outcomes. In addition, its modular and standardized architecture enables users to replace or extend components, encouraging collaboration to expand and refine the toolbox for broader use in digital health research.
+ParaDigMa addresses this need by providing pipelines for extracting digital biomarkers of PD from wrist sensor data obtained from passive monitoring in daily life. Instead of requiring researchers to build end-to-end pipelines from scratch, ParaDigMa streamlines the transformation of raw wrist sensor data into clinically meaningful outcomes. In addition, its modular and standardized architecture enables users to replace or extend components, encouraging collaboration to expand and refine the toolbox for broader use in digital health research.
 
 # Comparison
 
@@ -89,7 +89,7 @@ Beyond PD-specific toolboxes , ParaDigMa also differs from general-purpose toolb
 
 # Architecture
 
[…]
