+++ paper/paper.bib
@@ -138,22 +138,6 @@
   langid = {english}
 }
 
-@article{marekParkinsonProgressionMarker2011a,
-  title = {The {{Parkinson Progression Marker Initiative}} ({{PPMI}})},
-  author = {Marek, Kenneth and Jennings, Danna and Lasch, Shirley and Siderowf, Andrew and Tanner, Caroline and Simuni, Tanya and Coffey, Chris and Kieburtz, Karl and Flagg, Emily and Chowdhury, Sohini and Poewe, Werner and Mollenhauer, Brit and Klinik, Paracelsus-Elena and Sherer, Todd and Frasier, Mark and Meunier, Claire and Rudolph, Alice and Casaceli, Cindy and Seibyl, John and Mendick, Susan and Schuff, Norbert and Zhang, Ying and Toga, Arthur and Crawford, Karen and Ansbach, Alison and De Blasio, Pasquale and Piovella, Michele and Trojanowski, John and Shaw, Les and Singleton, Andrew and Hawkins, Keith and Eberling, Jamie and Brooks, Deborah and Russell, David and Leary, Laura and Factor, Stewart and Sommerfeld, Barbara and Hogarth, Penelope and Pighetti, Emily and Williams, Karen and Standaert, David and Guthrie, Stephanie and Hauser, Robert and Delgado, Holly and Jankovic, Joseph and Hunter, Christine and Stern, Matthew and Tran, Baochan and Leverenz, Jim and Baca, Marne and Frank, Sam and Thomas, Cathi-Ann and Richard, Irene and Deeley, Cheryl and Rees, Linda and Sprenger, Fabienne and Lang, Elisabeth and Shill, Holly and Obradov, Sanja and Fernandez, Hubert and Winters, Adrienna and Berg, Daniela and Gauss, Katharina and Galasko, Douglas and Fontaine, Deborah and Mari, Zoltan and Gerstenhaber, Melissa and Brooks, David and Malloy, Sophie and Barone, Paolo and Longo, Katia and Comery, Tom and Ravina, Bernard and Grachev, Igor and Gallagher, Kim and Collins, Michelle and Widnell, Katherine L. and Ostrowizki, Suzanne and Fontoura, Paulo and Ho, Tony and Luthman, Johan and Brug, Marcel Van Der and Reith, Alastair D. and Taylor, Peggy},
-  year = {2011},
-  month = dec,
-  journal = {Progress in Neurobiology},
-  volume = {95},
-  number = {4},
-  pages = {629--635},
-  issn = {03010082},
-  doi = {10.1016/j.pneurobio.2011.09.005},
-  urldate = {2025-08-11},
-  copyright = {https://www.elsevier.com/tdm/userlicense/1.0/},
-  langid = {english}
-}
-
 @article{masanneckEvidenceClinicalTrialsgovGrowth2023,
   title = {Evidence from {{ClinicalTrials}}.Gov on the Growth of {{Digital Health Technologies}} in Neurology Trials},
   author = {Masanneck, Lars and Gieseler, Pauline and Gordon, William J. and Meuth, Sven G. and Stern, Ariel D.},
@@ -166,6 +150,16 @@
   issn = {2398-6352},
   doi = {10.1038/s41746-023-00767-1},
   urldate = {2025-08-11},
+  langid = {english},
+}
+
+@article{mijnsbergenDisturbedRestactivityRhythm2025,
+  title = {Disturbed Rest-Activity Rhythm in {{Parkinson}}'s Disease: {{Associations}} with Motor Severity and Orthostatic Hypotension},
[…]
