### 2026-03-25 107986f
Merge main into paper
