+++ src/paradigma/pipelines/tremor_pipeline.py
@@ -1,7 +1,6 @@
 import json
 import logging
 from pathlib import Path
-from typing import List
 
 import numpy as np
 import pandas as pd
