+++ paper/paper.bib
@@ -156,7 +156,7 @@
 @article{mijnsbergenDisturbedRestactivityRhythm2025,
   title = {Disturbed Rest-Activity Rhythm in {{Parkinson}}'s Disease: {{Associations}} with Motor Severity and Orthostatic Hypotension},
   shorttitle = {Disturbed Rest-Activity Rhythm in {{Parkinson}}'s Disease},
-  author = {{Mijnsbergen, Manon D.} and {Exadaktylos, Vasileios} and {van Hilten, Jacobus J.} and {Hepp, Dagmar H.} and {Weijer, Roel HA.}},
+  author = {Mijnsbergen, Manon D. and Exadaktylos, Vasileios and van Hilten, Jacobus J. and Hepp, Dagmar H. and Weijer, Roel HA.},
   date = {2025-10-21},
   journaltitle = {Journal of Parkinson’s Disease},
   url = {https://journals.sagepub.com/doi/10.1177/1877718X251388890?url_ver=Z39.88-2003&rfr_id=ori%3Arid%3Acrossref.org&rfr_dat=cr_pub++0pubmed},
