### 2026-03-25 #9502 @Erikpostt
@Kevin-Mattheus-Moerman The updated paper is ready for review. I have made the following minor changes upon inspection:

* L102-L103: Added references to two preprints and rephrased slightly based on these added references (previously: "Ongoing work is focused on longitudinal validation of the pipelines, including assessments of test-retest reliability and sensitivity to disease progression"; currently: "Additionally, detailed assessments of test-retest reliability and sensitivity to disease progression have been conducted of the gait and tremor pipelines";
* L121: Changed PPMI to ProPark, as our recent research efforts have divested from the one to the other;
* L19, L36, L74 : Changed "wrist-worn sensor data" to "wrist sensor data" for correctness
* Modified incorrect ORCIDs

Furthermore, I have merged all recent changes into the `paper` branch.
