### 2026-03-25 #9502 @Erikpostt
* Authors, affiliations and ORCIDs have been updated and look correct;
* Latest release: v1.1.1 ([10.5281/zenodo.19207320](https://doi.org/10.5281/zenodo.19207320));
* Authors and affiliations are correct in release/archive;
* DOI updated;
* Titles align;
* Licenses align

I have manually updated the title and 3 ORCIDs in the Zenodo archive such that it aligns with the paper. These changes have also been implemented in the `paper` branch, which shall be integrated upon a next release. If you require a new release with these changes visible in CITATION.cff prior to publication of the paper please let me know.
