### 2026-03-25 #9502 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
