### 2026-03-26 #9502 @Kevin-Mattheus-Moerman
@Erikpostt As AEiC for JOSS I will now help to process this submission for acceptance in JOSS. Below are some final checks:

- [x] Copyedit pass/checked for typos
